==> 03_disturbance_and_succession.Rmd <==
---
title: "03_disturbance_and_succession"
author: "Adriana Uscanga"
date: "7/10/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Load packages and files:

```{r}
library(tidyverse)
library(ggplot2)
# library(vegan)
# library(factoextra)
# library(gridExtra)
# library(NbClust)
# library(goeveg)

load("output/sites_cf2.RData")
load("output/plots_cf2.RData")
load("output/trees_cf2.RData")
```

# Forest Disturbance and Succession

## Forest Disturbance

Read and order FI forest disturbance dataset.

```{r}
disturbance <- load("output/disturbance_sno.RData")

disturbance_cf <- sites_cf %>%
  dplyr::select(site) %>%
  left_join(disturbance_sno, by = "site") %>%
  dplyr::select(-c(on_soil, on_water))

disturbance_cf
rm(disturbance_sno)
  
```

Analyze the types of disturbance recorded

```{r}

disturbance_cf %>%
  group_by(on_veg, cause) %>%
  summarize(cause = unique(cause),
            cause_no = n())

disturbance_cf %>%
  group_by(cause, on_veg) %>%
  summarize(no_sites = n()) %>%
  #filter(on_veg == "Mayor") %>%
  ggplot(aes(x= cause, y= no_sites, fill= on_veg)) +
  geom_bar(stat = "identity") +
  theme_classic() +
  theme(axis.text.x = element_text(angle=60, vjust = 1, hjust = 1))

disturbance_cf %>%
  group_by(cause, notes) %>%
  summarize(notes = unique(notes))


#Disturbances related to agriculture and cattle ranches:
#Aprovechamientos forestales
#Incendios
#Pastoreo
#Uso del suelo diferente al forestal

disturbance_cf %>%
  group_by(site) %>%
  summarize(sites_list = unique(site))

disturbance_cf %>%
  filter(is.na(cause)) #14
```


```{r}

agriculture_grazing <- disturbance_cf %>%
  mutate(cause = ifelse(is.na(cause), "unknown", cause)) %>% #Keep sites that were not recorded on disturbance data set
  filter(cause == "Uso del suelo diferente al forestal" | #LULCC
           cause == "Aprovechamientos forestales" | #Logging
           cause == "Incendios" | #Fire -- swidden ag
           cause == "Pastoreo" | # Cattle grazing
           cause == "unknown") %>% #disturbance not reported
  mutate(cause = ifelse(cause == "Uso del suelo diferente al forestal", "lulcc",
                        ifelse(cause =="Aprovechamientos forestales", "logging",
                               ifelse(cause == "Incendios", "fire",
                                      ifelse(cause == "Pastoreo", "grazing", cause))))) %>%
  mutate(cause = ifelse(cause == "logging" & 
                          notes == "Tala para sembrar café." |
                          notes == "Roza-Tumba y quema." |
                          notes == "Roza tumba quema cambio a agricultura temporal.", "agriculture", cause)) %>%
  mutate(cause = ifelse(cause == "fire" &
                          notes == "Quemas agrícolas." |
                          notes == "Rosa, tumba y quema" |
                          notes == "Quema agrícola.", "agriculture", cause)) %>%
  mutate(cause = ifelse(cause == "lulcc" &
                          notes == "Forestal con ganadería extensiva y cultivo de maíz", "agriculture", cause)) %>% #both grazing and ag but this site has another entry for grazing (below) so I labelled as ag
  mutate(cause = ifelse(cause == "lulcc" &
                           notes == "Potrero" |
                           notes == "Forestal con ganadería" |
                           notes == "Forestal con ganadería extensiva" |
                           notes == "De forestal a potreros" |
                           notes == "Forestal con ganaderia extensiva.", "grazing", cause)) %>%
  mutate(cause = ifelse(cause == "lulcc", "agriculture", cause)) %>%
  filter(cause == "grazing & agriculture" |
           cause == "grazing" |
           cause == "agriculture" |
           is.na(cause)) %>%
  mutate(disturbance_type = as.character(cause),
         disturbance_severity = as.character(on_veg),
         disturbance_notes = as.character(notes)) %>%
  select(-state, -municipality, -cause, -on_veg, -notes) %>%
  mutate(disturbance_type = ifelse(is.na(disturbance_type), "unknown", disturbance_type)) %>%
  mutate(disturbance_severity = ifelse(disturbance_severity == "Mayor", 4,
                                       ifelse(disturbance_severity == "Mediana", 3,
                                              ifelse(disturbance_severity == "Menor", 2,
                                                     ifelse(disturbance_severity == "No perceptible", 1, NA))))) %>%
  mutate(disturbance_severity = as.numeric(disturbance_severity))
  
  
#Notes (causes in logging related to agriculture):
#Tala para sembrar café.
#Roza-Tumba y quema. 
#Roza tumba quema cambio a agricultura temporal.

#Notes in fire:
#Quemas agrícolas.
#Rosa, tumba y quema
#Quema agrícola.
#Notes in lulcc that are grazing:  
#Potrero
#Forestal con ganadería
#Forestal con ganadería extensiva
#De forestal a potreros
#Forestal con ganaderia extensiva.
#Forestal con ganadería extensiva y cultivo de maíz

agriculture_grazing %>%
  group_by(site) %>%
  summarize(sites = unique(site))

# result- 51 sites have either agricultural or grazing activities or both


```

51 sites have either agricultural or grazing activities and 14 are missing from the disturbance data set

Join to sites

```{r}

sites_cf <- agriculture_grazing %>%
  select(-disturbance_notes) %>%
  group_by(site, disturbance_type) %>%
  summarize(disturbance_type = unique(disturbance_type),
            disturbance_severity = max(disturbance_severity)) %>% #some sites have the same disturbance repeated but with different severity values, I kept the max value
  mutate(disturbance_severity = ifelse(disturbance_type == "unknown", -1, disturbance_severity )) %>% #Just to keep track of unknown values
  pivot_wider(names_from = disturbance_type, values_from = disturbance_severity, values_fill = 0) %>%
  right_join(sites_cf, by = "site") %>%
  mutate(agriculture = ifelse(is.na(agriculture), 0, agriculture),
         unknown = ifelse(is.na(unknown), 0, unknown),
         grazing = ifelse(is.na(grazing), 0, grazing)) %>%
  mutate(unknown = ifelse(unknown == -1, NA, unknown)) %>%
  mutate(agriculture = ifelse(is.na(unknown), NA, agriculture),
         grazing = ifelse(is.na(unknown), NA, grazing)) %>%
  select(-unknown) %>%
  relocate(c("agriculture", "grazing"), .after = "hill_shannon_ucl")

```

Logging and firewood collection

```{r}

logging <- disturbance_cf %>%
  mutate(cause = ifelse(is.na(cause), "unknown", cause)) %>% #Keep sites that were not recorded on disturbance data set
  filter(cause == "Aprovechamientos forestales") %>%  #Logging
  mutate(cause = ifelse(cause =="Aprovechamientos forestales", "logging", "unknown")) %>%
  mutate(cause = ifelse(cause == "logging" & 
                          notes == "Tala para sembrar café." |
                          notes == "Roza-Tumba y quema." |
                          notes == "Roza tumba quema cambio a agricultura temporal.", "agriculture", cause)) %>%
  filter(!cause == "agriculture") %>%
  mutate(disturbance_type = as.character(cause),
         disturbance_severity = as.character(on_veg),
         disturbance_notes = as.character(notes)) %>%
  select(-state, -municipality, -cause, -on_veg, -notes) %>%
  mutate(disturbance_type = ifelse(is.na(disturbance_type), "unknown", disturbance_type)) %>%
  mutate(disturbance_severity = ifelse(disturbance_severity == "Mayor", 4,
                                       ifelse(disturbance_severity == "Mediana", 3,
                                              ifelse(disturbance_severity == "Menor", 2,
                                                     ifelse(disturbance_severity == "No perceptible", 1, NA))))) %>%
  mutate(disturbance_severity = as.numeric(disturbance_severity))


```

Built-up

```{r}
# Apertura de caminos, Asentamientos humanos, Lineas Electricas

builtup <- disturbance_cf %>%
  mutate(cause = ifelse(is.na(cause), "unknown", cause)) %>% #Keep sites that were not recorded on disturbance data set
  filter(cause == "Apertura de caminos" | #Roads
           cause == "Asentamientos humanos" | #Urbanization
           cause == "Lineas Electricas" | #Power lines
           cause == "unknown") %>% #disturbance not reported
  mutate(cause = ifelse(cause == "Apertura de caminos" | 
                          cause =="Asentamientos humanos" |
                          cause == "Lineas Electricas", "builtup", cause)) %>%
  mutate(disturbance_type = as.character(cause),
         disturbance_severity = as.character(on_veg),
         disturbance_notes = as.character(notes)) %>%
  select(-state, -municipality, -cause, -on_veg, -notes) %>%
  mutate(disturbance_type = ifelse(is.na(disturbance_type), "unknown", disturbance_type)) %>%
  mutate(disturbance_severity = ifelse(disturbance_severity == "Mayor", 4,
                                       ifelse(disturbance_severity == "Mediana", 3,
                                              ifelse(disturbance_severity == "Menor", 2,
                                                     ifelse(disturbance_severity == "No perceptible", 1, NA))))) %>%
  mutate(disturbance_severity = as.numeric(disturbance_severity))

```

Other disturbances (fires, pest outbreaks, hurricanes)

```{r}

other_dist <- disturbance_cf %>%
  mutate(cause = ifelse(is.na(cause), "unknown", cause)) %>% #Keep sites that were not recorded on disturbance data set
  filter(cause == "Huracanes" | #Hurricanes
           cause == "Incendios" | #Fire
           cause == "Plagas y enfermedades" | #Pests
           cause == "unknown") %>% #disturbance not reported
  mutate(cause = ifelse(cause == "Incendios" &
                          notes == "Quemas agrícolas." |
                          notes == "Rosa, tumba y quema" |
                          notes == "Quema agrícola.", "agriculture", cause)) %>%
  filter(!cause == "agriculture") %>%
  mutate(cause = ifelse(cause == "Huracanes" | 
                          cause =="Plagas y enfermedades" |
                          cause == "Incendios", "other_dist", cause)) %>%
  mutate(disturbance_type = as.character(cause),
         disturbance_severity = as.character(on_veg),
         disturbance_notes = as.character(notes)) %>%
  select(-state, -municipality, -cause, -on_veg, -notes) %>%
  mutate(disturbance_type = ifelse(is.na(disturbance_type), "unknown", disturbance_type)) %>%
  mutate(disturbance_severity = ifelse(disturbance_severity == "Mayor", 4,
                                       ifelse(disturbance_severity == "Mediana", 3,
                                              ifelse(disturbance_severity == "Menor", 2,
                                                     ifelse(disturbance_severity == "No perceptible", 1, NA))))) %>%
  mutate(disturbance_severity = as.numeric(disturbance_severity))
```

Join all datasets

```{r}

sites_cf <- logging %>%
  select(-disturbance_notes) %>%
  pivot_wider(names_from = disturbance_type, values_from = disturbance_severity, values_fill = 0) %>%
  right_join(sites_cf, by = "site") %>%
  mutate(logging = ifelse(is.na(logging), 0, logging)) %>%
  mutate(logging = ifelse(is.na(agriculture), NA, logging)) %>%
  relocate(logging, .after = grazing)

sites_cf <- builtup %>%
  select(-disturbance_notes) %>%
  filter(disturbance_type == "builtup") %>%
  group_by(site, disturbance_type) %>%
  summarize(disturbance_type = unique(disturbance_type),
            disturbance_severity = max(disturbance_severity)) %>% #some sites have the same disturbance repeated but with different severity values, I kept the max value
  pivot_wider(names_from = disturbance_type, values_from = disturbance_severity, values_fill = 0) %>%
  right_join(sites_cf, by = "site") %>%
  mutate(builtup = ifelse(is.na(builtup), 0, builtup)) %>%
  mutate(builtup = ifelse(is.na(agriculture), NA, builtup)) %>%
  relocate(builtup, .after = logging)

sites_cf <- other_dist %>%
  select(-disturbance_notes) %>%
  filter(disturbance_type == "other_dist") %>%
  group_by(site, disturbance_type) %>%
  summarize(disturbance_type = unique(disturbance_type),
            disturbance_severity = max(disturbance_severity)) %>% #some sites have the same disturbance repeated but with different severity values, I kept the max value
  pivot_wider(names_from = disturbance_type, values_from = disturbance_severity, values_fill = 0) %>%
  right_join(sites_cf, by = "site") %>%
  mutate(other_dist = ifelse(is.na(other_dist), 0, other_dist)) %>%
  mutate(other_dist = ifelse(is.na(agriculture), NA, other_dist)) %>%
  relocate(other_dist, .after = builtup)
```

Join disturbance info to plots_cf data set

```{r}
plots_cf <- plots_cf %>%
  mutate(site = as.character(site))

sites_cf %>%
  select(site, agriculture, grazing, logging, builtup, other_dist) %>%
  right_join(plots_cf, by = "site") %>%
  relocate(c("agriculture", "grazing", "logging", "builtup", "other_dist"), .after = hill_shannon_ucl) -> plots_cf
```

## Forest Classification by Structure

### Non-hierarchical cluster analysis and PCA

Select structure attributes and scale dataset so that variables in different units can be compared

```{r}

# Plots

# tree_no_ave and av_treedensity_site are completely correlated.
# Remove tree_no_ave to avoid correlation in further analyses.

plots_str <- plots_cf %>%
  dplyr::select(plot_id,
         tree_density,
         loreys_height,
         basal_area_ha)

plot(plots_str)

# Scale data so that variables in different units can be compared
plots_str_scaled <- scale(plots_str[,c(3:5)])
row.names(plots_str_scaled) <- plots_str$plot_id

```

Compute K-means with several number of clusters and compare plots

```{r}

# Plots
k2_structure_plots <- kmeans(plots_str_scaled, centers = 2, nstart = 25)
k3_structure_plots <- kmeans(plots_str_scaled, centers = 3, nstart = 25)
k4_structure_plots <- kmeans(plots_str_scaled, centers = 4, nstart = 25)
k5_structure_plots <- kmeans(plots_str_scaled, centers = 5, nstart = 25)

# Compare plots:
p1_str_plots <- fviz_cluster(k2_structure_plots, geom = "point", data = plots_str_scaled) + ggtitle("k = 2")
p2_str_plots <- fviz_cluster(k3_structure_plots, geom = "point",  data = plots_str_scaled) + ggtitle("k = 3")
p3_str_plots <- fviz_cluster(k4_structure_plots, geom = "point",  data = plots_str_scaled) + ggtitle("k = 4")
p4_str_plots <- fviz_cluster(k5_structure_plots, geom = "point",  data = plots_str_scaled) + ggtitle("k = 5")
grid.arrange(p1_str_plots, p2_str_plots, p3_str_plots, p4_str_plots, nrow = 2)

```

Data can be classified in many different number of clusters.
To define the best number of clusters, we're using NbClust.

```{r}
# Plots

str_clusters_plots <- NbClust((plots_str_scaled^2), distance = "euclidean", min.nc = 2, max.nc = 10, method = "ward.D2", index = "all")


```

Most indices proposed 3 as the best number of clusters (for both sites and plots).
Now I can classify data in three classes. Also, I can visualize these three classes in a PCA to see how structure attributes defined them.

```{r}
# Plots

# Save a vector of 3 clusters:
str_k3_plots <- k3_structure_plots$cluster
as.numeric(str_k3_plots)

#pca with structural data

pca_str_plots <- rda(plots_str_scaled, scale = F)
summary(pca_str_plots) # 
plot(pca_str_plots)

ordiplot(pca_str_plots, display = 'sites', type = 'n')
points(pca_str_plots, col = str_k3_plots)
biplot(pca_str_plots, display = 'species')

ordiplot(pca_str_plots, display = 'sites', type = 'text')
biplot(pca_str_plots)
```

Add cluster class plots_cf dataset.

```{r}

# Plots

str_k3_plots <- as_tibble(str_k3_plots, rownames = "plot_id")

str_k3_plots <- str_k3_plots %>%
  mutate(str_class_plots = as.character(value)) %>%
  select(-value)

plots_cf <- plots_cf %>%
  left_join(str_k3_plots, by = "plot_id")


```

### Strucutre attributes across clusters 

Let's see how structure attributes change in these three clusters.

```{r}

# Plots

plots_cf %>%
  ggplot(aes(x = tree_density, group = str_class_plots, fill = str_class_plots, color = str_class_plots)) +
  geom_density(alpha= 0.5) +
  scale_fill_manual(values = c("tomato", "green3", "dodgerblue2")) +
  scale_color_manual(values = c("tomato", "green3", "dodgerblue2")) +
  ggtitle("Tree density")

plots_cf %>%
  ggplot(aes(x = loreys_height, group = str_class_plots, fill = str_class_plots, color = str_class_plots)) +
  geom_density(alpha= 0.5) +
  scale_fill_manual(values = c("tomato", "green3", "dodgerblue2")) +
  scale_color_manual(values = c("tomato", "green3", "dodgerblue2")) +
  ggtitle("Tree height")

plots_cf %>%
  ggplot(aes(x = basal_area_ha, group = str_class_plots, fill = str_class_plots, color = str_class_plots)) +
  geom_density(alpha= 0.5) +
  scale_fill_manual(values = c("tomato", "green3", "dodgerblue2")) +
  scale_color_manual(values = c("tomato", "green3", "dodgerblue2")) +
  ggtitle("Basal area")

```

```{r}

# Plots

plots_cf %>%
  ggplot(aes(y = tree_density, color = str_class_plots)) +
  geom_boxplot()

plots_cf %>%
  ggplot(aes(y = loreys_height, color = str_class_plots)) +
  geom_boxplot()

plots_cf %>%
  ggplot(aes(y = basal_area_ha, color = str_class_plots)) +
  geom_boxplot()

plots_cf %>%
  ggplot(aes(y = agb_plot_ha, color = str_class_plots)) +
  geom_boxplot()

plots_cf %>%
  ggplot(aes(y = hill_shannon, color = str_class_plots)) +
  geom_boxplot()

plots_cf %>%
  ggplot(aes(y = hill_richness, color = str_class_plots)) +
  geom_boxplot()

```

* Note: there's a plot with more than 600 Mg/ha of biomass!

Structure attributes are clearly different among clusters. It seems that plots split in those with very few trees, many trees of medium size, and plots with very big trees with medium tree density. This broadly coincides with succession stages: very young, young-medium, and mature forests. 

How can we know this classification is more related to disturbance/succession than climate?
We can compare the 3 clusters with info on disturbance (from FI)
We can explore whether plots at higher elevations with stunted trees may be passing by secondary forests,
but this is exactly the question we are after (the relationship between structure, environment, and land use)
We can also compare this classification with the disturbance history of this site using this
Tracking long-term (1990-2021) deforestation and degradation in tropical moist forests
https://forobs.jrc.ec.europa.eu/TMF/index.php

```{r}

plots_cf %>%
  ggplot(aes(x = altitude, y = agriculture, color = str_class_plots)) +
  geom_jitter()

plots_cf %>%
  ggplot(aes(x = altitude, y = logging, color = str_class_plots)) +
  geom_jitter()

plots_cf %>%
  ggplot(aes(x = altitude, y = grazing, color = str_class_plots)) +
  geom_jitter()

plots_cf %>%
  mutate(lu = (agriculture + grazing + logging),
         dist = (agriculture + grazing + logging + builtup)) %>%
  #filter(lu == 0) %>%
  ggplot(aes(x = altitude, y = lu, color = str_class_plots)) +
  geom_jitter()

plots_cf %>%
  filter(altitude > 2500,
         str_class_plots == 2)
# All plots above 2,500 m asl in class 2 have logging activities  
```

### Landscape composition: mosaic description

```{r}

mosaic <- plots_cf %>%
  mutate(str_class_plots = as.numeric(str_class_plots),
         site = as.character(site)) %>%
  group_by(site) %>%
  # normalized mosaic = value - min value / max value - min value
  # mosaic goes from 1 (all mature forest) to 0 (all agriculture/very young forest)
  summarize(mosaic4 = ifelse(n()==4, (sum(str_class_plots)-4)/(12-4), NA),
            mosaic3 = ifelse(n()==3, (sum(str_class_plots)-3)/(9-3), NA),
            mosaic2 = ifelse(n()==2, (sum(str_class_plots)-2)/(6-2), NA),
            mosaic1 = ifelse(n()==1, (sum(str_class_plots)-1)/(3-1), NA),
            mosaic = ifelse(n()==4, mosaic4,
                            ifelse(n()==3, mosaic3,
                                   ifelse(n()==2, mosaic2, mosaic1)))) %>%
  arrange(mosaic)

sites_cf <- sites_cf %>%
  left_join(mosaic, by = "site") %>%
  select(-c(mosaic1, mosaic2, mosaic3, mosaic4))

sites_cf %>%
  filter(plot_no == 4) %>%
  ggplot(aes(x= mosaic, y = site)) +
  geom_point()

sites_cf %>%
  filter(plot_no == 4,
         mosaic == 0.5)

#Note: this visualization only works with this classification of str_class_plots 
# (if k-means is repeated, classification may change)

sites_cf %>%
  filter(plot_no >3) %>%
  group_by(mosaic) %>%
  summarize(agb_mosaic = mean(av_agb_site), av_agb_site = as.numeric(av_agb_site)) %>%
  ggplot(aes(x = mosaic, y = av_agb_site)) +
  geom_point() +
  geom_point(aes(x= mosaic, y = agb_mosaic), color = "red")
  

```

## Forest classification by disturbance

Based on https://www.science.org/doi/10.1126/sciadv.abe1603

```{r}

forestclass <- read_csv("input/forestclass.csv") # extracted from JRC_TMF_TransitionMap_MainClasses_v1_1982_2022_SAM_ID65_N20_W100 and JRC_TMF_TransitionMap_Subtypes_v1_1982_2022_SAM_ID65_N20_W100

forestclass <- forestclass %>%
  transmute(plot_id = as.character(field_1),
              subtypes = as.character(JRC_TMF_TransitionMap_Subtypes_v1_1982_2022_SAM_ID65_N20_W100),
              mainclasses = as.character(JRC_TMF_TransitionMap_MainClasses_v1_1982_2022_SAM_ID65_N20_W100))

forestclass %>%
  group_by(subtypes) %>%
  summarize(subtypes_list = unique(subtypes))

forestclass %>%
  group_by(mainclasses) %>%
  summarize(mainclasses_list = unique(mainclasses), counts = n())

forestclass <- forestclass %>%
  mutate(mainclasses_legend = ifelse(mainclasses == "10", "Undisturbed",
                                     ifelse(mainclasses == "20", "Degraded",
                                            ifelse(mainclasses == "30", "Regrowth",
                                                   ifelse(mainclasses == "43", "Deforested",
                                                          ifelse(mainclasses == "50", "Ongoing deforestation/degradation", "Other"))))))

forestclass %>%
  group_by(mainclasses_legend) %>%
  summarize(counts = n())

forestclass <- forestclass %>%
  mutate(subtypes_legend = ifelse(subtypes == "10", "Undisturbed",
                                  ifelse(subtypes == "21", "Degraded forest with short-duration disturbance (started before 2012)",
                                         ifelse(subtypes == "22", "Degraded forest with short-duration disturbance (started in 2012-2020)",
                                                ifelse(subtypes == "23", "Degraded forest with long-duration disturbance (started before 2012)",
                                                       ifelse(subtypes == "24", "Degraded forest with long-duration disturbance (started in 2012-2020)",
                                                              ifelse(subtypes == "25", "Degraded forest with 2/3 short degradation periods (last degradation started before 2012)",
                                                                     ifelse(subtypes == "26", "Degraded forest with 2/3 short degradation periods (last degradation started in 2012-2020)",
                                                                            ifelse(subtypes == "31", "Old forest regrowth (disturbed before 2002)",
                                                                                   ifelse(subtypes == "32", "Young forest regrowth (disturbed in 2002-2011)",
                                                                                          ifelse(subtypes == "33", "Very young forest regrowth (disturbed in 2012-2018)",
                                                                                                 ifelse(subtypes == "41", "Deforestation started before 2011",
                                                                                                        ifelse(subtypes == "42", "Deforestation started in 2011-2018",
                                                                                                               ifelse(subtypes == "51", "Deforestation started in 2019",
                                                                                                                      ifelse(subtypes == "52", "Deforestation started in 2020",
                                                                                                                             ifelse(subtypes == "91", "Other LC without afforestation",
                                                                                                                                    ifelse(subtypes == "92", "Young afforestation (between 3 and 9 years of regrowth)", "Old afforestation (between 10 and 20 years of regrowth)")))))))))))))))))
       

forestclass %>%
  group_by(mainclasses_legend, subtypes_legend) %>%
  summarize(counts = n())

```
Compare to plots_cf data

```{r}
plots_cf <- plots_cf %>%
  left_join(forestclass)

plots_cf %>%
  group_by(mainclasses_legend) %>%
  summarize(counts = n())

plots_cf %>%
  dplyr::select(plot_id, agriculture, grazing, logging, builtup, str_class_plots, subtypes, mainclasses, subtypes_legend, mainclasses_legend) %>%
  group_by(str_class_plots, mainclasses_legend, subtypes_legend) %>%
  summarize(counts = n())

plots_cf %>%
  ggplot(aes(x= mainclasses_legend, y = agb_plot_ha)) +
  geom_boxplot()
```

Match time of disturbance to time of data collection

```{r}
plots_cf %>%
  filter(!is.na(agriculture)) %>%
  select(plot_id, year, agriculture, grazing, logging, builtup, str_class_plots, subtypes_legend, mainclasses_legend) %>%
  group_by(mainclasses_legend, subtypes_legend) %>%
  summarize(year = as.numeric(year),
            plot_id = as.character(plot_id),
            str_class_plots = as.numeric(str_class_plots),
            agriculture = as.numeric(agriculture),
            grazing = as.numeric(grazing),
            logging = as.numeric(logging),
            builtup = as.numeric(builtup))

plots_cf %>%
  filter(!is.na(agriculture)) %>%
  group_by(mainclasses_legend) %>%
  summarize(count = n())
```
Make new classification, from:
Deforested	17			
Degraded	54			
Ongoing deforestation/degradation	2			
Other	83			
Regrowth	20			
Undisturbed	92

To: 
Deforested - agriculture, grazing, and intensive logging
Degraded - some logging (wood harvesting- logging values 1 and 2)
Regrowth and Afforestation - forest growth and not noticeable disturbance (values 0 and 1) 
Undisturbed

```{r}
# Undisturbed (n = 92)

reclass1 <- plots_cf %>%
  filter(!is.na(agriculture)) %>%
  filter(mainclasses_legend == "Undisturbed") %>% # 92 plots
   #8 plots lack info on disturbance
  mutate(lu = (agriculture + grazing),
         dist = (agriculture + grazing + logging + builtup + other_dist)) %>%
  arrange(-lu) %>%
  mutate(reclass = ifelse(lu > 0, "as_deforested",
                          ifelse(lu == 0 & logging > 2, "as_deforested2",
                                 ifelse(lu == 0 & logging < 3 & logging > 0, "as_degraded", "as_undisturbed")))) %>%
  select(plot_id, reclass)

reclass_undisturbed <- reclass1
```
Plots that are "undisturbed" in forestclass product (mainclasses_legend), but have disturbance according to FI were re-classified as 'deforested' or 'degraded', according to the level and type of disturbance. 

Plots in other categories that were undisturbed by the time of data collection

```{r}
# Deforested (n=17)

reclass2 <- plots_cf %>%
  filter(!is.na(agriculture)) %>%
  filter(mainclasses_legend == "Deforested") %>% 
  filter(subtypes_legend == "Deforestation started before 2011") %>%
  # 9 plots, most of them have agriculture, grazing or logging > 2
  mutate(lu = (agriculture + grazing),
         dist = (agriculture + grazing + logging + builtup + other_dist)) %>%
  arrange(-lu) %>%
  mutate(reclass = ifelse(lu > 0, "as_deforested", "as_degraded")) %>%
  select(plot_id, reclass)

reclass3 <- plots_cf %>%
  filter(!is.na(agriculture)) %>%
  filter(mainclasses_legend == "Deforested") %>% 
  filter(subtypes_legend == "Deforestation started in 2011-2018") %>%
  # 8 plots, many have ag, grazing, logging = 0, re-classify to undisturbed
  mutate(lu = (agriculture + grazing),
         dist = (agriculture + grazing + logging + builtup + other_dist)) %>%
  mutate(reclass = ifelse(lu > 0, "as_deforested", "as_undisturbed")) %>%
  select(plot_id, reclass)

reclass_deforested <- rbind(reclass2, reclass3)
```

Degraded 

```{r}
# Degraded (n = 54)

plots_cf %>%
  filter(!is.na(agriculture)) %>%
  filter(mainclasses_legend == "Degraded") %>%
  # 67 plots
  # 13 plots lack info on disturbance (FI)
  group_by(mainclasses_legend, subtypes_legend) %>%
  summarize(counts = n()) # 6 subtypes

reclass4 <- plots_cf %>%
  filter(!is.na(agriculture)) %>%
  filter(mainclasses_legend == "Degraded",
         subtypes_legend == "Degraded forest with 2/3 short degradation periods (last degradation started before 2012)") %>%
  arrange(year) %>%
  mutate(lu = (agriculture + grazing),
         dist = (agriculture + grazing + logging + builtup + other_dist)) %>%
  arrange(-lu) %>%
  mutate(reclass = ifelse(lu > 0, "as_deforested",
                          ifelse(lu == 0 & logging > 2, "as_deforested2","as_degraded"))) %>%
  select(plot_id, reclass)
# Keep some as degraded, re-classify some as deforested

plots_cf %>%
  filter(!is.na(agriculture)) %>%
  filter(mainclasses_legend == "Degraded",
         subtypes_legend == "Degraded forest with 2/3 short degradation periods (last degradation started in 2012-2020)") %>%
  arrange(year) 

reclass5 <- plots_cf %>%
  filter(!is.na(agriculture)) %>%
  filter(mainclasses_legend == "Degraded",
         subtypes_legend == "Degraded forest with 2/3 short degradation periods (last degradation started in 2012-2020)") %>%
  arrange(year) %>%
  mutate(lu = (agriculture + grazing),
         dist = (agriculture + grazing + logging + builtup + other_dist)) %>%
  arrange(-lu) %>%
  mutate(reclass = ifelse(lu > 0, "as_deforested",
                          ifelse(year < 2012 & lu == 0 & logging == 0, "as_undisturbed","as_degraded"))) %>%
  select(plot_id, reclass)
# All plots have disturbance according to FI even before 2012
# Keep some as degraded, re-classify some as deforested

reclass6 <- plots_cf %>%
  filter(!is.na(agriculture)) %>%
  filter(mainclasses_legend == "Degraded",
         subtypes_legend == "Degraded forest with long-duration disturbance (started before 2012)") %>%
  mutate(lu = (agriculture + grazing),
         dist = (agriculture + grazing + logging + builtup + other_dist)) %>%
  arrange(-lu) %>%
  mutate(reclass = ifelse(lu > 0, "as_deforested", "as_degraded")) %>%
  select(plot_id, reclass)
# Keep as degraded

reclass7 <- plots_cf %>%
  filter(!is.na(agriculture)) %>%
  filter(mainclasses_legend == "Degraded",
         subtypes_legend == "Degraded forest with long-duration disturbance (started in 2012-2020)") %>%
  arrange(year) %>%
  mutate(lu = (agriculture + grazing),
         dist = (agriculture + grazing + logging + builtup + other_dist)) %>%
  arrange(-lu) %>%
  mutate(reclass = ifelse(lu > 0, "as_deforested",
                          ifelse(lu == 0 & logging > 2, "as_deforested2",
                                 ifelse(lu == 0 & logging < 3 & logging > 0, "as_degraded", "as_undisturbed")))) %>%
  select(plot_id, reclass)
# Some deforestation/degradation took place after data collection, 
# re-classify plots 

reclass8 <- plots_cf %>%
  filter(!is.na(agriculture)) %>%
  filter(mainclasses_legend == "Degraded",
         subtypes_legend == "Degraded forest with short-duration disturbance (started before 2012)") %>%
  mutate(lu = (agriculture + grazing),
         dist = (agriculture + grazing + logging + builtup + other_dist)) %>%
  arrange(-lu) %>%
  mutate(reclass = ifelse(lu > 0, "as_deforested",
                          ifelse(lu == 0 & logging > 2, "as_deforested2",
                                 ifelse(lu == 0 & logging < 3 & logging > 0, "as_degraded", "as_degraded2")))) %>%
  select(plot_id, reclass)
# Keep some as degraded, change as deforested the ones that have agriculture and/or grazing
# because degradation took place before data collection, none of the plots are reclassified as undisturbed

reclass9 <- plots_cf %>%
  filter(!is.na(agriculture)) %>%
  filter(mainclasses_legend == "Degraded",
         subtypes_legend == "Degraded forest with short-duration disturbance (started in 2012-2020)") %>%
  # 7 plots
  arrange(year) %>%
  filter(year < 2013) %>%
  mutate(lu = (agriculture + grazing),
         dist = (agriculture + grazing + logging + builtup + other_dist)) %>%
  arrange(-lu) %>%
  mutate(reclass = ifelse(lu > 0, "as_deforested",
                          ifelse(lu == 0 & logging > 2, "as_deforested2",
                                 ifelse(lu == 0 & logging < 3 & logging > 0, "as_degraded", "as_undisturbed")))) %>%
  select(plot_id, reclass)

reclass_degraded <- rbind(reclass4, reclass5, reclass6, reclass7, reclass8, reclass9)
```

Regrowth

```{r}
#Regrowth (n = 20)
# Classify as forest growth when meanclasses_legend = Regrowth and there's not noticeable disturbance (values 0 and 1)

plots_cf %>%
  filter(mainclasses_legend == "Regrowth") %>%
  # 21 plots
  filter(!is.na(agriculture)) %>%
  # 1 plot lacks info on disturbance (FI)
  group_by(mainclasses_legend, subtypes_legend) %>%
  summarize(counts = n())

reclass10 <- plots_cf %>%
  filter(!is.na(agriculture)) %>%
  filter(mainclasses_legend == "Regrowth",
         subtypes_legend == "Old forest regrowth (disturbed before 2002)") %>%
  mutate(lu = (agriculture + grazing),
         dist = (agriculture + grazing + logging + builtup + other_dist)) %>%
  arrange(-lu) %>%
  # All plots have disturbance according to FI dataset
  mutate(reclass = ifelse(lu > 0, "as_deforested", "as_regrowth")) %>%
  select(plot_id, reclass)
# Note that plots in secondary succession experience disturbance by logging

reclass11 <- plots_cf %>%
  filter(!is.na(agriculture)) %>%
  filter(mainclasses_legend == "Regrowth",
         subtypes_legend == "Very young forest regrowth (disturbed in 2012-2018)") %>%
  arrange(year) %>%
  mutate(lu = (agriculture + grazing),
         dist = (agriculture + grazing + logging + builtup + other_dist)) %>%
  arrange(-lu) %>%
  #filter(year < 2013) %>%
  mutate(reclass = ifelse(lu > 0, "as_deforested",
                          ifelse(lu == 0 & logging < 3 & logging > 0 & year < 2012, "as_degraded",
                                 ifelse(lu == 0 & logging == 0 & year < 2012, "as_undisturbed", "as_regrowth")))) %>%
  select(plot_id, reclass)

reclass12 <- plots_cf %>%
  filter(!is.na(agriculture)) %>%
  filter(mainclasses_legend == "Regrowth",
         subtypes_legend == "Young forest regrowth (disturbed in 2002-2011)") %>%
  mutate(lu = (agriculture + grazing),
         dist = (agriculture + grazing + logging + builtup + other_dist)) %>%
  arrange(-lu) %>%
  mutate(reclass = ifelse(lu > 0, "as_deforested", "as_regrowth")) %>%
  select(plot_id, reclass)

reclass_regrowth <- rbind(reclass10, reclass11, reclass12)
```
Other

```{r}

#Other (n= 83)

plots_cf %>%
  filter(mainclasses_legend == "Other") %>%
  # 91 plots
  filter(!is.na(agriculture)) %>%
  # 83 plots
  group_by(mainclasses_legend, subtypes_legend) %>%
  summarize(counts = n())

reclass13 <- plots_cf %>%
  filter(!is.na(agriculture)) %>%
  filter(mainclasses_legend == "Other",
         subtypes_legend == "Old afforestation (between 10 and 20 years of regrowth)") %>%
  mutate(lu = (agriculture + grazing),
         dist = (agriculture + grazing + logging + builtup + other_dist)) %>%
  arrange(-lu) %>%
  # only keep as regrowth when lu < 2
  mutate(reclass = ifelse(lu > 0, "as_deforested",
                          ifelse(lu == 0 & logging > 2, "as_deforested2", "as_afforestation"))) %>%
  select(plot_id, reclass)
# re classify plots that have ag, grazing, and intensive logging as deforested and the others as afforestation
# note that afforestation plots can present some 'mild' logging (logging < 3)

reclass14 <- plots_cf %>%
  filter(!is.na(agriculture)) %>%
  filter(mainclasses_legend == "Other",
         subtypes_legend == "Other LC without afforestation") %>%
  mutate(lu = (agriculture + grazing),
         dist = (agriculture + grazing + logging + builtup + other_dist)) %>%
  arrange(-lu) %>%
  mutate(reclass = ifelse(lu > 0, "as_deforested",
                          ifelse(lu == 0 & logging > 2, "as_deforested2",
                                 ifelse(lu == 0 & logging < 3 & logging > 0, "as_degraded", "other")))) %>%
  select(plot_id, reclass)
# keep as other
# Most of Other LC without afforestation plots have ag, grazing, or logging
# These are probably plots that have been 'managed' many years


reclass15 <- plots_cf %>%
  filter(!is.na(agriculture)) %>%
  filter(mainclasses_legend == "Other",
         subtypes_legend == "Young afforestation (between 3 and 9 years of regrowth)") %>%
  mutate(lu = (agriculture + grazing),
         dist = (agriculture + grazing + logging + builtup + other_dist)) %>%
  arrange(-lu) %>%
  mutate(reclass = ifelse(lu > 0, "as_deforested",
                          ifelse(lu == 0 & logging > 2, "as_deforested2",
                                 ifelse(lu == 0 & logging < 3 & logging > 0, "as_degraded", "other")))) %>%
  select(plot_id, reclass)
# Afforestation happen after data collection. These are forests with some logging activity (deforested or degraded)  

reclass_other <- rbind(reclass13, reclass14, reclass15)
```

Ongoing deforestation/degradation

```{r}

# Ongoing (n= 2)

plots_cf %>%
  filter(mainclasses_legend == "Ongoing deforestation/degradation") %>%
  # 3 plots
  filter(!is.na(agriculture)) %>%
  # 2 plots
  group_by(mainclasses_legend, subtypes_legend) %>%
  summarize(counts = n()) 

reclass16 <- plots_cf %>%
  filter(mainclasses_legend == "Ongoing deforestation/degradation") %>%
  # 3 plots
  filter(!is.na(agriculture)) %>%
  mutate(reclass = "as_undisturbed") %>%
  select(plot_id, reclass)
# Deforestation started after data collection and FI records don't show disturbances

reclass_ongoing <- reclass16
```
Bind datasets

```{r}

reclass_all <- rbind(reclass_deforested, reclass_degraded, reclass_ongoing, reclass_other, reclass_regrowth, reclass_undisturbed)

plots_cf %>%
  filter(!is.na(agriculture))

reclass_all %>%
  group_by(reclass) %>%
  summarize(counts = n())

reclass_all <- reclass_all %>%
  mutate(forest_classification = ifelse(reclass == "as_afforestation" |
                                          reclass == "as_regrowth", "regrowth",
                                        ifelse(reclass == "as_deforested" |
                                                 reclass == "as_deforested2", "deforested",
                                               ifelse(reclass == "as_degraded" |
                                                        reclass == "as_degraded2", "degraded",
                                                      ifelse(reclass == "as_undisturbed", "undisturbed", "other"))))) %>%
  select(-site)

reclass_all %>%
  group_by(reclass) %>%
  summarize(counts = n())

reclass_all %>%
  group_by(forest_classification) %>%
  summarize(counts = n())

plots_cf %>%
  left_join(reclass_all, by = "plot_id") %>%
  filter(forest_classification != "other",
         !is.na(forest_classification)) %>%
  group_by(forest_classification) %>%
  ggplot(aes(y = agb_plot_ha, color = forest_classification)) +
  geom_boxplot()

plots_cf <- plots_cf %>%
  left_join(reclass_all, by = "plot_id")

plots_cf <- plots_cf %>%
  mutate(site = as.character(site.x)) %>%
  select(-site.x, -site.y)

```




Year of disturbance

```{r}

plots_cf %>%
  left_join(reclass_all, by = "plot_id") %>%
  left_join(def_deg_year) %>%
  select(plot_id, year, forest_classification, JRC_TMF_DeforestationYear_INT_1982_2022_v1_SAM_ID65_N20_W100, JRC_TMF_DegradationYear_INT_1982_2022_v1_SAM_ID65_N20_W100) %>%
  mutate(def_year = as.numeric(JRC_TMF_DeforestationYear_INT_1982_2022_v1_SAM_ID65_N20_W100),
            deg_year = as.numeric(JRC_TMF_DegradationYear_INT_1982_2022_v1_SAM_ID65_N20_W100)) %>%
  select(-JRC_TMF_DeforestationYear_INT_1982_2022_v1_SAM_ID65_N20_W100,
         -JRC_TMF_DegradationYear_INT_1982_2022_v1_SAM_ID65_N20_W100)
```

## Save datasets

```{r}
save(sites_cf, file = "output/sites_cf3.RData")
save(plots_cf, file = "output/plots_cf3.RData")
save(trees_cf, file = "output/trees_cf3.RData")

```
