==> 04_data_preparation.Rmd <==
---
title: "04_data_preparation"
author: "Adriana Uscanga"
date: "14/2/2022"
output: html_document
---

# Data preparation

Load packages and data

```{r}
library(tidyverse)
library(ggpubr)

load("output/sites_cf3.RData")
load("output/plots_cf3.RData")
load("output/trees_cf3.RData")
```

Add slope and aspect to the data set.

```{r}
#Aspect and Slope from GEE 
aspect <- read_csv("input/aspect_plots.csv", col_names = F) 

slope <- read_csv("input/slope_plots.csv", col_names = F) 

head(aspect)

aspect <- aspect %>%
  transmute(plot_id = as.character(X5),
            lon = as.numeric(X3),
            lat = as.numeric(X4),
            aspect = as.numeric(X2))

head(aspect)

slope <- slope %>%
  transmute(plot_id = as.character(X5),
            lon = as.numeric(X3),
            lat = as.numeric(X4),
            slope = as.numeric(X2))
head(slope)

aspect_slope <- aspect %>%
  left_join(slope, by = c("plot_id", "lon", "lat"))

plots_cf <- plots_cf %>%
  left_join(aspect_slope, by = c("plot_id")) %>%
  select(-lon, -lat) %>%
  relocate(c(aspect, slope), .after = altitude)

plots_cf %>%
  filter(duplicated(longitude), duplicated(latitude)) %>%
  group_by(site) %>%
  summarize(sites = unique(site), plots = unique(plot_id))

plots_cf <- plots_cf %>%
  group_by(site) %>%
  summarize(plot_id = plot_id,
            aspect_first = first(aspect),
            dup = ifelse(duplicated(longitude) & duplicated(latitude), "duplicated", NA),
            aspectdups = ifelse(dup == "duplicated", aspect_first, NA),
            slope_first = first(slope),
            dup_s = ifelse(duplicated(longitude) & duplicated(latitude), "duplicated", NA),
            slopedups = ifelse(dup_s == "duplicated", slope_first, NA)) %>%
  filter(!is.na(dup)) %>%
  select(plot_id, aspectdups, slopedups) %>%
  right_join(plots_cf) %>%
  mutate(aspect = ifelse(is.na(aspect), aspectdups, aspect),
         slope = ifelse(is.na(slope), slopedups, slope)) %>%
  select(-aspectdups, -slopedups)

plots_cf %>%
  filter(is.na(aspect))
#Weird, plot 67878_2 doesn't have aspect and slope, it must have been missing from the GEE feature collection. I'll extract data for this plot and add it.
# aspect = 334.7877502441406
# slope= 34.21200942993164

plots_cf <- plots_cf %>%
  mutate(aspect = ifelse(plot_id == "67878_2", 334.787750, aspect),
         slope = ifelse(plot_id == "67878_2", 34.212009, slope))

plots_cf <- plots_cf %>%
  mutate(northness = cos(aspect),
         eastness = sin(aspect)) %>%
  relocate(c("northness", "eastness"), .after = aspect)

```

Average sites

```{r}
# Average aspect and slope by site

calcSE<-function(x){
  x <- x[is.na(x)==F]
  sd(x)/sqrt(length(x))
}

aspect_slope_sites <- plots_cf %>%
  select(site, plot_id, aspect, northness, eastness, slope) %>%
  group_by(site) %>%
  summarize(aspect_gee = mean(aspect),
            aspect_se_gee = calcSE(aspect),
            northness_gee = mean(northness),
            northness_se_gee = calcSE(northness),
            eastness_gee = mean(eastness),
            eastness_se_gee = calcSE(eastness),
            slope_gee = mean(slope),
            slope_se_gee = calcSE(slope)) %>%
  mutate(site = as.character(site))

sites_cf <- sites_cf %>%
  left_join(aspect_slope_sites, by = "site") %>%
  relocate(c("aspect_gee",
             "aspect_se_gee",
             "northness_gee",
             "northness_se_gee",
             "eastness_gee",
             "eastness_se_gee",
             "slope_gee",
             "slope_se_gee"), .after = aspect)

```


Site 71213 in sites_cf data set lacks geographic coordinates.
Which are: 17.50506, -96.30769

```{r}
sites_cf <- sites_cf %>%
  mutate(longitude = ifelse(site == "71213", -96.30769, longitude),
         latitude = ifelse(site == "71213", 17.50506, latitude))
```

Add a column with subregions to the data frame.

```{r}
sites_huautla <- read.csv("input/sites_huautla.csv", header = T)
sites_ixtlan <- read.csv("input/sites_ixtlan.csv", header = T)
sites_mixe <- read.csv("input/sites_mixe.csv", header = T)
sites_guevea <- read.csv("input/sites_guevea.csv", header = T)
```

Bind all subregions data sets in one and join to main data set

```{r}
subregions <- rbind(sites_huautla, sites_ixtlan, sites_mixe, sites_guevea)
subregions <- mutate(subregions, subregion = as.character(SUBREGIONE))
subregions <- select(subregions, -SUBREGIONE)
subregions <- subregions %>%
  mutate(site = as.character(site))

sites_cf <- sites_cf %>%
  left_join(subregions, by = "site") %>%
  relocate(subregion, .after = site)

plots_cf <- sites_cf %>%
  select(site, subregion) %>%
  right_join(plots_cf)

```

Change direction of mosaic values (so that 0-forest and 1-ag) and rename variable

```{r}
# sites_cf <- sites_cf %>%
#  mutate(landscape_str = 1- mosaic,
#         landscape_dist = 1 - mosaic_dist)
```

Visual inspection of data

```{r}
# First, remove sites with less than 3 plots because they're noisy

sites_cf2 <- sites_cf %>%
  filter(plot_no > 2)

#density plot
ggdensity(sites_cf2$av_agb_site)
ggdensity(sites_cf2$hill_shannon)

#qqplot
ggqqplot(sites_cf2$av_agb_site)
ggqqplot(sites_cf2$hill_shannon)
```

Shapiro-Wilk's normality test:

```{r}

# Plots

shapiro.test(plots_cf$agb_plot_ha)
shapiro.test(plots_cf$temp)
shapiro.test(plots_cf$precip)
shapiro.test(plots_cf$northness) 
shapiro.test(plots_cf$eastness)
shapiro.test(plots_cf$slope) 
shapiro.test(plots_cf$logging)
shapiro.test(plots_cf$agriculture)
shapiro.test(plots_cf$grazing)
shapiro.test(plots_cf$hill_richness) 
shapiro.test(plots_cf$hill_shannon) 

ggdensity(plots_cf$agb_plot_ha)
ggdensity(plots_cf$temp)
ggdensity(plots_cf$precip)
ggdensity(plots_cf$northness) #
ggdensity(plots_cf$eastness)
ggdensity(plots_cf$slope) 
ggdensity(plots_cf$logging)
ggdensity(plots_cf$agriculture)
ggdensity(plots_cf$grazing)
ggdensity(plots_cf$hill_richness) 
ggdensity(plots_cf$hill_shannon)

# Sites

shapiro.test(sites_cf2$av_agb_site)
shapiro.test(sites_cf2$temp)
shapiro.test(sites_cf2$precip)
shapiro.test(sites_cf2$aspect_gee) # normal p = 0.42
shapiro.test(sites_cf2$slope_gee) # marginally normal p = 0.02
#shapiro.test(sites_cf2$mosaic)
shapiro.test(sites_cf2$agriculture)
shapiro.test(sites_cf2$grazing)
shapiro.test(sites_cf2$hill_richness) # normal p=0.15
shapiro.test(sites_cf2$hill_shannon) # normal p = 0.60


ggdensity(sites_cf2$temp)
ggdensity(sites_cf2$precip)
ggdensity(sites_cf2$aspect_gee)
ggdensity(sites_cf2$slope_gee)
#ggdensity(sites_cf2$mosaic)
ggdensity(sites_cf2$agriculture)
ggdensity(sites_cf2$grazing)
ggdensity(sites_cf2$hill_richness)
ggdensity(sites_cf2$hill_shannon)
ggdensity(sites_cf2$av_treedensity_site)
ggdensity(sites_cf2$av_loreys_site)
```

In general, data is not normal (p<0.05).
Although more important is to see if the residuals of OLS are normal or not.
Data transformation may solve the problem.

```{r}

shapiro.test(log1p(plots_cf$agb_plot_ha))
ggdensity(log1p(plots_cf$agb_plot_ha))

shapiro.test(log(sites_cf2$av_agb_site))
ggdensity(log(sites_cf2$av_agb_site))
#p = 0.07
#log(agb) works
```

## Select data

Still noisy, let's select only sites with 4 plots and none NAs

```{r}

# Sites

sites_cf_4p <- sites_cf %>%
  filter(plot_no == 4) %>%
  filter(!is.na(agriculture)) %>%
  filter(!is.na(grazing))

# Plots
# subset plots_cf data set

plots_cf <- plots_cf %>%
  mutate(site = as.character(site))

plots_cf_4p <- sites_cf_4p %>%
  select(site) %>%
  left_join(plots_cf, by ="site") #160 plots

# Trees

trees_cf_4p <- plots_cf_4p %>%
  select(plot_id) %>%
  left_join(trees_cf, by = "plot_id")

# Let's double check that the number of trees is the same

plots_cf_4p %>%
  summarise(total_trees = sum(tree_no)) %>%
  summarise(sum(total_trees))  # 4,106 trees

trees_cf_4p %>%
  filter(life_form == "Arbol" | is.na(life_form)) %>%
  filter(status == "Vivo") %>%
  # 4,106 trees
  filter(is.na(species)) #out of which 12 do not have sp name

```
Replace NAs by 0 when needed

```{r}

# plots_cf <- plots_cf %>%
#   mutate(hill_shannon = ifelse(is.na(hill_shannon), 0, hill_shannon),
#          hill_richness = ifelse(is.na(hill_richness), 0, hill_richness))
# 
# plots_cf_4p <- plots_cf_4p %>%
#   mutate(hill_shannon = ifelse(is.na(hill_shannon), 0, hill_shannon),
#          hill_richness = ifelse(is.na(hill_richness), 0, hill_richness))

```


## Save data

```{r}
# Large data sets

save(sites_cf, file = "output/sites_cf4.RData")
save(plots_cf, file = "output/plots_cf4.RData")
save(trees_cf, file = "output/trees_cf4.RData")

# 4 plots data sets

save(sites_cf_4p, file = "output/sites_cf_4p.RData")
save(plots_cf_4p, file = "output/plots_cf_4p.RData")
save(trees_cf_4p, file = "output/trees_cf_4p.RData")

write.csv2(sites_cf_4p, file = "output/sites_cf_4p.csv")
write.csv2(plots_cf_4p, file = "output/plots_cf_4p.csv")
```
