==> 05_analysis.Rmd <==
---
title: "05_analysis"
author: "Adriana Uscanga"
date: "14/2/2022"
output: html_document
---

# Analysis

Load packages and files

```{r}

# Libraries

library(tidyverse)
library(ggplot2)
library(rstatix)
library(corrplot)
library(vegan)
library(leaps)
library(asbio)
library(spdep)
library(ggpubr)
library(multcompView)
library(lme4)


# Files

load("output/sites_cf_4p.RData")
load("output/plots_cf_4p.RData")
load("output/trees_cf_4p.RData")

# Functions

calcSE<-function(x){
  x <- x[is.na(x)==F]
  sd(x)/sqrt(length(x))
}
```


## 1. Summary statistics and data exploration

Summary statistics

```{r}
summary(sites_cf_4p)
sd(sites_cf_4p$shannon)
```


Correlations

```{r}
numeric_variables <- sites_cf_4p %>%
  select(temp, precip, altitude, aspect_gee, slope_gee, sp_richness_site_av, shannon_av_site, simpson_av_site, av_agb_site, landscape, agriculture, grazing) %>%
  mutate(log_agb = log(av_agb_site))

corr <- cor(drop_na(numeric_variables[,-1]))
corrplot(corr, method = 'ellipse')
corrplot(corr, method = 'number')
plot(numeric_variables)

plot(numeric_variables$slope_gee, numeric_variables$log_agb)
plot(numeric_variables$agriculture, numeric_variables$log_agb)
plot(numeric_variables$grazing, numeric_variables$log_agb)
plot(numeric_variables$landscape, numeric_variables$log_agb)

plot(numeric_variables$agriculture, numeric_variables$landscape)
plot(numeric_variables$grazing, numeric_variables$landscape)

```


## 2. AGB distribution in TMCF

Tree size classes and contribution to AGB

```{r}

# Estimate total agb per class and plot

agb_class_plots <- trees_cf_4p %>%
  right_join(plots_cf_4p, by = "plot_id") %>%
  filter(status == "Vivo") %>%
  filter(is.na(life_form) | life_form == "Arbol") %>%
  select(plot_id, species, genus, sp, family, diam_bh, agb, str_class_plots) %>%
  mutate(diam_class = ifelse(diam_bh <= 10, "1",
                             ifelse(diam_bh > 10 & diam_bh <= 20, "2",
                                    ifelse(diam_bh > 20 & diam_bh<= 30, "3",
                                           ifelse(diam_bh > 30 & diam_bh <= 40, "4",
                                                  ifelse(diam_bh > 40 & diam_bh <= 50, "5", "6")))))) %>%
  filter(!is.na(agb)) %>%
  group_by(plot_id, diam_class) %>%
  summarise(agb_class_plot = sum(agb))

agb_class_plots
```

```{r}
# Contribution of tree size to AGB

plots_cf_4p %>%
  select(plot_id, agb_plot) %>%
  right_join(agb_class_plots, by = "plot_id") %>%
  mutate(agb_prop = agb_class_plot/agb_plot)  %>%
  group_by(diam_class) %>%
  get_summary_stats(agb_prop, type = "mean_sd") 

# Plot

plots_cf_4p %>%
  select(plot_id, agb_plot) %>%
  right_join(agb_class_plots, by = "plot_id") %>%
  mutate(agb_prop = agb_class_plot/agb_plot)  %>%
  group_by(diam_class) %>%
  ggplot(aes(x = diam_class, y = agb_prop)) +
  geom_boxplot() +
  theme_classic() +
  # stat_compare_means(comparisons = list(c("1", "2"), c("1", "3"), c("1", "4"), c("1", "5"), c("1", "6"), c("2", "3"), c("2", "4"), c("2", "5"), c("2", "6"), c("3", "4"), c("3", "5"), c("3", "6"), c("4", "5"), c("4", "6"), c("5", "6")),
  #                    label = "p.signif") +
  stat_compare_means(method = "anova",
                     label.y = 1.1) +
  geom_text(aes(x= 3.5,
                y= -0.05,
                label = "a                       b                       b                       b                       b                        c",),
            size = 4,
            color = "gray30") +
  theme_classic() +
  labs(x = "Size class",
       y = "Proportion of AGB") +
  scale_y_continuous(breaks = c(0.0, 0.25, 0.50, 0.75, 1.0))

# Anova

agb_prop <- plots_cf_4p %>%
  select(plot_id, agb_plot) %>%
  right_join(agb_class_plots, by = "plot_id") %>%
  mutate(agb_prop = agb_class_plot/agb_plot) %>%
  arrange(diam_class)

pairwise.wilcox.test(x= agb_prop$agb_prop, g= agb_prop$diam_class, p.adjust.method = "bonferroni")

anova <- aov(agb_prop ~ diam_class, data = agb_prop)
summary(anova)
TukeyHSD(anova)
tukey_hsd(anova)

multcompLetters4(anova, (TukeyHSD(anova)))  

```

Big trees (size class 6) contributes the most to AGB. Classes 1 and 6 are statistically different from the others.

Tree size classes and contribution to total number of stems:

```{r}
# Contribution of size to stems

stems_class_plot <- trees_cf_4p %>%
  right_join(plots_cf_4p, by = "plot_id") %>%
  filter(status == "Vivo") %>%
  filter(is.na(life_form) | life_form == "Arbol") %>%
  select(plot_id, species, genus, sp, family, diam_bh, agb) %>%
  mutate(diam_class = ifelse(diam_bh <= 10, "1",
                             ifelse(diam_bh > 10 & diam_bh <= 20, "2",
                                    ifelse(diam_bh > 20 & diam_bh<= 30, "3", 
                                           ifelse(diam_bh > 30 & diam_bh <= 40, "4",
                                                  ifelse(diam_bh > 40 & diam_bh <= 50, "5", "6")))))) %>%
  filter(!is.na(agb)) %>%
  group_by(plot_id, diam_class) %>%
  summarise(stems = n())

stems_class_plot
```

Contribution of tree size classes to stem density:

```{r}
plots_cf_4p %>%
  select(plot_id, tree_no) %>%
  right_join(stems_class_plot, by = "plot_id") %>%
  mutate(stem_prop = stems/tree_no)  %>%
  group_by(diam_class) %>%
  get_summary_stats(stem_prop, type = "mean_sd") 

bp <- plots_cf_4p %>%
  select(plot_id, tree_no) %>%
  right_join(stems_class_plot, by = "plot_id") %>%
  mutate(stem_prop = stems/tree_no)  %>%
  group_by(diam_class)

bp %>%
  ggplot(aes(x = diam_class, y = stem_prop)) +
  geom_boxplot() +
  stat_compare_means(method = "anova",
                     label.y = 1.1) +
  geom_text(aes(x= 3.5,
                y= -0.05,
                label = "a                       b                       c                        d                       d                      d",),
            size = 4,
            color = "gray30") + 
  theme_classic() +
  labs(x = "Size class",
       y = "Proportion of stems") +
  scale_y_continuous(breaks = c(0.0, 0.25, 0.50, 0.75, 1.0))

# Anova

stems_prop <- plots_cf_4p %>%
  select(plot_id, tree_no) %>%
  right_join(stems_class_plot, by = "plot_id") %>%
  mutate(stems_prop = stems/tree_no) %>%
  arrange(diam_class)

pairwise.wilcox.test(x= stems_prop$stems_prop, g= stems_prop$diam_class, p.adjust.method = "bonferroni")

anova2 <- aov(stems_prop ~ diam_class, data = stems_prop)
summary(anova2)
TukeyHSD(anova2)
tukey_hsd(anova2) %>%
  get_test_label()

multcompLetters4(anova2, (TukeyHSD(anova2)))
```

Trees of class 2 (10-20 cm of DBH) are the most abundant. All classes are statistically different except for classes 4, 5 and 6.

### Tree size contribution to AGB and stems in relation to successional stage

Contribution to AGB

```{r}

plots_cf_4p %>%
  select(plot_id, agb_plot, str_class_plots) %>%
  right_join(agb_class_plots, by = "plot_id") %>%
  mutate(agb_prop = agb_class_plot/agb_plot)  %>%
  mutate(succ = ifelse(str_class_plots == "1", "vy",
                       ifelse(str_class_plots == "2", "y", "m"))) %>%
  group_by(str_class_plots, diam_class) %>%
  ggplot(aes(x = diam_class, y = agb_prop, fill = factor(succ, levels = c('vy','y', 'm'),ordered = TRUE))) +
  geom_boxplot() +
  theme_classic() +
  labs(fill ='Successional Stage') +
  scale_fill_brewer(palette = "Accent")

plots_cf_4p %>%
  select(plot_id, str_class_plots) %>%
  right_join(agb_prop, by = "plot_id") %>%
  select(-site.x, -site.y) %>%
  group_by(str_class_plots, diam_class) %>%
  get_summary_stats(agb_prop, type = "mean_sd") 

agb_prop_str <- plots_cf_4p %>%
  select(plot_id, str_class_plots) %>%
  right_join(agb_prop, by = "plot_id") %>%
  select(-site.x, -site.y) %>%
  mutate(succ = ifelse(str_class_plots == "1", "s1",
                       ifelse(str_class_plots == "2", "s2", "s3"))) %>%
  group_by(str_class_plots, diam_class)

anova3 <- aov(agb_prop ~ diam_class*succ, data = agb_prop_str)
summary(anova3)
TukeyHSD(anova3)
multcompLetters4(anova3, TukeyHSD(anova3), reversed = T)

agb_vy <- agb_prop_str %>%
  ungroup() %>%
  filter(succ == "vy") 

# anova5 <- aov(agb_prop ~ diam_class, data = agb_vy)
# summary(anova5)
# TukeyHSD(anova5)

# multcompLetters4(anova5, (TukeyHSD(anova5)), reversed = T)

# agb_y <- agb_prop_str %>%
#   ungroup() %>%
#   filter(succ == "y") 
# 
# anova6 <- aov(agb_prop ~ diam_class, data = agb_y)
# summary(anova6)
# TukeyHSD(anova6)
# 
# multcompLetters4(anova6, (TukeyHSD(anova6)), reversed = T)
```

Contribution to Stem Number

```{r}
plots_cf_4p %>%
  select(plot_id, tree_no, str_class_plots) %>%
  right_join(stems_class_plot, by = "plot_id") %>%
  mutate(stem_prop = stems/tree_no)  %>%
  mutate(succ = ifelse(str_class_plots == "1", "vy",
                       ifelse(str_class_plots == "2", "y", "m"))) %>%
  group_by(str_class_plots, diam_class) %>%
  ggplot(aes(x = diam_class, y = stem_prop, fill = factor(succ, levels = c('vy','y', 'm'),ordered = TRUE))) +
  geom_boxplot() +
  theme_classic() +
  labs(fill ='Successional Stage') +
  scale_fill_brewer(palette = "Accent")

plots_cf_4p %>%
  select(plot_id, str_class_plots) %>%
  right_join(stems_prop, by = "plot_id") %>%
  select(-site.x, -site.y) %>%
  group_by(str_class_plots, diam_class) %>%
  get_summary_stats(stems_prop, type = "mean_sd") 

stem_prop_str <- plots_cf_4p %>%
  select(plot_id, str_class_plots) %>%
  right_join(stems_prop, by = "plot_id") %>%
  select(-site.x, -site.y) %>%
  mutate(succ = ifelse(str_class_plots == "1", "vy",
                       ifelse(str_class_plots == "2", "y", "m"))) %>%
  group_by(str_class_plots, diam_class)

anova4 <- aov(stems_prop ~ succ * diam_class, data = stem_prop_str)
summary(anova4)
TukeyHSD(anova4)
```

### Structural attributes and AGB in three successional stages

```{r}

# Tree density

plots_cf_4p %>%
  mutate(succ = ifelse(str_class_plots == "1", "vy",
                       ifelse(str_class_plots == "2", "y", "m"))) %>%
  group_by(succ) %>%
  get_summary_stats(tree_density, type = "mean_sd")

summary(aov(tree_density ~ str_class_plots, data = plots_cf_4p))
TukeyHSD(aov(tree_density ~ str_class_plots, data = plots_cf_4p))
tukey_hsd(aov(tree_density ~ str_class_plots, data = plots_cf_4p))

# Tree height

plots_cf_4p %>%
  mutate(succ = ifelse(str_class_plots == "1", "vy",
                       ifelse(str_class_plots == "2", "y", "m"))) %>%
  group_by(succ) %>%
  get_summary_stats(loreys_height, type = "mean_sd")

summary(aov(loreys_height ~ str_class_plots, data = plots_cf_4p))
tukey_hsd(aov(loreys_height ~ str_class_plots, data = plots_cf_4p))
TukeyHSD(aov(loreys_height ~ str_class_plots, data = plots_cf_4p))

# Basal area

plots_cf_4p %>%
  mutate(succ = ifelse(str_class_plots == "1", "vy",
                       ifelse(str_class_plots == "2", "y", "m"))) %>%
  group_by(succ) %>%
  get_summary_stats(basal_area_ha, type = "mean_sd")

summary(aov(basal_area_ha ~ str_class_plots, data = plots_cf_4p))
tukey_hsd(aov(basal_area_ha ~ str_class_plots, data = plots_cf_4p))
TukeyHSD(aov(basal_area_ha ~ str_class_plots, data = plots_cf_4p))

# AGB

plots_cf_4p %>%
  mutate(succ = ifelse(str_class_plots == "1", "vy",
                       ifelse(str_class_plots == "2", "y", "m"))) %>%
  group_by(succ) %>%
  get_summary_stats(agb_plot_ha, type = "mean_sd")

summary(aov(agb_plot_ha ~ str_class_plots, data = plots_cf_4p))
TukeyHSD(aov(agb_plot_ha ~ str_class_plots, data = plots_cf_4p))
tukey_hsd(aov(agb_plot_ha ~ str_class_plots, data = plots_cf_4p))

# plots_cf_4p %>%
#   ungroup() %>%
#   anova_test(agb_plot_ha ~ str_class_plots, detailed = T, options(contrasts=c('contr.sum','contr.poly')))

plots_cf_4p %>%
  ungroup() %>%
  select(agb_plot_ha, str_class_plots) %>%
  filter(!is.na(agb_plot_ha), !is.na(str_class_plots)) %>%
  t_test(agb_plot_ha ~ str_class_plots, p.adjust.method = "bonferroni") %>%
  add_significance("p.adj") 
```


## 3. AGB patterns along environmental and land-use gradients

### PCA of environmental variables

```{r}
env <- sites_cf_4p %>%
  select(temp, precip, altitude, slope_gee)
env <- as.data.frame(env)
env$site -> row.names(env)
env <- env[,-1]

pca_env <- rda(env, scale = T)
summary(pca_env)
plot(pca_env)
biplot(pca_env)

```

PC1 explains 68.23% of variation and is correlated to temp, altitude and precip.

```{r}
pc1 <- as.data.frame(pca_env$CA$u[,1])
colnames(pc1) <- "env_pc1"
pc1

plot(pc1[,1], sites_cf_4p$precip)
plot(pc1[,1], sites_cf_4p$temp)
plot(pc1[,1], sites_cf_4p$altitude)

summary(lm(pc1[,1] ~ sites_cf_4p$precip))
summary(lm(pc1[,1] ~ sites_cf_4p$temp))
summary(lm(pc1[,1] ~ sites_cf_4p$altitude))
```

### Multiple linear regression model with all variables:

```{r}
# Add PC1 to data set

class(pc1)

env_pc1 <- as_tibble(pc1, .name_repair = "minimal", rownames = "site")


sites_cf_4p <- sites_cf_4p %>%
  left_join(env_pc1, by = "site")
  
# Run regression model with env_pc1

lm <- lm(log(av_agb_site) ~ env_pc1 + slope_gee + aspect_gee_c + shannon + sp_richness_site + landscape + agriculture + grazing, data = sites_cf_4p)
summary(lm)
```

### Stepwise model selection:

```{r}
ms <- regsubsets(log(av_agb_site) ~ env_pc1 + slope_gee + shannon + sp_richness_site + landscape + agriculture + grazing, data = sites_cf_4p)

ms_sum <- summary(ms)

# Best model:

data.frame(
  Adj.R2 = (ms_sum$adjr2),
  CP = (ms_sum$cp),
  BIC = (ms_sum$bic)
)

data.frame(
  Adj.R2 = which.max(ms_sum$adjr2),
  CP = which.min(ms_sum$cp),
  BIC = which.min(ms_sum$bic)
)

ms_sum

# Best model according to BIC has 3 predictors: 
#  slope_gee
#  landscape
#  agriculture

#Best model according to CP has four:
# Same but adding env_pc1
```

Best models:

```{r}
lm2 <- lm(log(av_agb_site) ~ slope_gee + landscape + agriculture, data = sites_cf_4p)
summary(lm2)
partial.resid.plot(lm2)

lm2_res <- lm2$residuals #save residuals

lm2b <- lm(log(av_agb_site) ~ slope_gee + landscape + agriculture + env_pc1, data = sites_cf_4p)
summary(lm2b)
partial.resid.plot(lm2b)

lm2b_res <- lm2b$residuals #save residuals

lm2c <- lm(log(av_agb_site) ~ slope_gee + landscape + agriculture + env_pc1 + grazing + shannon, data = sites_cf_4p)
summary(lm2c)
partial.resid.plot(lm2c)

lm2c_res <- lm2c$residuals #save residuals

summary(lm(landscape ~ shannon, data = sites_cf_4p))
```
P value comparisons of all models:

```{r}

# 1 predictor
summary(lm(log(av_agb_site) ~ landscape, data = sites_cf_4p))

# 2 predictors
summary(lm(log(av_agb_site) ~ landscape + agriculture, data = sites_cf_4p))

# 3 predictors
summary(lm(log(av_agb_site) ~ slope_gee + landscape + agriculture, data = sites_cf_4p))

# 4 predictors
summary(lm(log(av_agb_site) ~ slope_gee + landscape + agriculture + env_pc1, data = sites_cf_4p))

# 5 predicotrs
summary(lm(log(av_agb_site) ~ slope_gee + landscape + agriculture + env_pc1 + grazing, data = sites_cf_4p))

# 6 predictors
summary(lm(log(av_agb_site) ~ slope_gee + landscape + agriculture + env_pc1 + grazing + shannon, data = sites_cf_4p))

# 7 predictors
summary(lm(log(av_agb_site) ~ slope_gee + landscape + agriculture + env_pc1 + grazing + shannon + sp_richness_site, data = sites_cf_4p))
```


### Spatial autocorrelation test:

```{r}

sites_coords_4p <- sites_cf_4p %>%
  dplyr::select(longitude, latitude)

#distance neighbors
#loc_subs <- cbind(sites_coords_4p$longitude, sites_coords_4p$latitude)
loc_subs <- cbind(plots_cf_4p$longitude, plots_cf_4p$latitude)
d_subs <- 10 #distance has to be larger than the distance that exists between sites, which is 2.5 km
neighbors_dist <- dnearneigh(loc_subs, 0, d_subs, longlat = T)
str(neighbors_dist)
plot(neighbors_dist, loc_subs)

# Extract model residuals
model_res <- gamm4_4_int$gam$residuals

#Moran test
# spatial autocorrelation of model residuals
moran.test(model_res, zero.policy=T, 
     nb2listw(neighbors_dist, zero.policy=T, style="W"))

# p-value = 0.5176 with d= 10 Moran I= -0.04971499
# p-value = 0.6049 with d= 25 Moran I= -0.050440447
# p-value = 0.1875 with d= 50 Moran I= 0.020220697
# No spatial autocorrelation!
```

There's no spatial autocorrelation.

Map of residuals:

```{r}
sites_coords_4p %>%
  ggplot(aes(x = longitude, y = latitude)) +
  geom_point(aes(color = lm2_res), size = 4) +
  scale_colour_gradient2(low = "#d8b365", mid= "#f5f5f5", high = "#5ab4ac") +
  theme_classic()

```

### Why is environmental gradient not relevant?

```{r}
lm3 <- lm(log(av_agb_site) ~ env_pc1, data = sites_cf_4p)
summary(lm3)

lm4 <- lm(landscape ~ env_pc1, data = sites_cf_4p)
summary(lm4)

lm5 <- lm(log(av_agb_site) ~landscape, data = sites_cf_4p)
summary(lm5)

ggplot(sites_cf_4p, aes(x = env_pc1, y = log(av_agb_site), color = landscape)) +
  geom_point(size = 4) +
  scale_color_gradient(low = "darkgreen", high = "yellow") +
  geom_smooth(method = "lm", color = "darkgray")

ggplot(sites_cf_4p, aes(x = env_pc1, y = landscape, color = av_agb_site)) +
  geom_point(size = 4) +
  scale_color_gradient(high = "darkgreen", low = "yellow") +
  geom_smooth(method = "lm", color = "darkgray")

```

It is relevant but completely overshadowed by landscape composition. 
Also, landscape composition is correlated to environmental gradient (adjusted R2= 0.372) thus there's little variation left to be explained by the environment. 
Moreover, environment is already restricted because all sites are constrained to TMCF.


## 4. AGB and tree diversity

### Loess curves

Species richness 

```{r}

# Sp richness

lo1 <- loess(sp_richness_site_av ~ log(av_agb_site), data = sites_cf_4p)
summary(lo1)

# poor man's R-squared value
cor(sites_cf_4p$sp_richness_site_av, lo1$fitted)^2

sites_cf_4p %>%
  ggplot(aes(x = av_agb_site, y = sp_richness_site_av)) +
  geom_smooth(alpha = 0.2, method = "loess", span = 1, color= "gray") +
  geom_point(size=2) +
  theme_classic()

```

Simpson 


```{r}

lo2 <- loess(simpson_av_site ~ log(av_agb_site), data = sites_cf_4p)
summary(lo2)

# poor man's R-squared value
cor(sites_cf_4p$simpson_av_site, lo2$fitted)^2

sites_cf_4p %>%
  ggplot(aes(x = av_agb_site, y = simpson_av_site)) +
  geom_point(size=2) +
  geom_smooth(alpha = 0.2, method = "loess", span = 1, color= "gray") +
  theme_classic()

```

Stem density


```{r}

# Stem density ~ sp richness

lo3 <- loess(sp_richness_site_av ~ av_treedensity_site, data = sites_cf_4p)
summary(lo3)
# poor man's R-squared value
cor(sites_cf_4p$sp_richness_site_av, lo3$fitted)^2

sites_cf_4p %>%
  ggplot(aes(x = av_treedensity_site, y = sp_richness_site_av)) +
  geom_smooth(alpha = 0.2, method = "loess", span = 1, color= "gray") +
  #geom_smooth(method = "lm", alpha = 0.2) +
  geom_point(size=2) +
  theme_classic()


```

Stem density ~ simpson

```{r}
lo4 <- loess(simpson_av_site ~ av_treedensity_site, data = sites_cf_4p)
summary(lo4)
# poor man's R-squared value
cor(sites_cf_4p$simpson_av_site, lo4$fitted)^2

sites_cf_4p %>%
  ggplot(aes(x = av_treedensity_site, y = simpson_av_site, color = landscape)) +
  geom_smooth(alpha = 0.1, method = "loess", span = 0.75, color= "gray", linetype = "dashed") +
  #geom_smooth(method = "lm", alpha = 0.2) +
  geom_point(size= 3) +
  theme_classic() +
  scale_colour_gradient(high = "darkolivegreen2", low = "darkgreen") 

sites_cf_4p %>%
  ggplot(aes(x = av_treedensity_site, y = shannon_av_site, color = landscape)) +
  geom_smooth(alpha = 0.1, method = "loess", span = 0.75, color= "gray", linetype = "dashed") +
  #geom_smooth(method = "lm", alpha = 0.2) +
  geom_point(size= 3) +
  theme_classic() +
  scale_colour_gradient(high = "darkolivegreen2", low = "darkgreen") 
```

At plot level

```{r}

plots_cf_4p %>%
  filter(agb_plot_ha < 600) %>%
  filter(!is.na(simpson)) %>%
  filter(!is.na(agb_plot_ha)) %>%
  group_by(str_class_plots) %>%
  summarise(mean_simpson = mean(simpson),
            mean_agb = mean(agb_plot_ha),
            agb_plot_ha = as.numeric(agb_plot_ha),
            simpson = as.numeric(simpson)) %>%
  ggplot(aes(x = agb_plot_ha, y = simpson)) +
  geom_smooth(method = "loess", span = 1, color = "gray", se= T, alpha = 0.1) +
  geom_point(aes(x= mean_agb, y = mean_simpson, color = str_class_plots, size = 3)) +
  geom_point(aes(color = str_class_plots), size= 1.5, alpha = 0.3) +
  theme_classic() +
  scale_color_brewer(palette = "Accent") +
  labs(color = "Successional Stage",
       size = "Average")

plots_cf_4p %>%
  filter(agb_plot_ha < 600) %>%
  filter(!is.na(simpson)) %>%
  filter(!is.na(agb_plot_ha)) %>%
  group_by(str_class_plots) %>%
  summarise(mean_shannon = mean(shannon),
            mean_agb = mean(agb_plot_ha),
            agb_plot_ha = as.numeric(agb_plot_ha),
            shannon = as.numeric(shannon)) %>%
  ggplot(aes(x = agb_plot_ha, y = shannon)) +
  geom_smooth(method = "loess", span = 1, color = "gray", se= T, alpha = 0.1) +
  geom_point(aes(x= mean_agb, y = mean_shannon, color = str_class_plots, size = 3)) +
  geom_point(aes(color = str_class_plots), size= 1.5, alpha = 0.3) +
  theme_classic() +
  scale_color_brewer(palette = "Accent") +
  labs(color = "Successional Stage",
       size = "Average")

plots_cf_4p %>%
  filter(agb_plot_ha < 600) %>%
  filter(!is.na(simpson)) %>%
  filter(!is.na(agb_plot_ha)) %>%
  group_by(str_class_plots) %>%
  summarise(mean_sp = mean(sp_richness_plot),
            mean_agb = mean(agb_plot_ha),
            agb_plot_ha = as.numeric(agb_plot_ha),
            sp_richness_plot = as.numeric(sp_richness_plot)) %>%
  ggplot(aes(x = agb_plot_ha, y = sp_richness_plot, color = str_class_plots)) +
  #geom_smooth(method = "lm", se =F) +
  geom_smooth(method = "loess", span = 1, color = "gray", se= T, alpha = 0.1) +
  geom_point(aes(x= mean_agb, y = mean_sp, color = str_class_plots, size = 3)) +
  geom_point(aes(color = str_class_plots), size= 1.5, alpha = 0.5) +
  theme_classic() +
  scale_color_brewer(palette = "Accent") +
  labs(color = "Successional Stage",
       size = "Average")



```
```{r}
# Total number of spp

head(trees_cf_4p)
  
trees_cf_4p %>%
  group_by(species) %>%
  summarise(list = unique(species), count = n()) %>%
  arrange(-count)

trees_cf_4p %>%
  #filter(is.na(species)) %>%
  filter(status == "Vivo") %>%
  filter(is.na(agb_ae))

# Wood density at site level

trees_cf_4p %>%
  filter(!is.na(mean_wd)) %>%
  group_by(site) %>%
  summarise(mean_wd_site = mean(mean_wd),
            se_wd_site = calcSE(mean_wd)) %>%
  arrange(-mean_wd_site) %>%
  mutate(average = mean(mean_wd_site),
         se = calcSE(mean_wd_site))

mean(trees_cf_4p$mean_wd, na.rm = T)
sd(trees_cf_4p$mean_wd, na.rm = T)

# Wood density at plot level

trees_cf_4p %>%
  filter(!is.na(mean_wd)) %>%
  group_by(plot_id) %>%
  summarise(mean_wd_plot = mean(mean_wd),
            sd_wd_plot = sd(mean_wd)) %>%
  right_join(plots_cf_4p) %>%
  mutate(mean_wd_plot = ifelse(tree_no == 0, 0, mean_wd_plot),
         sd_wd_plot = ifelse(tree_no == 0, 0, sd_wd_plot)) %>%
  #group_by(str_class_plots) %>%
  # summarise(mean_wd_str = mean(mean_wd_plot),
  #           sd_wd_str = sd(mean_wd_plot)) %>%
  tukey_hsd(mean_wd_plot ~ str_class_plots)

# Ordination of plots according to their structural attributes

plots_str <- trees_cf_4p %>%
  filter(!is.na(mean_wd)) %>%
  group_by(plot_id) %>%
  summarise(mean_wd_plot = mean(mean_wd),
            sd_wd_plot = sd(mean_wd)) %>%
  right_join(plots_cf_4p) %>%
  mutate(mean_wd_plot = ifelse(tree_no == 0, 0, mean_wd_plot),
         sd_wd_plot = ifelse(tree_no == 0, 0, sd_wd_plot)) %>%
  select(plot_id,
         tree_density,
         loreys_height,
         basal_area_ha,
         mean_wd_plot,
         agb_plot_ha)


plot(plots_str)

# Scale data so that variables in different units can be compared
plots_str_scaled <- scale(plots_str[,-c(1,5,6)])
row.names(plots_str_scaled) <- plots_str$plot_id

# Save a vector of 3 clusters:
str_k3_plots <- plots_cf_4p$str_class_plots
as.numeric(str_k3_plots)

#pca with structural data

pca_str_plots <- rda(plots_str_scaled, scale = F)
summary(pca_str_plots) # 
plot(pca_str_plots)

ordiplot(pca_str_plots, display = 'sites', type = 'n')
points(pca_str_plots, col = plots_cf_4p$str_class_plots)
biplot(pca_str_plots, display = 'species')

pca_scores <- scores(pca_str_plots)
pca_scores_plots <- as.data.frame(pca_scores$sites)
pca_scores_plots %>%
  as_tibble(rownames = "plot_id") %>%
  left_join(plots_cf_4p) %>%
  ggplot(aes(x = PC1, y = PC2)) +
  geom_point(aes(color = str_class_plots))

pca_scores$species

```

Proportion of AGB by successional stage

```{r}

plots_cf_4p %>%
  group_by(str_class_plots) %>%
  summarise(agb_total = sum(agb_plot_ha))

# AGB total:
#2303.270+9404.818+10291.113 =
#21999.2

plots_cf_4p %>%
  group_by(str_class_plots) %>%
  summarise(agb_total = sum(agb_plot_ha)) %>%
  mutate(per100 = (agb_total / 21999.2) *100)

```

```{r}
plots_cf_4p %>% filter(str_class_plots == "2") %>% arrange(-agb_plot_ha)
```
