==> 05b_analysis.Rmd <==
---
title: "05b_analysis"
author: "Adriana Uscanga"
date: "2023-02-07"
output: html_document
---

# Analysis: second round

Load packages and files

```{r}
# Files

load("output/sites_cf_4p.RData")
load("output/plots_cf_4p.RData")
load("output/trees_cf_4p.RData")

# Libraries

library(tidyverse)
library(ggplot2)
library(lme4)
library(corrplot)
library(vegan)
library(stargazer)
library(sjPlot)
library(cAIC4)
library(mgcv)
library(gamm4)
library(multcomp)
library(car)

library(rstatix)
#library(corrplot)
#library(vegan)
#library(leaps)
#library(asbio)
#library(spdep)
#library(ggpubr)
#library(multcompView)
```

## 1. Summary statistics and data exploration

```{r}

summary(plots_cf_4p)
summary(sites_cf_4p)

plots_cf_4p %>%
  ggplot(aes(x= agb_plot_ha)) +
  geom_histogram(bins = 25)

plots_cf_4p %>%
  ggplot(aes(x= log1p(agb_plot_ha))) +
  geom_histogram(bins = 20)

plots_cf_4p %>%
  ggplot(aes(x= sqrt(agb_plot_ha))) +
  geom_histogram(bins = 20)

plots_cf_4p %>%
  ggplot(aes(x= hill_shannon)) +
  geom_histogram(bins = 20)

plots_cf_4p %>%
  ggplot(aes(x= hill_richness)) +
  geom_histogram(bins = 10)

```
### Correlations

```{r}
numeric_variables <- plots_cf_4p %>%
  dplyr::select(temp, precip, ai, altitude, latitude, longitude, northness, eastness, slope, hill_shannon, hill_richness, agb_plot_ha, agriculture, grazing, logging) %>%
  mutate(log_agb = log1p(agb_plot_ha)) %>%
  dplyr::select(-agb_plot_ha)

corr <- cor(drop_na(numeric_variables[,-1]))
corrplot(corr, method = 'ellipse')
corrplot(corr, method = 'number')
plot(numeric_variables)
```
From this correlation matrix, I'm going to select variables that have correlation > 0.3
Environmental data is highly correlated, so I'm going to get rid of some of that collinearity with a PCA

## 2. PCA of environmental variables

```{r}
env <- plots_cf_4p %>%
  dplyr::select(plot_id, temp, precip, ai, altitude, northness, eastness, slope)
env <- as.data.frame(env)
env$plot_id -> row.names(env)
env <- env[,c(-1, -2)]

pca_env <- rda(env, scale = T)
summary(pca_env)
plot(pca_env)
biplot(pca_env, display = c('species', 'sites'), type = c('text', 'points'), col = c('gray', 'black'))
biplot(pca_env)
# PC1 (0.43) and PC2 (0.19) together explain 0.62 of variation

```

PC1:

```{r}
pc1 <- as.data.frame(pca_env$CA$u[,1])
colnames(pc1) <- "env_pc1"
pc1

plot(pc1[,1], plots_cf_4p$precip)
plot(pc1[,1], plots_cf_4p$ai)
plot(pc1[,1], plots_cf_4p$temp)
plot(pc1[,1], plots_cf_4p$altitude)

summary(lm(pc1[,1] ~ plots_cf_4p$precip))
summary(lm(pc1[,1] ~ plots_cf_4p$ai))
summary(lm(pc1[,1] ~ plots_cf_4p$temp))
summary(lm(pc1[,1] ~ plots_cf_4p$altitude))
```

PC2:

```{r}
pc2 <- as.data.frame(pca_env$CA$u[,2])
colnames(pc2) <- "env_pc2"
pc2

plot(pc2[,1], plots_cf_4p$slope)
plot(pc2[,1], plots_cf_4p$northness)
plot(pc2[,1], plots_cf_4p$eastness)
#plot(pc2[,1], plots_cf_4p$latitude)
#plot(pc2[,1], plots_cf_4p$longitude)

summary(lm(pc2[,1] ~ plots_cf_4p$slope))
summary(lm(pc2[,1] ~ plots_cf_4p$northness))
summary(lm(pc2[,1] ~ plots_cf_4p$eastness))
#summary(lm(pc2[,1] ~ plots_cf_4p$latitude))
#summary(lm(pc2[,1] ~ plots_cf_4p$longitude))

```
Slope and aspect are more related to PC2 but relationships are weak.
Correlation among them is also weak. 
When lat and lon included, PC2 is mainly driven by lat and lon.
To avoid collinearity, we are going to use PC1 (which encompasses temp, precip, elevation, and ai), and slope.
Lat, lon, and aspect don't seem to be related to other variables (see corrplot).

```{r}
# Add PC1 values to dataset

class(pc1)

env_pc1 <- as_tibble(pc1, .name_repair = "minimal", rownames = "plot_id")


plots_cf_4p <- plots_cf_4p %>%
  left_join(env_pc1, by = "plot_id")

plots_cf_4p <- plots_cf_4p %>%
  relocate(env_pc1, .after = ai)

```

Once we addressed collinearity, and checked which variables are correlated,
the selected variables that will be predictors in the model are:
* env_pc1
* slope (?)
* hill_shannon
* hill_richness
* agriculture
* grazing
* logging (?)
* log_agb

(?) are not related to agb nor tree diversity but are correlated.
Also, slope is always included in agb models of tropical montane forests.

## 3. Linear Mixed Effects Models

We are going to fit models for predicting AGB using the plots_cf_4p dataset
and using the variable 'site' as random factor.

1) First fit a model of agb with each predictor
2) Test all possible models and select the best ones
3) Test interactions between predictors and assess whether adding
interactions in the model is worth it 

### LMM: each predictor

```{r}

lmm_env <- lmer(log1p(agb_plot_ha) ~ env_pc1 + (1| site), data = plots_cf_4p)
summary(lmm_env)
plot(lmm_env)
stargazer(lmm_env, type = "text")
cAIC(lmm_env)

```

```{r}
lmm_slope <- lmer(log1p(agb_plot_ha) ~ slope + (1| site), data = plots_cf_4p)
summary(lmm_slope)
plot(lmm_slope)
stargazer(lmm_slope, type = "text")
cAIC(lmm_slope)
```

```{r}

lmm_hill_shannon <- lmer(log1p(agb_plot_ha) ~ hill_shannon + (1| site), data = plots_cf_4p)
summary(lmm_hill_shannon)
plot(lmm_hill_shannon)
stargazer(lmm_hill_shannon, type = "text")
cAIC(lmm_hill_shannon)

```


```{r}

lmm_sp_rich <- lmer(log1p(agb_plot_ha) ~ hill_richness + (1| site), data = plots_cf_4p)
summary(lmm_sp_rich)
plot(lmm_sp_rich)
stargazer(lmm_sp_rich, type = "text")
cAIC(lmm_sp_rich)

```


```{r}

lmm_ag <- lmer(log1p(agb_plot_ha) ~ agriculture + (1| site), data = plots_cf_4p)
summary(lmm_ag)
plot(lmm_ag)
stargazer(lmm_ag, type = "text")
cAIC(lmm_ag)

```
```{r}

lmm_graz <- lmer(log1p(agb_plot_ha) ~ grazing + (1| site), data = plots_cf_4p)
summary(lmm_graz)
plot(lmm_graz)
stargazer(lmm_graz, type = "text")
cAIC(lmm_graz)

```
```{r}

lmm_logg <- lmer(log1p(agb_plot_ha) ~ logging + (1| site), data = plots_cf_4p)
summary(lmm_logg)
plot(lmm_logg)
stargazer(lmm_logg, type = "text")
cAIC(lmm_logg)

```

```{r}
anocAIC(lmm_env, lmm_slope, lmm_hill_shannon, lmm_sp_rich, lmm_ag, lmm_graz, lmm_logg, digits = 3)
```

Best model with only 1 predictor is agb ~ sp_richness!

```{r}

lmm_all <- lmer(log1p(agb_plot_ha) ~ env_pc1 +
           slope +
           hill_shannon +
           hill_richness +
           agriculture +
           grazing +
           logging +
           (1| site), data = plots_cf_4p)

lmm_all6 <- lmer(log1p(agb_plot_ha) ~ env_pc1 +
           slope +
           hill_shannon +
           hill_richness +
           agriculture +
           grazing +
           #logging +
           (1| site), data = plots_cf_4p)

lmm_all5 <- lmer(log1p(agb_plot_ha) ~ env_pc1 +
           slope +
           hill_shannon +
           hill_richness +
           agriculture +
           #grazing +
           #logging +
           (1| site), data = plots_cf_4p)

lmm_all4 <- lmer(log1p(agb_plot_ha) ~ env_pc1 +
           slope +
           #hill_shannon +
           hill_richness +
           agriculture +
           #grazing +
           #logging +
           (1| site), data = plots_cf_4p)


anocAIC(lmm_all, lmm_all6, lmm_all5, lmm_all4, digits = 3)
```

### LMM: all posible models

```{r}
# Two predictors
combn(c("env","slope","sp_rich", "hill_shannon","ag","graz","logg"), 2, simplify = T)

# env_pc1

lmm_2p_a <- lmer(log1p(agb_plot_ha) ~ env_pc1 +
           slope +
           #hill_shannon +
           #hill_richness +
           #agriculture +
           #grazing +
           #logging +
           (1| site), data = plots_cf_4p)

lmm_2p_b <- lmer(log1p(agb_plot_ha) ~ env_pc1 +
           #slope +
           hill_shannon +
           #hill_richness +
           #agriculture +
           #grazing +
           #logging +
           (1| site), data = plots_cf_4p)

lmm_2p_c <- lmer(log1p(agb_plot_ha) ~ env_pc1 +
           #slope +
           #hill_shannon +
           hill_richness +
           #agriculture +
           #grazing +
           #logging +
           (1| site), data = plots_cf_4p)

lmm_2p_d <- lmer(log1p(agb_plot_ha) ~ env_pc1 +
           #slope +
           #hill_shannon +
           #hill_richness +
           agriculture +
           #grazing +
           #logging +
           (1| site), data = plots_cf_4p)

lmm_2p_e <- lmer(log1p(agb_plot_ha) ~ env_pc1 +
           #slope +
           #hill_shannon +
           #hill_richness +
           #agriculture +
           grazing +
           #logging +
           (1| site), data = plots_cf_4p)

lmm_2p_f <- lmer(log1p(agb_plot_ha) ~ env_pc1 +
           #slope +
           #hill_shannon +
           #hill_richness +
           #agriculture +
           #grazing +
           logging +
           (1| site), data = plots_cf_4p)

# Slope 

lmm_2p_g <- lmer(log1p(agb_plot_ha) ~ #env_pc1 +
           slope +
           hill_shannon +
           #hill_richness +
           #agriculture +
           #grazing +
           #logging +
           (1| site), data = plots_cf_4p)

lmm_2p_h <- lmer(log1p(agb_plot_ha) ~ #env_pc1 +
           slope +
           #hill_shannon +
           hill_richness +
           #agriculture +
           #grazing +
           #logging +
           (1| site), data = plots_cf_4p)

lmm_2p_i <- lmer(log1p(agb_plot_ha) ~ #env_pc1 +
           slope +
           #hill_shannon +
           #hill_richness +
           agriculture +
           #grazing +
           #logging +
           (1| site), data = plots_cf_4p)

lmm_2p_j <- lmer(log1p(agb_plot_ha) ~ #env_pc1 +
           slope +
           #hill_shannon +
           #hill_richness +
           #agriculture +
           grazing +
           #logging +
           (1| site), data = plots_cf_4p)

lmm_2p_k <- lmer(log1p(agb_plot_ha) ~ #env_pc1 +
           slope +
           #hill_shannon +
           #hill_richness +
           #agriculture +
           #grazing +
           logging +
           (1| site), data = plots_cf_4p)

# hill_shannon

lmm_2p_l <- lmer(log1p(agb_plot_ha) ~ #env_pc1 +
           #slope +
           hill_shannon +
           hill_richness +
           #agriculture +
           #grazing +
           #logging +
           (1| site), data = plots_cf_4p)

lmm_2p_m <- lmer(log1p(agb_plot_ha) ~ #env_pc1 +
           #slope +
           hill_shannon +
           #hill_richness +
           agriculture +
           #grazing +
           #logging +
           (1| site), data = plots_cf_4p)

lmm_2p_n <- lmer(log1p(agb_plot_ha) ~ #env_pc1 +
           #slope +
           hill_shannon +
           #hill_richness +
           #agriculture +
           grazing +
           #logging +
           (1| site), data = plots_cf_4p)

lmm_2p_o <- lmer(log1p(agb_plot_ha) ~ #env_pc1 +
           #slope +
           hill_shannon +
           #hill_richness +
           #agriculture +
           #grazing +
           logging +
           (1| site), data = plots_cf_4p)

# hill_richness

lmm_2p_p <- lmer(log1p(agb_plot_ha) ~ #env_pc1 +
           #slope +
           #hill_shannon +
           hill_richness +
           agriculture +
           #grazing +
           #logging +
           (1| site), data = plots_cf_4p)

lmm_2p_q <- lmer(log1p(agb_plot_ha) ~ #env_pc1 +
           #slope +
           #hill_shannon +
           hill_richness +
           #agriculture +
           grazing +
           #logging +
           (1| site), data = plots_cf_4p)

lmm_2p_r <- lmer(log1p(agb_plot_ha) ~ #env_pc1 +
           #slope +
           #hill_shannon +
           hill_richness +
           #agriculture +
           #grazing +
           logging +
           (1| site), data = plots_cf_4p)

# Agriculture

lmm_2p_s <- lmer(log1p(agb_plot_ha) ~ #env_pc1 +
           #slope +
           #hill_shannon +
           #hill_richness +
           agriculture +
           grazing +
           #logging +
           (1| site), data = plots_cf_4p)

lmm_2p_t <- lmer(log1p(agb_plot_ha) ~ #env_pc1 +
           #slope +
           #hill_shannon +
           #hill_richness +
           agriculture +
           #grazing +
           logging +
           (1| site), data = plots_cf_4p)

# Grazing

lmm_2p_u <- lmer(log1p(agb_plot_ha) ~ #env_pc1 +
           #slope +
           #hill_shannon +
           #hill_richness +
           #agriculture +
           grazing +
           logging +
           (1| site), data = plots_cf_4p)

lmm2p_cAIC <- anocAIC(lmm_2p_a,
        lmm_2p_b,
        lmm_2p_c,
        lmm_2p_d,
        lmm_2p_e,
        lmm_2p_f,
        lmm_2p_g,
        lmm_2p_h,
        lmm_2p_i,
        lmm_2p_j,
        lmm_2p_k,
        lmm_2p_l,
        lmm_2p_m,
        lmm_2p_n,
        lmm_2p_o,
        lmm_2p_p,
        lmm_2p_q,
        lmm_2p_r,
        lmm_2p_s,
        lmm_2p_t,
        lmm_2p_u,
        digits = 3)

lmm2p_cAIC <- as_tibble(lmm2p_cAIC, rownames = "model")

lmm2p_cAIC %>%
  arrange(cAIC)
```
Best model with 2 predictors: sp_richness + agriculture cAIC 446.844
I'll remove hill_shannon because sp_richness seems to work better and they're very related

```{r}
# 3 predictors
combn(c("env","slope","sp_rich","ag","graz","logg"), 3, simplify = T)

# env_pc1

lmm_3p_a <- lmer(log1p(agb_plot_ha) ~ env_pc1 +
           slope +
           hill_richness +
           #agriculture +
           #grazing +
           #logging +
           (1| site), data = plots_cf_4p)

lmm_3p_b <- lmer(log1p(agb_plot_ha) ~ env_pc1 +
           slope +
           #hill_richness +
           agriculture +
           #grazing +
           #logging +
           (1| site), data = plots_cf_4p)

lmm_3p_c <- lmer(log1p(agb_plot_ha) ~ env_pc1 +
           slope +
           #hill_richness +
           #agriculture +
           grazing +
           #logging +
           (1| site), data = plots_cf_4p)

lmm_3p_d <- lmer(log1p(agb_plot_ha) ~ env_pc1 +
           slope +
           #hill_richness +
           #agriculture +
           #grazing +
           logging +
           (1| site), data = plots_cf_4p)

lmm_3p_e <- lmer(log1p(agb_plot_ha) ~ env_pc1 +
           #slope +
           hill_richness +
           agriculture +
           #grazing +
           #logging +
           (1| site), data = plots_cf_4p)

lmm_3p_f <- lmer(log1p(agb_plot_ha) ~ env_pc1 +
           #slope +
           hill_richness +
           #agriculture +
           grazing +
           #logging +
           (1| site), data = plots_cf_4p)

lmm_3p_g <- lmer(log1p(agb_plot_ha) ~ env_pc1 +
           #slope +
           hill_richness +
           #agriculture +
           #grazing +
           logging +
           (1| site), data = plots_cf_4p)

lmm_3p_h <- lmer(log1p(agb_plot_ha) ~ env_pc1 +
           #slope +
           #hill_richness +
           agriculture +
           grazing +
           #logging +
           (1| site), data = plots_cf_4p)

lmm_3p_i <- lmer(log1p(agb_plot_ha) ~ env_pc1 +
           #slope +
           #hill_richness +
           agriculture +
           #grazing +
           logging +
           (1| site), data = plots_cf_4p)

lmm_3p_j <- lmer(log1p(agb_plot_ha) ~ env_pc1 +
           #slope +
           #hill_richness +
           #agriculture +
           grazing +
           logging +
           (1| site), data = plots_cf_4p)

# slope

lmm_3p_k <- lmer(log1p(agb_plot_ha) ~ #env_pc1 +
           slope +
           hill_richness +
           agriculture +
           #grazing +
           #logging +
           (1| site), data = plots_cf_4p)

lmm_3p_l <- lmer(log1p(agb_plot_ha) ~ #env_pc1 +
           slope +
           hill_richness +
           #agriculture +
           grazing +
           #logging +
           (1| site), data = plots_cf_4p)

lmm_3p_m <- lmer(log1p(agb_plot_ha) ~ #env_pc1 +
           slope +
           hill_richness +
           #agriculture +
           #grazing +
           logging +
           (1| site), data = plots_cf_4p)

lmm_3p_n <- lmer(log1p(agb_plot_ha) ~ #env_pc1 +
           slope +
           #hill_richness +
           agriculture +
           grazing +
           #logging +
           (1| site), data = plots_cf_4p)

lmm_3p_o <- lmer(log1p(agb_plot_ha) ~ #env_pc1 +
           slope +
           #hill_richness +
           agriculture +
           #grazing +
           logging +
           (1| site), data = plots_cf_4p)

lmm_3p_p <- lmer(log1p(agb_plot_ha) ~ #env_pc1 +
           slope +
           #hill_richness +
           #agriculture +
           grazing +
           logging +
           (1| site), data = plots_cf_4p)

# sp_richness

lmm_3p_q <- lmer(log1p(agb_plot_ha) ~ #env_pc1 +
           #slope +
           hill_richness +
           agriculture +
           grazing +
           #logging +
           (1| site), data = plots_cf_4p)

lmm_3p_r <- lmer(log1p(agb_plot_ha) ~ #env_pc1 +
           #slope +
           hill_richness +
           agriculture +
           #grazing +
           logging +
           (1| site), data = plots_cf_4p)

lmm_3p_s <- lmer(log1p(agb_plot_ha) ~ #env_pc1 +
           #slope +
           hill_richness +
           #agriculture +
           grazing +
           logging +
           (1| site), data = plots_cf_4p)

# agriculture

lmm_3p_t <- lmer(log1p(agb_plot_ha) ~ #env_pc1 +
           #slope +
           #hill_richness +
           agriculture +
           grazing +
           logging +
           (1| site), data = plots_cf_4p)

lmm3p_cAIC <- anocAIC(lmm_3p_a,
        lmm_3p_b,
        lmm_3p_c,
        lmm_3p_d,
        lmm_3p_e,
        lmm_3p_f,
        lmm_3p_g,
        lmm_3p_h,
        lmm_3p_i,
        lmm_3p_j,
        lmm_3p_k,
        lmm_3p_l,
        lmm_3p_m,
        lmm_3p_n,
        lmm_3p_o,
        lmm_3p_p,
        lmm_3p_q,
        lmm_3p_r,
        lmm_3p_s,
        lmm_3p_t,
        digits = 3)

lmm3p_cAIC <- as_tibble(lmm3p_cAIC, rownames = "model")

lmm3p_cAIC %>%
  arrange(cAIC)

```
Best model with 3 predictors is slope + sp_richness + agriculture cAIC 443.707

```{r}
# 4 predictors
combn(c("env","slope","sp_rich","ag","graz","logg"), 4, simplify = T)

# env_pc1

lmm_4p_a <- lmer(log1p(agb_plot_ha) ~ env_pc1 +
           slope +
           hill_richness +
           agriculture +
           #grazing +
           #logging +
           (1| site), data = plots_cf_4p)

lmm_4p_b <- lmer(log1p(agb_plot_ha) ~ env_pc1 +
           slope +
           hill_richness +
           #agriculture +
           grazing +
           #logging +
           (1| site), data = plots_cf_4p)

lmm_4p_c <- lmer(log1p(agb_plot_ha) ~ env_pc1 +
           slope +
           hill_richness +
           #agriculture +
           #grazing +
           logging +
           (1| site), data = plots_cf_4p)

lmm_4p_d <- lmer(log1p(agb_plot_ha) ~ env_pc1 +
           slope +
           #hill_richness +
           agriculture +
           grazing +
           #logging +
           (1| site), data = plots_cf_4p)

lmm_4p_e <- lmer(log1p(agb_plot_ha) ~ env_pc1 +
           slope +
           #hill_richness +
           agriculture +
           #grazing +
           logging +
           (1| site), data = plots_cf_4p)

lmm_4p_f <- lmer(log1p(agb_plot_ha) ~ env_pc1 +
           slope +
           #hill_richness +
           #agriculture +
           grazing +
           logging +
           (1| site), data = plots_cf_4p)

lmm_4p_g <- lmer(log1p(agb_plot_ha) ~ env_pc1 +
           #slope +
           hill_richness +
           agriculture +
           grazing +
           #logging +
           (1| site), data = plots_cf_4p)

lmm_4p_h <- lmer(log1p(agb_plot_ha) ~ env_pc1 +
           #slope +
           hill_richness +
           agriculture +
           #grazing +
           logging +
           (1| site), data = plots_cf_4p)

lmm_4p_i <- lmer(log1p(agb_plot_ha) ~ env_pc1 +
           #slope +
           hill_richness +
           #agriculture +
           grazing +
           logging +
           (1| site), data = plots_cf_4p)

lmm_4p_j <- lmer(log1p(agb_plot_ha) ~ env_pc1 +
           #slope +
           #hill_richness +
           agriculture +
           grazing +
           logging +
           (1| site), data = plots_cf_4p)

lmm_4p_k <- lmer(log1p(agb_plot_ha) ~ #env_pc1 +
           slope +
           hill_richness +
           agriculture +
           grazing +
           #logging +
           (1| site), data = plots_cf_4p)

lmm_4p_l <- lmer(log1p(agb_plot_ha) ~ #env_pc1 +
           slope +
           hill_richness +
           agriculture +
           #grazing +
           logging +
           (1| site), data = plots_cf_4p)

lmm_4p_m <- lmer(log1p(agb_plot_ha) ~ #env_pc1 +
           slope +
           hill_richness +
           #agriculture +
           grazing +
           logging +
           (1| site), data = plots_cf_4p)

lmm_4p_n <- lmer(log1p(agb_plot_ha) ~ #env_pc1 +
           slope +
           #hill_richness +
           agriculture +
           grazing +
           logging +
           (1| site), data = plots_cf_4p)

lmm_4p_o <- lmer(log1p(agb_plot_ha) ~ #env_pc1 +
           #slope +
           hill_richness +
           agriculture +
           grazing +
           logging +
           (1| site), data = plots_cf_4p)

lmm4p_cAIC <- anocAIC(lmm_4p_a,
        lmm_4p_b,
        lmm_4p_c,
        lmm_4p_d,
        lmm_4p_e,
        lmm_4p_f,
        lmm_4p_g,
        lmm_4p_h,
        lmm_4p_i,
        lmm_4p_j,
        lmm_4p_k,
        lmm_4p_l,
        lmm_4p_m,
        lmm_4p_n,
        lmm_4p_o,
        digits = 3)

lmm4p_cAIC <- as_tibble(lmm4p_cAIC, rownames = "model")

lmm4p_cAIC %>%
  arrange(cAIC)

```
Best model with 4 predictors is env_pc1 + slope + sp_richness + agriculture cAIC: 443.194

```{r}

# 5 predictors
combn(c("env","slope","sp_rich","ag","graz","logg"), 5, simplify = T)

lmm_5p_a <- lmer(log1p(agb_plot_ha) ~ env_pc1 +
           slope +
           hill_richness +
           agriculture +
           grazing +
           #logging +
           (1| site), data = plots_cf_4p)

lmm_5p_b <- lmer(log1p(agb_plot_ha) ~ env_pc1 +
           slope +
           hill_richness +
           agriculture +
           #grazing +
           logging +
           (1| site), data = plots_cf_4p)

lmm_5p_c <- lmer(log1p(agb_plot_ha) ~ env_pc1 +
           slope +
           hill_richness +
           #agriculture +
           grazing +
           logging +
           (1| site), data = plots_cf_4p)

lmm_5p_d <- lmer(log1p(agb_plot_ha) ~ env_pc1 +
           slope +
           #hill_richness +
           agriculture +
           grazing +
           logging +
           (1| site), data = plots_cf_4p)

lmm_5p_e <- lmer(log1p(agb_plot_ha) ~ env_pc1 +
           #slope +
           hill_richness +
           agriculture +
           grazing +
           logging +
           (1| site), data = plots_cf_4p)

lmm_5p_f <- lmer(log1p(agb_plot_ha) ~ #env_pc1 +
           slope +
           hill_richness +
           agriculture +
           grazing +
           logging +
           (1| site), data = plots_cf_4p)

lmm5p_cAIC <- anocAIC(lmm_5p_a,
        lmm_5p_b,
        lmm_5p_c,
        lmm_5p_d,
        lmm_5p_e,
        lmm_5p_f,
        digits = 3)

lmm5p_cAIC <- as_tibble(lmm5p_cAIC, rownames = "model")

lmm5p_cAIC %>%
  arrange(cAIC)

```

Best model with 5 predictors is env_pc1 + slope + hill_richness + agriculture + logging cAIC 443.769

```{r}
# All 6 predictors

lmm_6p_a <- lmer(log1p(agb_plot_ha) ~ env_pc1 +
           slope +
           hill_richness +
           agriculture +
           grazing +
           logging +
           (1| site), data = plots_cf_4p)

cAIC(lmm_6p_a)
```

Best model has 4 predictors: 
```{r}

summary(lmm_4p_a)
plot(lmm_4p_a)
plot_model(lmm_4p_a, show.p = T, show.values = T)
stargazer(lmm_4p_a, type = "text")
cAIC(lmm_4p_a)

qqnorm(resid(lmm_4p_a))
qqline(resid(lmm_4p_a))

shapiro.test(resid(lmm_4p_a))

rsq.lmm(lmm_4p_a, adj= T)

plot(plots_cf_4p$log_agb~fitted(lmm_4p_a))

model_performance(lmm_4p_a, estimator = "ML")
# returns to the un-transformed response variable

#RMSE

sqrt(mean((plots_cf_4p$log_agb - fitted(lmm_4p_a))^2))

yhat_lmm_4p_a <- fitted(lmm_4p_a)
res_lmm_4p_a <- residuals(lmm_4p_a)

# plot observed vs predicted and observed vs residuals and model coefficients
```

As we would expect some variables to be related, we also made models with interactions

### LMM: best model with interactions

```{r}


lmm_4p_int <- lmer(log1p(agb_plot_ha) ~ env_pc1 * agriculture +
                     hill_richness * agriculture +
                     slope +
                     (1| site), data = plots_cf_4p)

cAIC(lmm_4p_int)
stargazer(lmm_4p_int, type = "text")

summary(lmm_4p_int)
shapiro.test(resid(lmm_4p_int))

plot(lmm_4p_int)
plot_model(lmm_4p_int, show.p = T, show.values = T)

qqnorm(resid(lmm_4p_int))
qqline(resid(lmm_4p_int))

rsq.lmm(lmm_4p_int, adj= T)

sqrt(mean((plots_cf_4p$log_agb - fitted(lmm_4p_int))^2))
```


```{r}
cAIC(lmm_4p_a)
cAIC(lmm_4p_int)

anova(lmm_4p_a, lmm_4p_int)
```

So far, the best models are with 4 predictors. 
When adding an interactions between agriculture and sp_richness, the model improves a lot,
although residuals are not normal.

# GLMM

```{r}
plots_cf_4p <- plots_cf_4p %>%
  mutate(log_agb = log1p(agb_plot_ha))

plots_cf_4p %>%
  mutate(log_agb_scaled = scale(log_agb)) %>%
  select(log_agb, log_agb_scaled) %>%
  arrange(log_agb_scaled) # -1.499155119	

glmm_4p <- glmer(log_agb+0.1 ~ env_pc1 +
           scale(slope) +
           scale(hill_richness) +
           scale(agriculture) +
           (1| site),
           Gamma(link="inverse"),
           data = plots_cf_4p,
           glmerControl(optimizer = "bobyqa", optCtrl = list(maxfun = 100000)))
summary(glmm_4p)
stargazer(glmm_4p, type = "text")

shapiro.test(resid(glmm_4p))

plot(glmm_4p)
plot_model(glmm_4p, show.p = T, show.values = T)

qqnorm(resid(glmm_4p))
qqline(resid(glmm_4p))

```


#################################################################################
Linear mixed model

LMM with all variables and site as random effect

```{r}
lmm <- lmer(log1p(agb_plot_ha) ~ env_pc1 + slope + northness + eastness + hill_shannon + agriculture + grazing + (1| site), data = plots_cf_4p)

summary(lmm)
plot(lmm)
qqnorm(resid(lmm))
qqline(resid(lmm))

plot_model(lmm, show.values = T)
stargazer(lmm, type = "text")
```

LMM with all variables, interactions among them, and sites as random effect

```{r}
lmm2 <- lmer(log1p(agb_plot_ha) ~ env_pc1 + slope + northness + eastness + hill_shannon + agriculture + grazing + env_pc1*slope + slope*agriculture + slope*hill_shannon + slope*grazing + env_pc1*grazing + env_pc1*agriculture + env_pc1*hill_shannon + hill_shannon*grazing + hill_shannon*agriculture + grazing*agriculture + (1|site), data = plots_cf_4p)

summary(lmm2)
plot(lmm2)
qqnorm(resid(lmm2))
qqline(resid(lmm2))

plot_model(lmm2, show.values = T)

stargazer(lmm2, type = "text")
```
Lower AIC. 
I'll remove non-relevant variables first, and then non-relevant interactions:

```{r}
lmm3 <- lmer(log1p(agb_plot_ha) ~ env_pc1 + slope + hill_shannon + agriculture + grazing + (1|site), data = plots_cf_4p)

summary(lmm3)
plot(lmm3)
qqnorm(resid(lmm3))
qqline(resid(lmm3))

plot_model(lmm3, show.values = T)

stargazer(lmm3, type = "text")
```
This is the best model so far, although grazing is non-significant.
Let's add the relevant interactions:

```{r}
lmm4 <- lmer(log1p(agb_plot_ha) ~ env_pc1 + slope + hill_shannon + agriculture + hill_shannon*agriculture + slope*hill_shannon + env_pc1*grazing + (1|site), data = plots_cf_4p)

summary(lmm4)
stargazer(lmm4, type = "text")
plot_model(lmm4, show.values = T)

```
I'll remove the only interaction that wasn't statistically significant:

```{r}
lmm5 <- lmer(log1p(agb_plot_ha) ~ env_pc1 + slope + hill_shannon + agriculture + hill_shannon*agriculture + env_pc1*grazing + (1|site), data = plots_cf_4p)

summary(lmm5)
stargazer(lmm5, type = "text")
plot_model(lmm5, show.values = T)
```
Best model thus far.

```{r}

plots_cf_4p %>%
  ggplot(aes(x = hill_shannon, y = log1p(agb_plot_ha))) +
  geom_point(aes(color = forest_classification)) +
  geom_smooth(method = "lm")

plots_cf_4p %>%
  ggplot(aes(y = hill_shannon, x = agriculture)) +
  geom_point(aes(color = forest_classification)) +
  geom_smooth(method = "lm")

plots_cf_4p %>%
  ggplot(aes(y = hill_shannon, x = grazing)) +
  geom_point(aes(color = forest_classification)) +
  geom_smooth(method = "lm")

plots_cf_4p %>%
  ggplot(aes(x = agriculture, y = log1p(agb_plot_ha))) +
  geom_point(aes(color = forest_classification)) +
  geom_smooth(method = "lm")

plots_cf_4p %>%
  ggplot(aes(x = slope, y = log1p(agb_plot_ha))) +
  geom_point(aes(color = forest_classification)) +
  geom_smooth(method = "lm")

plots_cf_4p %>%
  ggplot(aes(x = grazing, y = log1p(agb_plot_ha))) +
  geom_point(aes(color = forest_classification)) +
  geom_smooth(method = "lm")

plots_cf_4p %>%
  ggplot(aes(x = env_pc1, y = agriculture)) +
  geom_point(aes(color = forest_classification)) +
  geom_smooth(method = "lm")

plots_cf_4p %>%
  ggplot(aes(x = altitude, y = agriculture)) +
  geom_point(aes(color = forest_classification)) +
  geom_smooth(method = "lm")

plots_cf_4p %>%
  ggplot(aes(x = altitude, y = grazing)) +
  geom_point(aes(color = forest_classification)) +
  geom_smooth(method = "lm")

plots_cf_4p %>%
  ggplot(aes(x = env_pc1, y = grazing)) +
  geom_point(aes(color = forest_classification)) +
  geom_smooth(method = "lm")

plots_cf_4p %>%
  ggplot(aes(x = env_pc1, y = hill_shannon)) +
  geom_point(aes(color = forest_classification)) +
  geom_smooth(method = "lm")

plots_cf_4p %>%
  ggplot(aes(x = env_pc1, y = log1p(agb_plot_ha))) +
  geom_point(aes(color = forest_classification)) +
  geom_smooth(method = "lm")

```

GLMM: agb ~ diversity

Name | Link Function | Distribution
General linear model | Identity | Normal
Logistic regression | Logit | Binomial
Poisson | Log | Poisson
Gamma | Inverse | Gamma

```{r}

glmm1 <- glmer((scale(agb_plot_ha)+1) ~ hill_shannon + (1|site), Gamma(link = "inverse"), data = plots_cf_4p, nAGQ = 2, glmerControl(optimizer = "bobyqa", optCtrl = list(maxfun = 100000)))
summary(glmm1)

glmm1b <- glmer((scale(agb_plot_ha)+1) ~ hill_richness + (1|site), Gamma(link = "inverse"), data = plots_cf_4p, nAGQ = 2, glmerControl(optimizer = "bobyqa", optCtrl = list(maxfun = 100000)))
summary(glmm1b)

lmer1 <- lmer(log1p(agb_plot_ha) ~ hill_shannon + (1|site), data = plots_cf_4p)
summary(lmer1)
plot(lmer1)
plot_model(lmer1, show.p = T, show.values = T)
stargazer(lmer1, type = "text")
# This model is significant but the relationship doesn't seem to be normal
# Also, about 60% of variation is explained by the random effect
# which I think it means that there's a stronger relationship between
# agb and diversity between plots than between sites 

lmer1b <- lmer(agb_plot_ha ~ hill_shannon + (1|site), data = plots_cf_4p)
summary(lmer1b)
plot(lmer1b)
stargazer(lmer1b, type = "text")

glmm1c <- glmer((log1p(agb_plot_ha+1)) ~ hill_shannon + (1|site), Gamma(link = "inverse"), data = plots_cf_4p, nAGQ = 2, glmerControl(optimizer = "bobyqa", optCtrl = list(maxfun = 100000)))
summary(glmm1c)
plot(glmm1c)

# GAMM?
```

GMM: agb ~ all variables

```{r}

glmm2 <- glmer((scale(agb_plot_ha)+1) ~ env_pc1 + scale(slope) + hill_shannon + agriculture + grazing + env_pc1*scale(slope) + scale(slope)*agriculture + scale(slope)*hill_shannon + scale(slope)*grazing + env_pc1*grazing + env_pc1*agriculture + env_pc1*hill_shannon + hill_shannon*grazing + hill_shannon*agriculture + grazing*agriculture + (1|site), Gamma(link="inverse"), data = plots_cf_4p, glmerControl(optimizer = "bobyqa", optCtrl = list(maxfun = 100000)))
summary(glmm2)

# dropping variables that are not significant in 'all variables and all interactions model'
glmm2b <- glmer((scale(agb_plot_ha)+1) ~ env_pc1 + scale(slope) + agriculture + scale(slope)*agriculture + scale(slope)*grazing + env_pc1*grazing + hill_shannon*agriculture + grazing*agriculture + (1|site), Gamma(link="inverse"), data = plots_cf_4p, glmerControl(optimizer = "bobyqa", optCtrl = list(maxfun = 100000)))
summary(glmm2b)

# Agriculture ~ environmental gradient
g <- glmer.nb(agriculture ~ env_pc1 + (1|site), data = plots_cf_4p, verbose = T)
getME(g, "glmer.nb.theta")

glmm3 <- glmer(agriculture ~ env_pc1 + (1|site), negative.binomial(theta= 87730.47, link="log"), data = plots_cf_4p, glmerControl(optimizer = "bobyqa", optCtrl = list(maxfun = 100000)))
summary(glmm3)
shapiro.test(resid(glmm3))
plot(glmm3)

## *Significant interaction!*

# Agriculture ~ slope

gb <- glmer.nb(agriculture ~ scale(slope) + (1|site), data = plots_cf_4p, verbose = T)
getME(gb, "glmer.nb.theta")

glmm3b <- glmer(agriculture ~ scale(slope) + (1|site), negative.binomial(theta= 88474.5, link="log"), data = plots_cf_4p, glmerControl(optimizer = "bobyqa", optCtrl = list(maxfun = 100000)))
summary(glmm3b)

## Non-significant

# Grazing ~ environmental gradient

# Grazing only has two values so I'm making it binomial 
ds <- plots_cf_4p %>%
  mutate(graz_bi = ifelse(grazing == 2, 1, grazing)) 
glmm4 <- glmer(graz_bi ~ env_pc1 + (1|site), binomial(link="logit"), data = ds, glmerControl(optimizer = "bobyqa", optCtrl = list(maxfun = 100000)))
summary(glmm4) #as binomial variable

g2 <- glmer.nb(grazing ~ env_pc1 + (1|site), data = plots_cf_4p, verbose = T)
getME(g2, "glmer.nb.theta")

glmm4b <- glmer(grazing ~ env_pc1 + (1|site), negative.binomial(theta= 75716.24, link="log"), data = plots_cf_4p, glmerControl(optimizer = "bobyqa", optCtrl = list(maxfun = 100000)))
summary(glmm4b) #as discrete variable

## Non-significant

# Grazing ~ slope

glmm4c <- glmer(graz_bi ~ scale(slope) + (1|site), binomial(link="logit"), data = ds, glmerControl(optimizer = "bobyqa", optCtrl = list(maxfun = 100000)))
summary(glmm4c) #as binomial variable

## Non-significant

# Tree diversity (hill_shannon) ~ environmental gradient

lmer_tshannon1 <- lmer(sqrt(hill_shannon) ~ env_pc1 + (1|site), data = plots_cf_4p)
summary(lmer_tshannon1)
stargazer(lmer_tshannon1, type = "text")
plot_model(lmer_tshannon1, show.values = T)
qqnorm(resid(lmer_tshannon1))
qqline(resid(lmer_tshannon1))

## Non-significant

# Tree diversity ~ agriculture

glmm5b <- glmer((hill_shannon+0.1) ~ agriculture + (1|site), Gamma(link="inverse"), data = plots_cf_4p, glmerControl(optimizer = "bobyqa", optCtrl = list(maxfun = 100000)))
summary(glmm5b)

lmer_tshannon <- lmer(hill_shannon ~ agriculture + (1|site), data = plots_cf_4p)
summary(lmer_tshannon)
stargazer(lmer_tshannon, type = "text")
plot_model(lmer_tshannon, show.values = T)
qqnorm(resid(lmer_tshannon))
qqline(resid(lmer_tshannon))

# sqrt transformation for slightly right-skewed data
lmer_tshannon2 <- lmer(sqrt(hill_shannon) ~ agriculture + (1|site), data = plots_cf_4p)
summary(lmer_tshannon2)
stargazer(lmer_tshannon2, type = "text")
plot_model(lmer_tshannon2, show.values = T)
qqnorm(resid(lmer_tshannon2))
qqline(resid(lmer_tshannon2))

## *Significant!*

# Tree diversity ~ grazing

lmer_tshannon3 <- lmer(sqrt(hill_shannon) ~ grazing + (1|site), data = plots_cf_4p)
summary(lmer_tshannon3)
stargazer(lmer_tshannon3, type = "text")
plot_model(lmer_tshannon3, show.values = T)
qqnorm(resid(lmer_tshannon3))
qqline(resid(lmer_tshannon3))

## *Significant!*

# Tree diversity ~ slope 
lmer_tshannon_slope <- lmer(sqrt(hill_shannon) ~ scale(slope) + (1|site), data = plots_cf_4p)
summary(lmer_tshannon_slope)
stargazer(lmer_tshannon_slope, type = "text")

## Non-significant
```


With interactions:

```{r}

glmm_w_interactions <- glmer((scale(agb_plot_ha)+1) ~  env_pc1*agriculture + 
                                         scale(slope) +
                                         hill_shannon*agriculture + 
                                         hill_shannon*grazing +
                                         (1|site), Gamma(link="inverse"), data = plots_cf_4p, glmerControl(optimizer = "bobyqa", optCtrl = list(maxfun = 100000)))
summary(glmm_w_interactions)
plot(glmm_w_interactions)
qqnorm(resid(glmm_w_interactions))
qqline(resid(glmm_w_interactions))

plot_model(glmm_w_interactions, show.values = T)
stargazer(glmm_w_interactions, type = "text")

# Reduced:

glmm_w_interactions_r <- glmer((scale(agb_plot_ha)+1) ~  env_pc1*agriculture + 
                                         scale(slope) +
                                         hill_shannon*agriculture + 
                                         #hill_shannon*grazing +
                                         (1|site), Gamma(link="inverse"), data = plots_cf_4p, glmerControl(optimizer = "bobyqa", optCtrl = list(maxfun = 100000)))
summary(glmm_w_interactions_r)
plot(glmm_w_interactions_r)
qqnorm(resid(glmm_w_interactions_r))
qqline(resid(glmm_w_interactions_r))

plot_model(glmm_w_interactions_r, show.values = T)
stargazer(glmm_w_interactions_r, type = "text")

glmm_wo_interactions <- glmer((scale(agb_plot_ha)+1) ~  env_pc1 + 
                                                        agriculture + 
                                                        scale(slope) +
                                                        hill_shannon + 
                                                        grazing +
                                         (1|site), Gamma(link="inverse"), data = plots_cf_4p, glmerControl(optimizer = "bobyqa", optCtrl = list(maxfun = 100000)))
summary(glmm_wo_interactions)
plot(glmm_wo_interactions)
qqnorm(resid(glmm_wo_interactions))
qqline(resid(glmm_wo_interactions))

plot_model(glmm_wo_interactions, show.values = T)
stargazer(glmm_wo_interactions, type = "text")
```

```{r}

lmer_w_interactions <- lmer(log1p(agb_plot_ha) ~ env_pc1*agriculture + 
                                         #env_pc1*grazing + #If added to the model, AIC decreases and env turns very relevant, but it seems weird because these two are not really related (see test above)
                                         scale(slope) +
                                         hill_shannon*agriculture + 
                                         hill_shannon*grazing +
                                         (1|site), data = plots_cf_4p)

summary(lmer_w_interactions)
plot(lmer_w_interactions)
qqnorm(resid(lmer_w_interactions))
qqline(resid(lmer_w_interactions))

plot_model(lmer_w_interactions, show.values = T)
stargazer(lmer_w_interactions, type = "text")
```


Plots classification

Option 1: k-means including PC1
```{r}
plots_str2 <- plots_cf_4p %>%
  select(plot_id,
         tree_density,
         loreys_height,
         basal_area_ha,
         agb_plot_ha,
         env_pc1)

plot(plots_str2)

# Scale data so that variables in different units can be compared
plots_str_scaled2 <- scale(plots_str2[,-c(1,2,5)])
row.names(plots_str_scaled2) <- plots_str2$plot_id

k2_structure_plots2 <- kmeans(plots_str_scaled2, centers = 2, nstart = 25)
k3_structure_plots2 <- kmeans(plots_str_scaled2, centers = 3, nstart = 25)
k4_structure_plots2 <- kmeans(plots_str_scaled2, centers = 4, nstart = 25)
k5_structure_plots2 <- kmeans(plots_str_scaled2, centers = 5, nstart = 25)

p1_str_plots2 <- fviz_cluster(k2_structure_plots2, geom = "point", data = plots_str_scaled2) + ggtitle("k = 2")
p2_str_plots2 <- fviz_cluster(k3_structure_plots2, geom = "point",  data = plots_str_scaled2) + ggtitle("k = 3")
p3_str_plots2 <- fviz_cluster(k4_structure_plots2, geom = "point",  data = plots_str_scaled2) + ggtitle("k = 4")
p4_str_plots2 <- fviz_cluster(k5_structure_plots2, geom = "point",  data = plots_str_scaled2) + ggtitle("k = 5")
grid.arrange(p1_str_plots2, p2_str_plots2, p3_str_plots2, p4_str_plots2, nrow = 2)

str_clusters_plots2 <- NbClust((plots_str_scaled2^2), distance = "euclidean", min.nc = 2, max.nc = 10, method = "ward.D2", index = "all")

# Save a vector of 2 clusters:
str_k2_plots2 <- k2_structure_plots2$cluster
as.numeric(str_k2_plots2)

#pca with structural data

pca_str_plots2 <- rda(plots_str_scaled2, scale = F)
summary(pca_str_plots2) # 
plot(pca_str_plots2)

ordiplot(pca_str_plots2, display = 'sites', type = 'n')
points(pca_str_plots2, col = str_k2_plots2)
biplot(pca_str_plots2, display = 'species')

ordiplot(pca_str_plots, display = 'sites', type = 'text')
biplot(pca_str_plots)

ordiplot(pca_str_plots2, display = 'sites', type = 'n')
points(pca_str_plots, col = str_k3_plots)
biplot(pca_str_plots, display = 'species')

ordiplot(pca_str_plots, display = 'sites', type = 'text')
biplot(pca_str_plots)

plots_cf_4p %>%
  ggplot(aes(x= forest_classification, y = agriculture, color = basal_area_ha)) +
  geom_jitter()

plots_cf_4p %>%
  ggplot(aes(x= forest_classification, y = grazing)) +
  geom_jitter(color = 'blue')

plots_cf_4p %>%
  filter(forest_classification == 1) %>%
  ggplot(aes(x = env_pc1, y = loreys_height, color = as.factor(agriculture))) +
  geom_point()

plots_cf_4p %>%
  filter(forest_classification == 2) %>%
  ggplot(aes(x = altitude, y = agb_plot_ha, color = as.factor(agriculture))) +
  geom_point()

plots_cf_4p %>%
  filter(forest_classification == 3) %>%
  ggplot(aes(x = altitude, y = agb_plot_ha, color = as.factor(agriculture))) +
  geom_point()


plots_cf_4p %>%
  filter(forest_classification == 1) %>%
  arrange(agriculture, -altitude)

plots_cf_4p %>%
  mutate(land_use = agriculture + grazing) %>%
  ggplot(aes(x = forest_classification, y = land_use)) +
  geom_jitter()
  
```

# GAMMs

Null model with random effects 

```{r}

gamm4_null <- gamm4(log1p(agb_plot_ha) ~ 1, random = ~(1|site), data = plots_cf_4p)
summary(gamm4_null)
summary(gamm4_null$mer)
summary(gamm4_null$gam)
gam0 <- gamm4_null$gam

```

## GAMM: 1 predictor

```{r}
# env_pc1, slope, hill_shannon, hill_richness, agriculture, grazing, logging

gamm4_1a <- gamm4(log1p(agb_plot_ha) ~ env_pc1, # relationship looks linear so I removed t2
                  random = ~(1|site), data = plots_cf_4p) 

gamm4_1b <-  gamm4(log1p(agb_plot_ha) ~ t2(hill_shannon, bs="tp", k=4),
                   random = ~(1|site), data = plots_cf_4p)
gamm4_1b_4 <-  gamm4(log1p(agb_plot_ha) ~ s(hill_shannon, bs="tp", k=4),
                   random = ~(1|site), data = plots_cf_4p)

agb_h <- lmer(log1p(agb_plot_ha) ~ hill_shannon + (1|site), data = plots_cf_4p)
stargazer(agb_h,type="text",report=("vc*p"))

agb_h2 <- lmer(log1p(agb_plot_ha) ~ hill_shannon*forest_classification + (1|site), data = plots_cf_4p)
summary(agb_h2)
plot_model(agb_h2)
stargazer(agb_h2,type="text",report=("vc*p"))

agb_h3 <- lmer(log1p(agb_plot_ha) ~ hill_shannon + (1|forest_classification) + (1|site) , data = plots_cf_4p)
summary(agb_h3)
stargazer(agb_h3,type="text",report=("vc*p"))

# Ganador (falta hacer cAICs) --Escribe sobre modelo con y sin forest_classification y sobre la variacion que Site explica. 
gamm4_1b_3 <-  gamm4(log1p(agb_plot_ha) ~ s(hill_shannon, bs= "tp", k= 4),
                   random = ~(1|site) + (1|forest_classification), data = plots_cf_4p)
summary(gamm4_1b_3$gam)

gamm4_1b_5 <-  gamm4(log1p(agb_plot_ha) ~ forest_classification + s(hill_shannon, bs= "tp", k= 4) ,
                   random = ~(1|site), data = plots_cf_4p) #forest class as fixed effect because it has only 4 levels
summary(gamm4_1b_5$gam)

gamm4_1c <- gamm4(log1p(agb_plot_ha) ~ t2(hill_richness, bs="tp", k=4),
                   random = ~(1|site), data = plots_cf_4p)

gamm4_1d <- gamm4(log1p(agb_plot_ha) ~ agriculture,
                   random = ~(1|site), data = plots_cf_4p)

gamm4_1e <- gamm4(log1p(agb_plot_ha) ~ grazing,
                   random = ~(1|site), data = plots_cf_4p)

gamm4_1f <- gamm4(log1p(agb_plot_ha) ~ logging,
                   random = ~(1|site), data = plots_cf_4p)

gamm4_1g <-  gamm4(log1p(agb_plot_ha) ~ slope,
                   random = ~(1|site), data = plots_cf_4p)

summary(gamm4_1a$mer)
summary(gamm4_1a$gam)
#plot.gam(gamm4_1a$gam, residuals=T, se=T, shade=T, pch=16, cex=0.5)
gam.check(gamm4_1a$gam)
aic1a <- cAIC(gamm4_1a)

summary(gamm4_1b$mer)
summary(gamm4_1b$gam)
plot.gam(gamm4_1b$gam, residuals=T, se=T, shade=T, pch=16, cex=0.5)
gam.check(gamm4_1b$gam)
aic1b <- cAIC(gamm4_1b)

summary(gamm4_1c$mer)
summary(gamm4_1c$gam)
plot.gam(gamm4_1c$gam, residuals=T, se=T, shade=T, pch=16, cex=0.5)
gam.check(gamm4_1c$gam)
aic1c <- cAIC(gamm4_1c)

summary(gamm4_1d$mer)
summary(gamm4_1d$gam)
#plot.gam(gamm4_1d$gam, residuals=T, se=T, shade=T, pch=16, cex=0.5)
gam.check(gamm4_1d$gam)
aic1d <- cAIC(gamm4_1d)

summary(gamm4_1e$mer)
summary(gamm4_1e$gam)
#plot.gam(gamm4_1e$gam, residuals=T, se=T, shade=T, pch=16, cex=0.5)
gam.check(gamm4_1e$gam)
aic1e <- cAIC(gamm4_1e)

summary(gamm4_1f$mer)
summary(gamm4_1f$gam)
#plot.gam(gamm4_1f$gam, residuals=T, se=T, shade=T, pch=16, cex=0.5)
gam.check(gamm4_1f$gam)
aic1f <- cAIC(gamm4_1f)

summary(gamm4_1g$mer)
summary(gamm4_1g$gam)
#plot.gam(gamm4_1g$gam, residuals=T, se=T, shade=T, pch=16, cex=0.5)
gam.check(gamm4_1g$gam)
aic1g <- cAIC(gamm4_1g)


gamms1_aic <- c(aic1a$caic,
                aic1b$caic,
                aic1c$caic,
                aic1d$caic,
                aic1e$caic,
                aic1f$caic,
                aic1g$caic)

gamms1_aic
# Best model with 1 predictor (based on cAIC) is agb ~ sp richness 
summary(gamm4_1b$mer)
summary(gamm4_1b$gam)
aic1b$caic

plots_cf_4p_wo <- plots_cf_4p %>%
  filter(hill_shannon < 9)

gamm4_1b_2 <-  gamm4(log1p(agb_plot_ha) ~ t2(hill_shannon, bs="tp", k=4),
                   random = ~(1|site), data = plots_cf_4p_wo)
summary(gamm4_1b_2$mer)
summary(gamm4_1b_2$gam)
cAIC(gamm4_1b_2)
plot.gam(gamm4_1b_2$gam, residuals=T, se=T, shade=T, pch=16, cex=0.5)
```

## GAMM: 2 predictors

```{r}

gamm4_2a <- gamm4(log1p(agb_plot_ha) ~ env_pc1 + # relationship looks linear so I removed t2
            slope,
           #hill_shannon +
           #hill_richness +
           #agriculture +
           #grazing +
           #logging +
           random = ~(1|site), data = plots_cf_4p)

gamm4_2b <- gamm4(log1p(agb_plot_ha) ~ env_pc1 +
           #slope +
           t2(hill_shannon, bs="tp", k=4),
           #hill_richness +
           #agriculture +
           #grazing +
           #logging +
           random = ~(1|site), data = plots_cf_4p)

gamm4_2c <- gamm4(log1p(agb_plot_ha) ~ env_pc1 +
           #slope +
           #hill_shannon +
           t2(hill_richness, bs="tp", k=4),
           #agriculture +
           #grazing +
           #logging +
           random = ~(1|site), data = plots_cf_4p)

gamm4_2d <- gamm4(log1p(agb_plot_ha) ~ env_pc1 +
           #slope +
           #hill_shannon +
           #hill_richness +
           agriculture,
           #grazing +
           #logging +
           random = ~(1|site), data = plots_cf_4p)

gamm4_2e <- gamm4(log1p(agb_plot_ha) ~ env_pc1 +
           #slope +
           #hill_shannon +
           #hill_richness +
           #agriculture +
           grazing, 
           #logging +
           random = ~(1|site), data = plots_cf_4p)

gamm4_2f <- gamm4(log1p(agb_plot_ha) ~ env_pc1 +
           #slope +
           #hill_shannon +
           #hill_richness +
           #agriculture +
           #grazing +
           logging,
           random = ~(1|site), data = plots_cf_4p)

# Slope 

gamm4_2g <- gamm4(log1p(agb_plot_ha) ~ #env_pc1 +
           slope +
           t2(hill_shannon, bs="tp", k=4),
           #hill_richness +
           #agriculture +
           #grazing +
           #logging +
           random = ~(1|site), data = plots_cf_4p)

gamm4_2h <- gamm4(log1p(agb_plot_ha) ~ #env_pc1 +
           slope +
           #hill_shannon +
           t2(hill_richness, bs="tp", k=4),
           #agriculture +
           #grazing +
           #logging +
           random = ~(1|site), data = plots_cf_4p)

gamm4_2i <- gamm4(log1p(agb_plot_ha) ~ #env_pc1 +
           slope +
           #hill_shannon +
           #hill_richness +
           agriculture,
           #grazing +
           #logging +
           random = ~(1|site), data = plots_cf_4p)

gamm4_2j <- gamm4(log1p(agb_plot_ha) ~ #env_pc1 +
           slope +
           #hill_shannon +
           #hill_richness +
           #agriculture +
           grazing,
           #logging +
           random = ~(1|site), data = plots_cf_4p)

gamm4_2k <- gamm4(log1p(agb_plot_ha) ~ #env_pc1 +
           slope +
           #hill_shannon +
           #hill_richness +
           #agriculture +
           #grazing +
           logging,
           random = ~(1|site), data = plots_cf_4p)

# hill_shannon

gamm4_2l <- gamm4(log1p(agb_plot_ha) ~ #env_pc1 +
           #slope +
           t2(hill_shannon, bs="tp", k=4) +
           t2(hill_richness, bs="tp", k=4),
           #agriculture +
           #grazing +
           #logging +
           random = ~(1|site), data = plots_cf_4p)

gamm4_2m <- gamm4(log1p(agb_plot_ha) ~ #env_pc1 +
           #slope +
           t2(hill_shannon, bs="tp", k=4) +
           #hill_richness +
           agriculture,
           #grazing +
           #logging +
           random = ~(1|site), data = plots_cf_4p)

gamm4_2n <- gamm4(log1p(agb_plot_ha) ~ #env_pc1 +
           #slope +
           t2(hill_shannon, bs="tp", k=4) +
           #hill_richness +
           #agriculture +
           grazing,
           #logging +
           random = ~(1|site), data = plots_cf_4p)

gamm4_2o <- gamm4(log1p(agb_plot_ha) ~ #env_pc1 +
           #slope +
           t2(hill_shannon, bs="tp", k=4) +
           #hill_richness +
           #agriculture +
           #grazing +
           logging,
           random = ~(1|site), data = plots_cf_4p)

# hill_richness

gamm4_2p <- gamm4(log1p(agb_plot_ha) ~ #env_pc1 +
           #slope +
           #hill_shannon +
           t2(hill_richness, bs="tp", k=4) +
           agriculture,
           #grazing +
           #logging +
           random = ~(1|site), data = plots_cf_4p)

gamm4_2q <- gamm4(log1p(agb_plot_ha) ~ #env_pc1 +
           #slope +
           #hill_shannon +
           t2(hill_richness, bs="tp", k=4) +
           #agriculture +
           grazing,
           #logging +
           random = ~(1|site), data = plots_cf_4p)

gamm4_2r <- gamm4(log1p(agb_plot_ha) ~ #env_pc1 +
           #slope +
           #hill_shannon +
           t2(hill_richness, bs="tp", k=4) +
           #agriculture +
           #grazing +
           logging,
           random = ~(1|site), data = plots_cf_4p)

# Agriculture

gamm4_2s <- gamm4(log1p(agb_plot_ha) ~ #env_pc1 +
           #slope +
           #hill_shannon +
           #hill_richness +
           agriculture +
           grazing,
           #logging +
           random = ~(1|site), data = plots_cf_4p)

gamm4_2t <- gamm4(log1p(agb_plot_ha) ~ #env_pc1 +
           #slope +
           #hill_shannon +
           #hill_richness +
           agriculture +
           #grazing +
           logging,
           random = ~(1|site), data = plots_cf_4p)

# Grazing

gamm4_2u <- gamm4(log1p(agb_plot_ha) ~ #env_pc1 +
           #slope +
           #hill_shannon +
           #hill_richness +
           #agriculture +
           grazing +
           logging,
           random = ~(1|site), data = plots_cf_4p)

# cAIC

# aic2a <- cAIC(gamm4_2a)
# aic2b <- cAIC(gamm4_2b)
# aic2c <- cAIC(gamm4_2c)
# aic2d <- cAIC(gamm4_2d)
# aic2e <- cAIC(gamm4_2e)
# aic2f <- cAIC(gamm4_2f)
# aic2g <- cAIC(gamm4_2g)
# aic2h <- cAIC(gamm4_2h)
# aic2i <- cAIC(gamm4_2i)
# aic2j <- cAIC(gamm4_2j)
# aic2k <- cAIC(gamm4_2k)
# aic2l <- cAIC(gamm4_2l)
# aic2m <- cAIC(gamm4_2m)
# aic2n <- cAIC(gamm4_2n)
# aic2o <- cAIC(gamm4_2o)
# #aic2p <- cAIC(gamm4_2p) #Error: After removing the terms with zero variance components and refitting the model cAIC can be called again.
# aic2q <- cAIC(gamm4_2q)
# aic2r <- cAIC(gamm4_2r)
# aic2s <- cAIC(gamm4_2s)
# aic2t <- cAIC(gamm4_2t)
# aic2u <- cAIC(gamm4_2u)


# gamms2_aic <- c("gamm4_2a", aic2a$caic,
#                 "gamm4_2b", aic2b$caic,
#                 "gamm4_2c", aic2c$caic,
#                 "gamm4_2d", aic2d$caic,
#                 "gamm4_2e", aic2e$caic,
#                 "gamm4_2f", aic2f$caic,
#                 "gamm4_2g", aic2g$caic,
#                 "gamm4_2h", aic2h$caic,
#                 "gamm4_2i", aic2i$caic,
#                 "gamm4_2j", aic2j$caic,
#                 "gamm4_2k", aic2k$caic,
#                 "gamm4_2l", aic2l$caic,
#                 "gamm4_2m", aic2m$caic,
#                 "gamm4_2n", aic2n$caic,
#                 "gamm4_2o", aic2o$caic,
#                 #"gamm4_2p", aic2p$caic,
#                 "gamm4_2q", aic2q$caic,
#                 "gamm4_2r", aic2r$caic,
#                 "gamm4_2s", aic2s$caic,
#                 "gamm4_2t", aic2t$caic,
#                 "gamm4_2u", aic2u$caic)

# Best model of 2 predictors is agb ~ slope + shannon cAIC = 437.6538
summary(gamm4_2g$mer)
summary(gamm4_2g$gam)
summary(gamm4_2h$mer)
summary(gamm4_2h$gam)
```
## GAMM: 3 predictors

```{r}
gamm4_3a <- gamm4(log1p(agb_plot_ha) ~ env_pc1 +
           slope +
           t2(hill_shannon, bs="tp", k=4),
           #agriculture +
           #grazing +
           #logging +
           random = ~(1|site), data = plots_cf_4p)

gamm4_3b <- gamm4(log1p(agb_plot_ha) ~ env_pc1 +
           slope +
           #hill_shannon +
           agriculture,
           #grazing +
           #logging +
           random = ~(1|site), data = plots_cf_4p)

gamm4_3c <- gamm4(log1p(agb_plot_ha) ~ env_pc1 +
           slope +
           #hill_shannon +
           #agriculture +
           grazing,
           #logging +
           random = ~(1|site), data = plots_cf_4p)

gamm4_3d <- gamm4(log1p(agb_plot_ha) ~ env_pc1 +
           slope +
           #hill_shannon +
           #agriculture +
           #grazing +
           logging,
           random = ~(1|site), data = plots_cf_4p)

gamm4_3e <- gamm4(log1p(agb_plot_ha) ~ env_pc1 +
           #slope +
           t2(hill_shannon, bs="tp", k=4) +
           agriculture,
           #grazing +
           #logging +
           random = ~(1|site), data = plots_cf_4p)

gamm4_3f <- gamm4(log1p(agb_plot_ha) ~ env_pc1 +
           #slope +
           t2(hill_shannon, bs="tp", k=4) +
           #agriculture +
           grazing,
           #logging +
           random = ~(1|site), data = plots_cf_4p)

gamm4_3g <- gamm4(log1p(agb_plot_ha) ~ env_pc1 +
           #slope +
           t2(hill_shannon, bs="tp", k=4) +
           #agriculture +
           #grazing +
           logging, 
           random = ~(1|site), data = plots_cf_4p)

gamm4_3h <- gamm4(log1p(agb_plot_ha) ~ env_pc1 +
           #slope +
           #hill_shannon +
           agriculture +
           grazing, 
           #logging +
           random = ~(1|site), data = plots_cf_4p)

gamm4_3i <- gamm4(log1p(agb_plot_ha) ~ env_pc1 +
           #slope +
           #hill_shannon +
           agriculture +
           #grazing +
           logging, 
           random = ~(1|site), data = plots_cf_4p)

gamm4_3j <- gamm4(log1p(agb_plot_ha) ~ env_pc1 +
           #slope +
           #hill_shannon +
           #agriculture +
           grazing +
           logging, 
           random = ~(1|site), data = plots_cf_4p)

# slope

gamm4_3k <- gamm4(log1p(agb_plot_ha) ~ #env_pc1 +
           slope +
           t2(hill_shannon, bs="tp", k=4) +
           agriculture,
           #grazing +
           #logging +
           random = ~(1|site), data = plots_cf_4p)

gamm4_3l <- gamm4(log1p(agb_plot_ha) ~ #env_pc1 +
           slope +
           t2(hill_shannon, bs="tp", k=4) +
           #agriculture +
           grazing,
           #logging +
           random = ~(1|site), data = plots_cf_4p)

gamm4_3m <- gamm4(log1p(agb_plot_ha) ~ #env_pc1 +
           slope +
           t2(hill_shannon, bs="tp", k=4) +
           #agriculture +
           #grazing +
           logging,
           random = ~(1|site), data = plots_cf_4p)

gamm4_3n <- gamm4(log1p(agb_plot_ha) ~ #env_pc1 +
           slope +
           #hill_shannon +
           agriculture +
           grazing,
           #logging +
           random = ~(1|site), data = plots_cf_4p)

gamm4_3o <- gamm4(log1p(agb_plot_ha) ~ #env_pc1 +
           slope +
           #hill_shannon +
           agriculture +
           #grazing +
           logging,
           random = ~(1|site), data = plots_cf_4p)

gamm4_3p <- gamm4(log1p(agb_plot_ha) ~ #env_pc1 +
           slope +
           #hill_shannon +
           #agriculture +
           grazing +
           logging,
           random = ~(1|site), data = plots_cf_4p)

# sp_richness

gamm4_3q <- gamm4(log1p(agb_plot_ha) ~ #env_pc1 +
           #slope +
           t2(hill_shannon, bs="tp", k=4) +
           agriculture +
           grazing,
           #logging +
           random = ~(1|site), data = plots_cf_4p)

gamm4_3r <- gamm4(log1p(agb_plot_ha) ~ #env_pc1 +
           #slope +
           t2(hill_shannon, bs="tp", k=4) +
           agriculture +
           #grazing +
           logging,
           random = ~(1|site), data = plots_cf_4p)

gamm4_3s <- gamm4(log1p(agb_plot_ha) ~ #env_pc1 +
           #slope +
           t2(hill_shannon, bs="tp", k=4) +
           #agriculture +
           grazing +
           logging,
           random = ~(1|site), data = plots_cf_4p)

# agriculture

gamm4_3t <- gamm4(log1p(agb_plot_ha) ~ #env_pc1 +
           #slope +
           #hill_shannon +
           agriculture +
           grazing +
           logging,
           random = ~(1|site), data = plots_cf_4p)

aic3a <- cAIC(gamm4_3a)
aic3b <- cAIC(gamm4_3b)
aic3c <- cAIC(gamm4_3c)
aic3d <- cAIC(gamm4_3d)
aic3e <- cAIC(gamm4_3e)
aic3f <- cAIC(gamm4_3f)
aic3g <- cAIC(gamm4_3g)
aic3h <- cAIC(gamm4_3h)
aic3i <- cAIC(gamm4_3i)
aic3j <- cAIC(gamm4_3j)
#aic3k <- cAIC(gamm4_3k) # Error: After removing the terms with zero variance components and refitting the model cAIC can be called again.
aic3l <- cAIC(gamm4_3l)
aic3m <- cAIC(gamm4_3m)
aic3n <- cAIC(gamm4_3n)
aic3o <- cAIC(gamm4_3o)
aic3p <- cAIC(gamm4_3p)
#aic3q <- cAIC(gamm4_3q) # Error: After removing the terms with zero variance components and refitting the model cAIC can be called again.
aic3r <- cAIC(gamm4_3r)
aic3s <- cAIC(gamm4_3s)
aic3t <- cAIC(gamm4_3t)

gamms3_aic <- c("gamm4_3a", aic3a$caic,
                "gamm4_3b", aic3b$caic,
                "gamm4_3c", aic3c$caic,
                "gamm4_3d", aic3d$caic,
                "gamm4_3e", aic3e$caic,
                "gamm4_3f", aic3f$caic,
                "gamm4_3g", aic3g$caic,
                "gamm4_3h", aic3h$caic,
                "gamm4_3i", aic3i$caic,
                "gamm4_3j", aic3j$caic,
                #"gamm4_3k", aic3k$caic,
                "gamm4_3l", aic3l$caic,
                "gamm4_3m", aic3m$caic,
                "gamm4_3n", aic3n$caic,
                "gamm4_3o", aic3o$caic,
                "gamm4_3p", aic3p$caic,
                #"gamm4_3q", aic3q$caic,
                "gamm4_3r", aic3r$caic,
                "gamm4_3s", aic3s$caic,
                "gamm4_3t", aic3t$caic)

# Best model with 3 predictors is agb ~ slope + sp_richness + agriculture cAIC= 434.859
```
## GAMM: 4 predictors

```{r}
gamm4_4a <- gamm4(log1p(agb_plot_ha) ~ env_pc1 +
           slope +
           t2(hill_shannon, bs="tp", k=4) +
           agriculture,
           #grazing +
           #logging +
           random = ~(1|site), data = plots_cf_4p)

gamm4_4b <- gamm4(log1p(agb_plot_ha) ~ env_pc1 +
           slope +
           t2(hill_shannon, bs="tp", k=4) +
           #agriculture +
           grazing,
           #logging +
           random = ~(1|site), data = plots_cf_4p)

gamm4_4c <- gamm4(log1p(agb_plot_ha) ~ env_pc1 +
           slope +
           t2(hill_shannon, bs="tp", k=4) +
           #agriculture +
           #grazing +
           logging,
           random = ~(1|site), data = plots_cf_4p)

gamm4_4d <- gamm4(log1p(agb_plot_ha) ~ env_pc1 +
           slope +
           #hill_shannon +
           agriculture +
           grazing,
           #logging +
           random = ~(1|site), data = plots_cf_4p)

gamm4_4e <- gamm4(log1p(agb_plot_ha) ~ env_pc1 +
           slope +
           #hill_shannon +
           agriculture +
           #grazing +
           logging,
           random = ~(1|site), data = plots_cf_4p)

gamm4_4f <- gamm4(log1p(agb_plot_ha) ~ env_pc1 +
           slope +
           #hill_shannon +
           #agriculture +
           grazing +
           logging,
           random = ~(1|site), data = plots_cf_4p)

gamm4_4g <- gamm4(log1p(agb_plot_ha) ~ env_pc1 +
           #slope +
           hill_shannon +
           agriculture +
           grazing,
           #logging +
           random = ~(1|site), data = plots_cf_4p)

gamm4_4h <- gamm4(log1p(agb_plot_ha) ~ env_pc1 +
           #slope +
           t2(hill_shannon, bs="tp", k=4) +
           agriculture +
           #grazing +
           logging,
           random = ~(1|site), data = plots_cf_4p)

gamm4_4i <- gamm4(log1p(agb_plot_ha) ~ env_pc1 +
           #slope +
           t2(hill_shannon, bs="tp", k=4) +
           #agriculture +
           grazing +
           logging,
           random = ~(1|site), data = plots_cf_4p)

gamm4_4j <- gamm4(log1p(agb_plot_ha) ~ env_pc1 +
           #slope +
           #hill_shannon +
           agriculture +
           grazing +
           logging,
           random = ~(1|site), data = plots_cf_4p)

gamm4_4k <- gamm4(log1p(agb_plot_ha) ~ #env_pc1 +
           slope +
           t2(hill_shannon, bs="tp", k=4) +
           agriculture +
           grazing,
           #logging +
           random = ~(1|site), data = plots_cf_4p)

gamm4_4l <- gamm4(log1p(agb_plot_ha) ~ #env_pc1 +
           slope +
           t2(hill_shannon, bs="tp", k=4) +
           agriculture +
           #grazing +
           logging,
           random = ~(1|site), data = plots_cf_4p)

gamm4_4m <- gamm4(log1p(agb_plot_ha) ~ #env_pc1 +
           slope +
           t2(hill_shannon, bs="tp", k=4) +
           #agriculture +
           grazing +
           logging,
           random = ~(1|site), data = plots_cf_4p)

gamm4_4n <- gamm4(log1p(agb_plot_ha) ~ #env_pc1 +
           slope +
           #hill_shannon +
           agriculture +
           grazing +
           logging,
           random = ~(1|site), data = plots_cf_4p)

gamm4_4o <- gamm4(log1p(agb_plot_ha) ~ #env_pc1 +
           #slope +
           t2(hill_shannon, bs="tp", k=4) +
           agriculture +
           grazing +
           logging,
           random = ~(1|site), data = plots_cf_4p)

aic4a <- cAIC(gamm4_4a)
aic4b <- cAIC(gamm4_4b)
aic4c <- cAIC(gamm4_4c)
aic4d <- cAIC(gamm4_4d)
aic4e <- cAIC(gamm4_4e)
aic4f <- cAIC(gamm4_4f)
#aic4g <- cAIC(gamm4_4g)
aic4h <- cAIC(gamm4_4h)
aic4i <- cAIC(gamm4_4i)
aic4j <- cAIC(gamm4_4j)
aic4k <- cAIC(gamm4_4k)
aic4l <- cAIC(gamm4_4l)
aic4m <- cAIC(gamm4_4m)
aic4n <- cAIC(gamm4_4n)
aic4o <- cAIC(gamm4_4o)

#gamms4_aic <- c("gamm4_4a", aic4a$caic,
                # "gamm4_4b", aic4b$caic,
                # "gamm4_4c", aic4c$caic,
                # "gamm4_4d", aic4d$caic,
                # "gamm4_4e", aic4e$caic,
                # "gamm4_4f", aic4f$caic,
                # "gamm4_4g", aic4g$caic,
                # "gamm4_4h", aic4h$caic,
                # "gamm4_4i", aic4i$caic,
                # "gamm4_4j", aic4j$caic,
                # "gamm4_4k", aic4k$caic,
                # "gamm4_4l", aic4l$caic,
                # "gamm4_4m", aic4m$caic,
                # "gamm4_4n", aic4n$caic,
                # "gamm4_4o", aic4o$caic)

# Best model with 4 predictors is agb ~ env + slope + richness + agriculture cAIC = 434.939
```
## GAMM: 5 predictors

```{r}
gamm4_5a <- gamm4(log1p(agb_plot_ha) ~ env_pc1 +
           slope +
           t2(hill_shannon, bs="tp", k=4) +
           agriculture +
           grazing,
           #logging +
           random = ~(1|site), data = plots_cf_4p)

gamm4_5b <- gamm4(log1p(agb_plot_ha) ~ env_pc1 +
           slope +
           t2(hill_shannon, bs="tp", k=4) +
           agriculture +
           #grazing +
           logging,
           random = ~(1|site), data = plots_cf_4p)

gamm4_5c <- gamm4(log1p(agb_plot_ha) ~ env_pc1 +
           slope +
           t2(hill_shannon, bs="tp", k=4) +
           #agriculture +
           grazing +
           logging,
           random = ~(1|site), data = plots_cf_4p)

gamm4_5d <- gamm4(log1p(agb_plot_ha) ~ env_pc1 +
           slope +
           #hill_shannon +
           agriculture +
           grazing +
           logging,
           random = ~(1|site), data = plots_cf_4p)

gamm4_5e <- gamm4(log1p(agb_plot_ha) ~ env_pc1 +
           #slope +
           t2(hill_shannon, bs="tp", k=4) +
           agriculture +
           grazing +
           logging,
           random = ~(1|site), data = plots_cf_4p)

gamm4_5f <- gamm4(log1p(agb_plot_ha) ~ #env_pc1 +
           slope +
           t2(hill_shannon, bs="tp", k=4) +
           agriculture +
           grazing +
           logging,
           random = ~(1|site), data = plots_cf_4p)

aic5a <- cAIC(gamm4_5a)
aic5b <- cAIC(gamm4_5b)
aic5c <- cAIC(gamm4_5c)
aic5d <- cAIC(gamm4_5d)
aic5e <- cAIC(gamm4_5e)
aic5f <- cAIC(gamm4_5f)

gamms5_aic <- c("gamm4_5a", aic5a$caic,
                "gamm4_5b", aic5b$caic,
                "gamm4_5c", aic5c$caic,
                "gamm4_5d", aic5d$caic,
                "gamm4_5e", aic5e$caic,
                "gamm4_5f", aic5f$caic)

# Best model with 5 predictors is agb ~ env + slope + sp_richness + agriculture + logging cAIC = 435.350
```

## GAMM: 6 predictors

```{r}

gamm4_6 <- gamm4(log1p(agb_plot_ha) ~ env_pc1 +
           slope +
           t2(hill_shannon, bs="tp", k=4) +
           agriculture +
           grazing +
           logging,
           random = ~(1|site), data = plots_cf_4p)

cAIC(gamm4_6)

# Model with 6 predictors cAIC= 439.09


```

Best models are:
1 predictor. gamm4_1c. cAIC = 439.3378
2 predictors. gamm4_2p. cAIC = 437.158 
3 predictors. gamm4_3k. cAIC= 434.859
4 predictors. gamm4_4a. cAIC = 434.939
5 predictors. gamm4_5b. cAIC = 435.350
6 predictors. gamm4_6. cAIC= 436.12

```{r}
summary(gamm4_1c)
summary(gamm4_1c$mer)
summary(gamm4_1c$gam)
plot.gam(gamm4_1c$gam, residuals=T, se=T, shade=T, pch=16, cex=0.5)
gam.check(gamm4_1c$gam)

plot.gam(gamm4_2p$gam, residuals=T, se=T, shade=T, pch=16, cex=0.5)
gam.check(gamm4_2p$gam)

plot.gam(gamm4_3k$gam, residuals=T, se=T, shade=T, pch=16, cex=0.5)
gam.check(gamm4_3k$gam)
summary(gamm4_3k$mer)
summary(gamm4_3k$gam)

summary(gamm4_4a$gam)

summary(gamm4_5b$gam)

summary(gamm4_6$gam)
```

## GAMM: Interactions

```{r}

gamm4_4_int <- gamm4(log1p(agb_plot_ha) ~ env_pc1*agriculture +
           slope +
           t2(hill_shannon*agriculture, bs= "tp", k= 4),
           random = ~(1|site), data = plots_cf_4p)

gamm4_4_int2 <- gamm4(log1p(agb_plot_ha) ~ env_pc1*agriculture +
           slope +
           t2(hill_shannon, bs= "tp", k= 4),
           random = ~(1|site), data = plots_cf_4p)

gamm4_4_int5 <- gamm4(log1p(agb_plot_ha) ~ env_pc1*agriculture +
           slope*agriculture +
           t2(hill_shannon, bs= "tp", k= 4),
           random = ~(1|site), data = plots_cf_4p)


gamm4_4_int3 <- gamm4(log1p(agb_plot_ha) ~ env_pc1*agriculture +
           slope +
           s(hill_shannon, bs= "tp", k= 4),
           random = ~(1|site), data = plots_cf_4p)

gamm4_4_int4 <- gamm4(log1p(agb_plot_ha) ~ env_pc1*agriculture +
           slope +
           forest_classification +
           t2(hill_shannon, bs= "tp", k= 4),
           random = ~(1|site), data = plots_cf_4p)

gamm4_4_int2 <- gamm4(log1p(agb_plot_ha) ~ env_pc1*agriculture +
           slope +
           t2(hill_shannon, bs= "tp", k= 4),
           random = ~(1|site), data = plots_cf_4p)


summary(gamm4_4_int$gam)

cAIC(gamm4_4_int)

plot(gamm4_4_int$gam, residuals=T, se=T, shade=T, pch=16, cex=0.5)


ag_env <- gamm4(agriculture ~ t2(env_pc1, bs="tp", k=4), random = ~(1|site), data = plots_cf_4p)
summary(ag_env$gam)

sp_env <- gamm4(hill_richness ~ t2(env_pc1, bs="tp", k=4), random = ~(1|site), data = plots_cf_4p)
summary(sp_env$gam)
plot(sp_env$gam)

sp_ag <- gamm4(hill_richness ~ t2(agriculture, bs="tp", k=4), random = ~(1|site), data = plots_cf_4p)
summary(sp_ag$gam)
plot(sp_ag$gam)

plot_model(gamm4_4_int, show.p = T, show.values = T)

gam.check(gamm4_4_int$gam)
plot.gam(gamm4_4_int$gam, residuals=T, se=T, shade=T, pch=16, cex=0.5)
plot.gam(gamm4_4_int2$gam, residuals=T, se=T, shade=T, pch=16, cex=0.5)
summary(gamm4_4_int$gam)
summary(gamm4_4_int2$gam)


```

## AGB distribution in TMCF

Tree size classes and contribution to AGB

```{r}

# Estimate total agb per class and plot

agb_class_plots <- trees_cf_4p %>%
  right_join(plots_cf_4p, by = "plot_id") %>%
  filter(status == "Vivo") %>%
  filter(is.na(life_form) | life_form == "Arbol") %>%
  dplyr::select(plot_id, species, genus, sp, family, diam_bh, agb, forest_classification) %>%
  mutate(diam_class = ifelse(diam_bh <= 10, "1",
                             ifelse(diam_bh > 10 & diam_bh <= 20, "2",
                                    ifelse(diam_bh > 20 & diam_bh<= 30, "3",
                                           ifelse(diam_bh > 30 & diam_bh <= 40, "4",
                                                  ifelse(diam_bh > 40 & diam_bh <= 50, "5", "6")))))) %>%
  filter(!is.na(agb)) %>%
  group_by(plot_id, diam_class) %>%
  summarise(agb_class_plot = sum(agb))

agb_class_plots
```

```{r}
# Contribution of tree size to AGB

agbprop <- plots_cf_4p %>%
  dplyr::select(plot_id, agb_plot) %>%
  right_join(agb_class_plots, by = "plot_id") %>%
  mutate(agb_prop = agb_class_plot/agb_plot)  %>%
  mutate(stem_prop_logit = logit(agb_prop))

lmm_agb_diam <- lmer(stem_prop_logit ~ diam_class + (1|site/plot_id), data = agbprop)
summary(lmm_agb_diam)
stargazer(lmm_agb_diam, type = "text")
summary(glht(lmm_agb_diam, linfct = mcp(diam_class = "Tukey")), test = adjusted("holm"))
  #group_by(diam_class) %>%
  #get_summary_stats(agb_prop, type = "mean_sd") 

# Plot

plots_cf_4p %>%
  dplyr::select(plot_id, agb_plot) %>%
  right_join(agb_class_plots, by = "plot_id") %>%
  mutate(agb_prop = agb_class_plot/agb_plot)  %>%
  group_by(diam_class) %>%
  ggplot(aes(x = diam_class, y = agb_prop)) +
  geom_boxplot() +
  theme_classic() +
  geom_text(aes(x= 3.5,
                y= -0.05,
                label = "a                        b                     b                        b                        c                        d",),
            size = 4,
            color = "gray30") +
  theme_classic() +
  labs(x = "Size class",
       y = "Proportion of AGB") +
  scale_y_continuous(breaks = c(0.0, 0.25, 0.50, 0.75, 1.0))

# Anova

# agb_prop <- plots_cf_4p %>%
#   dplyr::select(plot_id, agb_plot) %>%
#   right_join(agb_class_plots, by = "plot_id") %>%
#   mutate(agb_prop = agb_class_plot/agb_plot) %>%
#   arrange(diam_class)
# 
# pairwise.wilcox.test(x= agbprop$stem_prop_logit, g= agbprop$diam_class, p.adjust.method = "bonferroni")
# 
# anova <- aov(agb_prop ~ diam_class, data = agb_prop)
# summary(anova)
# TukeyHSD(anova)
# tukey_hsd(anova)
# 
# multcompLetters4(anova, (TukeyHSD(anova)))  

```

Big trees (size class 6) contributes the most to AGB. Classes 1, 5 and 6 are statistically different from the others.

Tree size classes and contribution to total number of stems:

```{r}
# Contribution of size to stems

stems_class_plot <- trees_cf_4p %>%
  right_join(plots_cf_4p, by = "plot_id") %>%
  filter(status == "Vivo") %>%
  filter(is.na(life_form) | life_form == "Arbol") %>%
  dplyr::select(plot_id, species, genus, sp, family, diam_bh, agb) %>%
  mutate(diam_class = ifelse(diam_bh <= 10, "1",
                             ifelse(diam_bh > 10 & diam_bh <= 20, "2",
                                    ifelse(diam_bh > 20 & diam_bh<= 30, "3", 
                                           ifelse(diam_bh > 30 & diam_bh <= 40, "4",
                                                  ifelse(diam_bh > 40 & diam_bh <= 50, "5", "6")))))) %>%
  filter(!is.na(agb)) %>%
  group_by(plot_id, diam_class) %>%
  summarise(stems = n())

stems_class_plot
```

Contribution of tree size classes to stem density:

```{r}
stemprop <- plots_cf_4p %>%
  dplyr::select(plot_id, tree_no) %>%
  right_join(stems_class_plot, by = "plot_id") %>%
  mutate(stem_prop = stems/tree_no)  %>%  
  mutate(stem_prop_logit = logit(stem_prop))

lmm_stem_diam <- lmer(stem_prop_logit ~ diam_class + (1|site/plot_id), data = stemprop)
summary(lmm_stem_diam)
stargazer(lmm_stem_diam, type = "text")
summary(glht(lmm_stem_diam, linfct = mcp(diam_class = "Tukey")), test = adjusted("holm"))


bp <- plots_cf_4p %>%
  dplyr::select(plot_id, tree_no) %>%
  right_join(stems_class_plot, by = "plot_id") %>%
  mutate(stem_prop = stems/tree_no)  %>%
  group_by(diam_class)

bp %>%
  ggplot(aes(x = diam_class, y = stem_prop)) +
  geom_boxplot() +
  geom_text(aes(x= 3.5,
                y= -0.05,
                label = "a                       b                       c                        d                       de                      e",),
            size = 4,
            color = "gray30") + 
  theme_classic() +
  labs(x = "Size class",
       y = "Proportion of stems") +
  scale_y_continuous(breaks = c(0.0, 0.25, 0.50, 0.75, 1.0))

# Anova

# stems_prop <- plots_cf_4p %>%
#   select(plot_id, tree_no) %>%
#   right_join(stems_class_plot, by = "plot_id") %>%
#   mutate(stems_prop = stems/tree_no) %>%
#   arrange(diam_class)
# 
# pairwise.wilcox.test(x= stems_prop$stems_prop, g= stems_prop$diam_class, p.adjust.method = "bonferroni")
# 
# anova2 <- aov(stems_prop ~ diam_class, data = stems_prop)
# summary(anova2)
# TukeyHSD(anova2)
# tukey_hsd(anova2) %>%
#   get_test_label()
# 
# multcompLetters4(anova2, (TukeyHSD(anova2)))
```

Trees of class 2 (10-20 cm of DBH) are the most abundant. All classes are statistically different except for classes 4 and 5, and 5 and 6.

### Tree size contribution to AGB and stems in relation to disturbance level

Contribution to AGB

```{r}

plots_cf_4p %>%
  dplyr::select(plot_id, agb_plot, forest_classification) %>%
  right_join(agb_class_plots, by = "plot_id") %>%
  mutate(agb_prop = agb_class_plot/agb_plot)  %>%
  group_by(forest_classification, diam_class) %>%
  ggplot(aes(x = diam_class, y = agb_prop, fill = factor(forest_classification, levels = c('undisturbed', 'regrowth', 'degraded', 'deforested'),ordered = TRUE))) +
  geom_boxplot() +
  theme_classic() +
  labs(fill ='Disturbance Class') +
  scale_fill_brewer(palette = "Accent")

# plots_cf_4p %>%
#   select(plot_id, forest_classification) %>%
#   right_join(agbprop, by = "plot_id") %>%
#   select(-site.x, -site.y) %>%
#   group_by(forest_classification, diam_class) %>%
#   get_summary_stats(agb_prop, type = "mean_sd") # remove rstatix package if not running these lines

agb_prop_fc <- plots_cf_4p %>%
  dplyr::select(plot_id, forest_classification) %>%
  right_join(agbprop, by = "plot_id") %>%
  mutate(site = site.x) %>%
  dplyr::select(-site.x, -site.y) %>%
  # mutate(succ = ifelse(forest_classification == "1", "s1",
  #                      ifelse(forest_classification == "2", "s2", "s3"))) %>%
  mutate(diam_class = ifelse(diam_class == 1, "diam1",
                             ifelse(diam_class == 2, "diam2",
                                    ifelse(diam_class == 3, "diam3",
                                           ifelse(diam_class == 4, "diam4",
                                                  ifelse(diam_class == 5, "diam5", "diam6")))))) %>%
  group_by(forest_classification, diam_class)


int <- interaction(agb_prop_fc$diam_class, agb_prop_fc$forest_classification)
lmm_agb_diam_fc <- lmer(agb_prop_logit ~ int + (1|site/plot_id), data = agb_prop_fc)
summary(lmm_agb_diam_fc)
summary(glht(lmm_agb_diam_fc, linfct = mcp(int = "Tukey"), test = adjusted("holm")))
# Differences between diam_class within disturbance level class
diam1 <- agb_prop_fc %>%
  filter(diam_class == "diam1")
summary(lmer(agb_prop_logit ~ forest_classification + (1|site), data = diam1))
summary(glht(lmer(agb_prop_logit ~ forest_classification + (1|site), data = diam1), linfct = mcp(forest_classification = "Tukey"), test = adjusted("holm")))
# summary(comp)
# comp_df <- as_tibble(comp[["linfct"]], rownames = "comps")

lmm_agb_diam_fc <- lmer(agb_prop_logit ~ diam_class*forest_classification + (1|site/plot_id), data = agb_prop_fc)
# summary(lmm_agb_diam_str)
#stargazer(lmm_agb_diam_fc, type = "text")
summary(glht(lmm_agb_diam_fc, linfct = mcp(diam_class= "Tukey", forest_classification = "Tukey")), test = adjusted("holm"))

diam1 <- agb_prop_fc %>%
  filter(diam_class == "diam1")
summary(lmer(agb_prop_logit ~ forest_classification + (1|site), data = diam1))
summary(glht(lmer(agb_prop_logit ~ forest_classification + (1|site), data = diam1), linfct = mcp(forest_classification = "Tukey"), test = adjusted("holm")))

diam2 <- agb_prop_fc %>%
  filter(diam_class == "diam2")
summary(lmer(agb_prop_logit ~ forest_classification + (1|site), data = diam2))
summary(glht(lmer(agb_prop_logit ~ forest_classification + (1|site), data = diam2), linfct = mcp(forest_classification = "Tukey"), test = adjusted("holm")))

diam3 <- agb_prop_fc %>%
  filter(diam_class == "diam3")
summary(lmer(agb_prop_logit ~ forest_classification + (1|site), data = diam3))
summary(glht(lmer(agb_prop_logit ~ forest_classification + (1|site), data = diam3), linfct = mcp(forest_classification = "Tukey"), test = adjusted("holm")))

diam4 <- agb_prop_fc %>%
  filter(diam_class == "diam4")
summary(lmer(agb_prop_logit ~ forest_classification + (1|site), data = diam4))
summary(glht(lmer(agb_prop_logit ~ forest_classification + (1|site), data = diam4), linfct = mcp(forest_classification = "Tukey"), test = adjusted("holm")))

diam5 <- agb_prop_fc %>%
  filter(diam_class == "diam5")
summary(lmer(agb_prop_logit ~ forest_classification + (1|site), data = diam5))
summary(glht(lmer(agb_prop_logit ~ forest_classification + (1|site), data = diam5), linfct = mcp(forest_classification = "Tukey"), test = adjusted("holm")))

diam6 <- agb_prop_fc %>%
  filter(diam_class == "diam6")
summary(lmer(agb_prop_logit ~ forest_classification + (1|site), data = diam6))
summary(glht(lmer(agb_prop_logit ~ forest_classification + (1|site), data = diam6), linfct = mcp(forest_classification = "Tukey"), test = adjusted("holm")))

agbprop_undis <- agb_prop_fc %>%
  filter(forest_classification == "undisturbed")
agbprop_und_lmm <- lmer(agb_prop_logit ~ diam_class + (1|site), data = agbprop_undis)
summary(agbprop_und_lmm)
stargazer(agbprop_und_lmm, type = "text", report=("vc*p"))
summary(glht(lmer(agb_prop_logit ~ diam_class + (1|site), data = agbprop_undis), linfct = mcp(diam_class = "Tukey"), test = adjusted("holm")))


agbprop_reg <- agb_prop_fc %>%
  filter(forest_classification == "regrowth")
agbprop_reg_lmm <- lmer(agb_prop_logit ~ diam_class + (1|site), data = agbprop_reg)
summary(agbprop_reg_lmm)
stargazer(agbprop_reg_lmm, type = "text", report=("vc*p"))
summary(glht(lmer(agb_prop_logit ~ diam_class + (1|site), data = agbprop_reg), linfct = mcp(diam_class = "Tukey"), test = adjusted("holm")))

agbprop_deg <- agb_prop_fc %>%
  filter(forest_classification == "degraded")
agbprop_deg_lmm <- lmer(agb_prop_logit ~ diam_class + (1|site), data = agbprop_deg)
summary(agbprop_deg_lmm)
stargazer(agbprop_deg_lmm, type = "text", report=("vc*p"))
summary(glht(lmer(agb_prop_logit ~ diam_class + (1|site), data = agbprop_deg), linfct = mcp(diam_class = "Tukey"), test = adjusted("holm")))


agbprop_def <- agb_prop_fc %>%
  filter(forest_classification == "deforested")
agbprop_def_lmm <- lmer(agb_prop_logit ~ diam_class + (1|site), data = agbprop_def)
summary(agbprop_def_lmm)
stargazer(agbprop_def_lmm, type = "text", report=("vc*p"))
summary(glht(lmer(agb_prop_logit ~ diam_class + (1|site), data = agbprop_def), linfct = mcp(diam_class = "Tukey"), test = adjusted("holm")))
```

Contribution to Stem Number

```{r}
plots_cf_4p %>%
  dplyr::select(plot_id, tree_no, forest_classification) %>%
  right_join(stems_class_plot, by = "plot_id") %>%
  mutate(stem_prop = stems/tree_no)  %>%
  group_by(forest_classification, diam_class) %>%
  ggplot(aes(x = diam_class, y = stem_prop, fill = factor(forest_classification, levels = c('undisturbed', 'regrowth', 'degraded', 'deforested'), ordered = TRUE))) +
  geom_boxplot() +
  theme_classic() +
  labs(fill ='Disturbance Class') +
  scale_fill_brewer(palette = "Accent")

stem_prop_str <- plots_cf_4p %>%
  dplyr::select(plot_id, forest_classification) %>%
  right_join(stemprop, by = "plot_id") %>%
  mutate(site = site.x) %>%
  dplyr::select(-site.x, -site.y) %>%
  mutate(diam_class = ifelse(diam_class == 1, "diam1",
                             ifelse(diam_class == 2, "diam2",
                                    ifelse(diam_class == 3, "diam3",
                                           ifelse(diam_class == 4, "diam4",
                                                  ifelse(diam_class == 5, "diam5", "diam6")))))) %>%
  group_by(forest_classification, diam_class)


int_stem <- interaction(stem_prop_str$diam_class, stem_prop_str$forest_classification)
lmm_stem_diam_str <- lmer(stem_prop_logit ~ int_stem + (1|site/plot_id), data = stem_prop_str)
summary(lmm_stem_diam_str)
summary(glht(lmm_stem_diam_str, linfct = mcp(int_stem = "Tukey"), test = adjusted("holm")))



plots_cf_4p %>%
  select(plot_id, forest_classification) %>%
  right_join(stems_prop, by = "plot_id") %>%
  select(-site.x, -site.y) %>%
  group_by(forest_classification, diam_class) %>%
  get_summary_stats(stems_prop, type = "mean_sd") 

diam1 <- stem_prop_str %>%
  filter(diam_class == "diam1")
summary(lmer(stem_prop_logit ~ forest_classification + (1|site), data = diam1))
summary(glht(lmer(stem_prop_logit ~ forest_classification + (1|site), data = diam1), linfct = mcp(forest_classification = "Tukey"), test = adjusted("holm")))

diam2 <- stem_prop_str %>%
  filter(diam_class == "diam2")
summary(lmer(stem_prop_logit ~ forest_classification + (1|site), data = diam2))
summary(glht(lmer(stem_prop_logit ~ forest_classification + (1|site), data = diam2), linfct = mcp(forest_classification = "Tukey"), test = adjusted("holm")))

diam3 <- stem_prop_str %>%
  filter(diam_class == "diam3")
summary(lmer(stem_prop_logit ~ forest_classification + (1|site), data = diam3))
summary(glht(lmer(stem_prop_logit ~ forest_classification + (1|site), data = diam3), linfct = mcp(forest_classification = "Tukey"), test = adjusted("holm")))

diam4 <- stem_prop_str %>%
  filter(diam_class == "diam4")
summary(lmer(stem_prop_logit ~ forest_classification + (1|site), data = diam4))
summary(glht(lmer(stem_prop_logit ~ forest_classification + (1|site), data = diam4), linfct = mcp(forest_classification = "Tukey"), test = adjusted("holm")))

diam5 <- stem_prop_str %>%
  filter(diam_class == "diam5")
summary(lmer(stem_prop_logit ~ forest_classification + (1|site), data = diam5))
summary(glht(lmer(stem_prop_logit ~ forest_classification + (1|site), data = diam5), linfct = mcp(forest_classification = "Tukey"), test = adjusted("holm")))

diam6 <- stem_prop_str %>%
  filter(diam_class == "diam6")
summary(lmer(stem_prop_logit ~ forest_classification + (1|site), data = diam6))
summary(glht(lmer(stem_prop_logit ~ forest_classification + (1|site), data = diam6), linfct = mcp(forest_classification = "Tukey"), test = adjusted("holm")))

stemprop_undis <- stem_prop_str %>%
  filter(forest_classification == "undisturbed")
stemprop_undis_lmm <- lmer(stem_prop_logit ~ diam_class + (1|site), data = stemprop_undis)
summary(stemprop_undis_lmm)
stargazer(stemprop_undis_lmm, type = "text", report=("vc*p"))
summary(glht(lmer(stem_prop_logit ~ diam_class + (1|site), data = stemprop_undis), linfct = mcp(diam_class = "Tukey"), test = adjusted("holm")))

stemprop_reg <- stem_prop_str %>%
  filter(forest_classification == "regrowth")
stemprop_reg_lmm <- lmer(stem_prop_logit ~ diam_class + (1|site), data = stemprop_reg)
summary(stemprop_reg_lmm)
stargazer(stemprop_reg_lmm, type = "text", report=("vc*p"))
summary(glht(lmer(stem_prop_logit ~ diam_class + (1|site), data = stemprop_reg), linfct = mcp(diam_class = "Tukey"), test = adjusted("holm")))

stemprop_deg <- stem_prop_str %>%
  filter(forest_classification == "degraded")
stemprop_deg_lmm <- lmer(stem_prop_logit ~ diam_class + (1|site), data = stemprop_deg)
summary(stemprop_deg_lmm)
stargazer(stemprop_deg_lmm, type = "text", report=("vc*p"))
summary(glht(lmer(stem_prop_logit ~ diam_class + (1|site), data = stemprop_deg), linfct = mcp(diam_class = "Tukey"), test = adjusted("holm")))

stemprop_def <- stem_prop_str %>%
  filter(forest_classification == "deforested")
stemprop_def_lmm <- lmer(stem_prop_logit ~ diam_class + (1|site), data = stemprop_def)
summary(stemprop_def_lmm)
stargazer(stemprop_def_lmm, type = "text", report=("vc*p"))
summary(glht(lmer(stem_prop_logit ~ diam_class + (1|site), data = stemprop_def), linfct = mcp(diam_class = "Tukey"), test = adjusted("holm")))
```

### Structural attributes, AGB and diversity across disturbance levels

```{r}

# Tree density

# Min, Mean, and Max values:
plots_cf_4p %>%
  group_by(forest_classification) %>%
  get_summary_stats(tree_density, type = "min")

plots_cf_4p %>%
  group_by(forest_classification) %>%
  get_summary_stats(tree_density, type = "mean_sd")

plots_cf_4p %>%
  group_by(forest_classification) %>%
  get_summary_stats(tree_density, type = "max")

# LMM:
td_fc <- lmer(tree_density ~ forest_classification + (1|site), data = plots_cf_4p)
summary(td_fc)
stargazer(td_fc, type = "text")
stargazer(td_fc,type="text",report=("vc*p"))
plot_model(td_fc, show.p = T, show.values = T)
summary(glht(td_fc, linfct = mcp(forest_classification = "Tukey")), test = adjusted("holm"))

# Tree height

# Min, Mean, and Max values:
plots_cf_4p %>%
  group_by(forest_classification) %>%
  get_summary_stats(loreys_height, type = "min")

plots_cf_4p %>%
  group_by(forest_classification) %>%
  get_summary_stats(loreys_height, type = "mean_sd")

plots_cf_4p %>%
  group_by(forest_classification) %>%
  get_summary_stats(loreys_height, type = "max")

# LMM:

th_fc <- lmer(loreys_height ~ forest_classification + (1|site), data = plots_cf_4p)
summary(th_fc)
stargazer(th_fc, type = "text")
summary(glht(th_fc, linfct = mcp(forest_classification = "Tukey")), test = adjusted("holm"))


# Basal area

# Min, Mean, and Max values:
plots_cf_4p %>%
  group_by(forest_classification) %>%
  get_summary_stats(basal_area_ha, type = "min")

plots_cf_4p %>%
  group_by(forest_classification) %>%
  get_summary_stats(basal_area_ha, type = "mean_sd")

plots_cf_4p %>%
  group_by(forest_classification) %>%
  get_summary_stats(basal_area_ha, type = "max")

# LMM:

ba_fc <- lmer(basal_area_ha ~ forest_classification + (1|site), data = plots_cf_4p)
summary(ba_fc)
stargazer(ba_fc, type = "text")
summary(glht(ba_fc, linfct = mcp(forest_classification = "Tukey")), test = adjusted("holm"))

# Wood density

# Min, Mean, and Max values:

trees_cf_4p %>%
  filter(!is.na(mean_wd)) %>%
  dplyr::group_by(plot_id, species) %>%
  dplyr::summarize(abundance = sum(n()),
                   mean_wd = as.numeric(mean_wd)) %>%
  group_by(plot_id) %>%
  summarize(mean_wd_plot = weighted.mean(mean_wd, abundance)) %>%
  right_join(plots_cf_4p) %>%
  filter(!is.na(mean_wd_plot)) %>%
  group_by(forest_classification) %>%
  summarize(min = min(mean_wd_plot),
            mean = mean(mean_wd_plot),
            sd = sd(mean_wd_plot),
            max = max(mean_wd_plot))
  
# LMM:

wd_df <- trees_cf_4p %>%
  filter(!is.na(mean_wd)) %>%
  dplyr::group_by(plot_id, species) %>%
  dplyr::summarize(abundance = sum(n()),
                   mean_wd = as.numeric(mean_wd)) %>%
  group_by(plot_id) %>%
  summarize(mean_wd_plot = weighted.mean(mean_wd, abundance)) %>%
  right_join(plots_cf_4p) %>%
  filter(!is.na(mean_wd_plot)) %>%
  select(site, plot_id, forest_classification, mean_wd_plot)

wd_fc <- lmer(mean_wd_plot ~ forest_classification + (1|site), data = wd_df)
summary(wd_fc)
stargazer(wd_fc, type = "text")
summary(glht(wd_fc, linfct = mcp(forest_classification = "Tukey")), test = adjusted("holm"))

# AGB

# Min, Mean, and Max values:

plots_cf_4p %>%
  group_by(forest_classification) %>%
  get_summary_stats(agb_plot_ha, type = "min")

plots_cf_4p %>%
  group_by(forest_classification) %>%
  get_summary_stats(agb_plot_ha, type = "mean_sd")

plots_cf_4p %>%
  group_by(forest_classification) %>%
  get_summary_stats(agb_plot_ha, type = "max")


# LMM:
agb_fc <- lmer(agb_plot_ha ~ forest_classification + (1|site), data = plots_cf_4p)
summary(agb_fc)
stargazer(agb_fc, type = "text")
summary(glht(agb_fc, linfct = mcp(forest_classification = "Tukey")), test = adjusted("holm"))


# Shannon

# Min, Mean, and Max values:

plots_cf_4p %>%
  group_by(forest_classification) %>%
  get_summary_stats(hill_shannon, type = "min")

plots_cf_4p %>%
  group_by(forest_classification) %>%
  get_summary_stats(hill_shannon, type = "mean_sd")

plots_cf_4p %>%
  group_by(forest_classification) %>%
  get_summary_stats(hill_shannon, type = "max")

# LMM:
hs_fc <- lmer(hill_shannon ~ forest_classification + (1|site), data = plots_cf_4p)
summary(hs_fc)
stargazer(hs_fc, type = "text")
summary(glht(hs_fc, linfct = mcp(forest_classification = "Tukey")), test = adjusted("holm"))

# All plots

min(plots_cf_4p$tree_density)
mean(plots_cf_4p$tree_density)
sd(plots_cf_4p$tree_density)
max(plots_cf_4p$tree_density)

min(plots_cf_4p$loreys_height)
mean(plots_cf_4p$loreys_height)
sd(plots_cf_4p$loreys_height)
max(plots_cf_4p$loreys_height)

min(plots_cf_4p$basal_area_ha)
mean(plots_cf_4p$basal_area_ha)
sd(plots_cf_4p$basal_area_ha)
max(plots_cf_4p$basal_area_ha)

min(plots_cf_4p$agb_plot_ha)
mean(plots_cf_4p$agb_plot_ha)
sd(plots_cf_4p$agb_plot_ha)
max(plots_cf_4p$agb_plot_ha)

trees_cf_4p %>%
  filter(!is.na(mean_wd)) %>%
  dplyr::group_by(plot_id, species) %>%
  dplyr::summarize(abundance = sum(n()),
                   mean_wd = as.numeric(mean_wd)) %>%
  group_by(plot_id) %>%
  summarize(meanwd = mean(mean_wd)) %>%
  summarize(sdwd = sd(meanwd))
  
min(plots_cf_4p$hill_shannon, na.rm = T)
mean(plots_cf_4p$hill_shannon, na.rm = T)
sd(plots_cf_4p$hill_shannon, na.rm = T)
max(plots_cf_4p$hill_shannon, na.rm = T)

plots_cf_4p %>%
  ungroup() %>%
  select(agb_plot_ha, forest_classification) %>%
  filter(!is.na(agb_plot_ha), !is.na(forest_classification)) %>%
  t_test(agb_plot_ha ~ forest_classification, p.adjust.method = "bonferroni") %>%
  add_significance("p.adj") 

plots_cf_4p %>%
  ggplot(aes(x = forest_classification, y = log1p(agb_plot_ha))) +
  geom_boxplot()

plots_cf_4p %>%
  #filter(hill_shannon < 5) %>%
  ggplot(aes(x = forest_classification, y = sqrt(hill_shannon))) +
  geom_boxplot()

plots_cf_4p %>%
  #filter(hill_shannon < 5) %>%
  ggplot(aes(x = agriculture, y = hill_shannon)) +
  geom_point()

plots_cf_4p %>%
  #filter(hill_shannon < 5) %>%
  ggplot(aes(x = agriculture, y = log1p(agb_plot_ha))) +
  geom_point()

plots_cf_4p %>%
  #filter(hill_shannon < 5) %>%
  ggplot(aes(x = sqrt(hill_shannon), y = log1p(agb_plot_ha), color = forest_classification)) +
  geom_point()
```
