==> 06_figures.Rmd <==
---
title: "06_figures"
author: "Adriana Uscanga"
date: "19/2/2022"
output: html_document
---

# Figures
 
```{r}
# Note: Run 05_analysis first to have all needed data loaded

# Load packages for making figures:
library(ggplot2)
library(ggpubr)
library(cowplot)
library(rlang)
library(ggsignif)
library(ggbreak)
```

## Figure 2. 


```{r}

# Structural attributes across disturbance classes

## Panel a: tree density 

pa <- plots_cf_4p %>%
  ggplot(aes(x= factor(forest_classification, levels = c('undisturbed', 'regrowth', 'degraded', 'deforested'),ordered = TRUE),
             y = tree_density,
             fill = factor(forest_classification, levels = c('undisturbed', 'regrowth', 'degraded', 'deforested'),ordered = TRUE))) +
  geom_boxplot(color = "gray25",
               outlier.alpha = 0.5,
               outlier.color = "gray40",
               outlier.size = 1.2) +
  theme_classic() +
  scale_fill_manual(values = c("#1b9e77", "#7570b3", "#e7298a", "#d95f02")) +
  theme(axis.title.x = element_blank()) +
  labs(y = (bquote('Tree density ( '*"tree"~ha^-1*')'))) +
  theme(axis.title.y = element_text(size = 10)) +
  theme(axis.text = element_text(size = 9)) +
  theme(legend.position = "none") +
  annotate("text", x = c(1,2,3,4),
           y = 2800,
           label = c("a", "ab", "b", "b"),
           size = 4,
           color = "gray30")

# Fit: lmer(formula = tree_density ~ forest_classification + (1 | site), 
#     data = plots_cf_4p)
# 
# Linear Hypotheses:
#                               Estimate Std. Error z value Pr(>|z|)    
# degraded - deforested == 0       140.6      112.3   1.253 0.631100    
# regrowth - deforested == 0       253.9      153.6   1.653 0.393488    
# undisturbed - deforested == 0    440.8      113.0   3.902 0.000571 ***
# regrowth - degraded == 0         113.3      145.9   0.776 0.631100    
# undisturbed - degraded == 0      300.2      115.3   2.604 0.046046 *  
# undisturbed - regrowth == 0      186.9      155.9   1.199 0.631100    
# ---
# Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
# (Adjusted p values reported -- holm method)

# Undisturbed a
# Regrowth ab
# Degraded b
# Deforested b

## Panel b: lorey's height

pb <- plots_cf_4p %>%
  # mutate(succ = ifelse(str_class_plots == "1", "F",
  #                      ifelse(str_class_plots == "2", "Y", "M"))) %>%
  ggplot(aes(x= factor(forest_classification, levels = c('undisturbed', 'regrowth', 'degraded', 'deforested'),ordered = TRUE),
             y = loreys_height,
             fill = factor(forest_classification, levels = c('undisturbed', 'regrowth', 'degraded', 'deforested'),ordered = TRUE))) +
  geom_boxplot(color = "gray25",
               outlier.alpha = 0.5,
               outlier.color = "gray40",
               outlier.size = 1.2) +
  # stat_compare_means(comparisons = comp,
  #                    label = "p.signif") +
  theme_classic() +
  scale_fill_manual(values = c("#1b9e77", "#7570b3", "#e7298a", "#d95f02")) +
  theme(axis.title.x = element_blank()) +
  labs(y = "Lorey's height (m)") +
  theme(axis.title.y = element_text(size = 10)) +
  theme(axis.text = element_text(size = 9)) +
  theme(legend.position = "none") +
  annotate("text", x = c(1,2,3,4),
           y = 29,
           label = c("a", "ab", "ab", "b"),
           size = 4,
           color = "gray30")

# Fit: lmer(formula = loreys_height ~ forest_classification + (1 | site), 
#     data = plots_cf_4p)
# 
# Linear Hypotheses:
#                               Estimate Std. Error z value Pr(>|z|)   
# degraded - deforested == 0      2.8181     1.3363   2.109  0.17478   
# regrowth - deforested == 0      2.3610     1.6063   1.470  0.56645   
# undisturbed - deforested == 0   4.6171     1.4047   3.287  0.00608 **
# regrowth - degraded == 0       -0.4571     1.3745  -0.333  0.73947   
# undisturbed - degraded == 0     1.7990     1.2877   1.397  0.56645   
# undisturbed - regrowth == 0     2.2561     1.5772   1.430  0.56645   
# ---
# Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
# (Adjusted p values reported -- holm method)

# Undisturbed a
# Regrowth ab
# Degraded ab
# Deforested b

## Panel c: basal area

pc <- plots_cf_4p %>%
  # mutate(succ = ifelse(str_class_plots == "1", "F",
  #                      ifelse(str_class_plots == "2", "Y", "M"))) %>%
  ggplot(aes(x= factor(forest_classification, levels = c('undisturbed', 'regrowth', 'degraded', 'deforested'),ordered = TRUE),
             y = basal_area_ha,
             fill = factor(forest_classification, levels = c('undisturbed', 'regrowth', 'degraded', 'deforested'),ordered = TRUE))) +
  geom_boxplot(color = "gray25",
               outlier.alpha = 0.5,
               outlier.color = "gray40",
               outlier.size = 1.2) +
  # stat_compare_means(comparisons = comp,
  #                    label = "p.signif") +
  theme_classic() +
  scale_fill_manual(values = c("#1b9e77", "#7570b3", "#e7298a", "#d95f02")) +
  theme(axis.title.x = element_blank()) +
  labs(y = (bquote('Basal area ('*"m"^2 ~ha^-1*')'))) +
  theme(axis.title.y = element_text(size = 10)) +
  theme(axis.text = element_text(size = 9)) +
  theme(legend.position = "none") +
  annotate("text", x = c(1,2,3,4),
           y = 75,
           label = c("a", "ab", "ab", "b"),
           size = 4,
           color = "gray30")

#Fit: lmer(formula = basal_area_ha ~ forest_classification + (1 | site), 
#     data = plots_cf_4p)
# 
# Linear Hypotheses:
#                               Estimate Std. Error z value Pr(>|z|)    
# degraded - deforested == 0      10.068      3.944   2.552 0.053481 .  
# regrowth - deforested == 0       8.096      4.999   1.620 0.304732    
# undisturbed - deforested == 0   16.959      4.064   4.173 0.000181 ***
# regrowth - degraded == 0        -1.972      4.497  -0.438 0.661043    
# undisturbed - degraded == 0      6.890      3.924   1.756 0.304732    
# undisturbed - regrowth == 0      8.862      4.998   1.773 0.304732    
# ---
# Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
# (Adjusted p values reported -- holm method)

# Undisturbed a
# Regrowth ab
# Degraded ab
# Deforested b

## Panel d: wood density

pd <- trees_cf_4p %>%
  filter(!is.na(mean_wd)) %>%
  dplyr::group_by(plot_id, species) %>%
  dplyr::summarize(abundance = sum(n()),
                   mean_wd = as.numeric(mean_wd)) %>%
  group_by(plot_id) %>%
  summarise(mean_wd_plot = weighted.mean(mean_wd, abundance),
            sd_wd_plot = sd(mean_wd, na.rm = T)) %>%
  right_join(plots_cf_4p) %>%
  # mutate(succ = ifelse(str_class_plots == "1", "F",
  #                      ifelse(str_class_plots == "2", "Y", "M"))) %>%
  mutate(mean_wd_plot = ifelse(tree_no == 0, 0, mean_wd_plot),
         sd_wd_plot = ifelse(tree_no == 0, 0, sd_wd_plot)) %>%
  ggplot(aes(x= factor(forest_classification, levels = c('undisturbed', 'regrowth', 'degraded', 'deforested'),ordered = TRUE),
             y = mean_wd_plot,
             fill = factor(forest_classification, levels = c('undisturbed', 'regrowth', 'degraded', 'deforested'),ordered = TRUE))) +
  geom_boxplot(color = "gray25",
               outlier.alpha = 0.5,
               outlier.color = "gray40",
               outlier.size = 1.2) +
  # stat_compare_means(comparisons = comp,
  #                    label = "p.signif") +
  theme_classic() +
  scale_fill_manual(values = c("#1b9e77", "#7570b3", "#e7298a", "#d95f02")) +
  theme(axis.title.x = element_blank()) +
  labs(y = (bquote('Wood density ('*"g"~cm^-3*')'))) +
  theme(axis.title.y = element_text(size = 10)) +
  theme(axis.text = element_text(size = 9)) +
  theme(legend.position = "none") +
  annotate("text", x = c(1,2,3,4),
           y = 1,
           label = c("a", "a", "a", "a"),
           size = 4,
           color = "gray30")

# Fit: lmer(formula = mean_wd_plot ~ forest_classification + (1 | site), 
#     data = wd_df)
# 
# Linear Hypotheses:
#                               Estimate Std. Error z value Pr(>|z|)
# degraded - deforested == 0    0.020101   0.033074   0.608        1
# regrowth - deforested == 0    0.023434   0.042057   0.557        1
# undisturbed - deforested == 0 0.030875   0.034035   0.907        1
# regrowth - degraded == 0      0.003333   0.037927   0.088        1
# undisturbed - degraded == 0   0.010774   0.032920   0.327        1
# undisturbed - regrowth == 0   0.007441   0.042054   0.177        1
# (Adjusted p values reported -- holm method)

# Panel e: AGB

pe <- plots_cf_4p %>%
  # mutate(succ = ifelse(str_class_plots == "1", "F",
  #                      ifelse(str_class_plots == "2", "Y", "M"))) %>%
  ggplot(aes(x= factor(forest_classification, levels = c('undisturbed', 'regrowth', 'degraded', 'deforested'),ordered = TRUE),
             y = agb_plot_ha,
             fill = factor(forest_classification, levels = c('undisturbed', 'regrowth', 'degraded', 'deforested'),ordered = TRUE))) +
  geom_boxplot(color = "gray25",
               outlier.alpha = 0.5,
               outlier.color = "gray40",
               outlier.size = 1.2) +
  # stat_compare_means(comparisons = comp,
  #                    label = "p.signif") +
  theme_classic() +
  scale_fill_manual(values = c("#1b9e77", "#7570b3", "#e7298a", "#d95f02")) +
  theme(axis.title.x = element_blank()) +
  labs(y = (bquote('AGB ('*"Mg"~ha^-1*')'))) +
  theme(axis.title.y = element_text(size = 10)) +
  theme(axis.text = element_text(size = 9)) +
  theme(legend.position = "none") +
  annotate("text", x = c(1,2,3,4),
           y = 750,
           label = c("a", "ab", "ab", "b"),
           size = 4,
           color = "gray30")

# Fit: lmer(formula = agb_plot_ha ~ forest_classification + (1 | site), 
#     data = plots_cf_4p)
# 
# Linear Hypotheses:
#                               Estimate Std. Error z value Pr(>|z|)  
# degraded - deforested == 0       88.58      37.81   2.343   0.0958 .
# regrowth - deforested == 0       57.54      47.22   1.218   0.7055  
# undisturbed - deforested == 0   121.09      39.17   3.092   0.0119 *
# regrowth - degraded == 0        -31.04      41.94  -0.740   0.7673  
# undisturbed - degraded == 0      32.52      37.32   0.871   0.7673  
# undisturbed - regrowth == 0      63.56      47.01   1.352   0.7055  
# ---
# Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
# (Adjusted p values reported -- holm method)

# Undisturbed a
# Regrowth ab
# Degraded ab
# Deforested b

## Panel f: Hill-Shannon

pf <- plots_cf_4p %>%
  ggplot(aes(x= factor(forest_classification, levels = c('undisturbed', 'regrowth', 'degraded', 'deforested'),ordered = TRUE),
             y = hill_shannon,
             fill = factor(forest_classification, levels = c('undisturbed', 'regrowth', 'degraded', 'deforested'),ordered = TRUE))) +
  geom_boxplot(color = "gray25",
               outlier.alpha = 0.5,
               outlier.color = "gray40",
               outlier.size = 1.2) +
  theme_classic() +
  scale_fill_manual(values = c("#1b9e77", "#7570b3", "#e7298a", "#d95f02")) +
  theme(axis.title.x = element_blank()) +
  labs(y = (bquote('Hill-Shannon'))) +
  theme(axis.title.y = element_text(size = 10)) +
  theme(axis.text = element_text(size = 9)) +
  theme(legend.position = "none") +
  scale_y_break(c(5.5,9), ticklabels = c(9,10), space = 0.3) +
  annotate("text", x = c(1,2,3,4),
           y = 10,
           label = c("a", "a", "a", "a"),
           size = 4,
           color = "gray30")



# pe2 <- pe + theme(legend.position = "none")
# #pg2 <- pg + theme(legend.position = "none")
# l <- get_legend(pf)

figure2 <- ggarrange(print(pa), print(pb), print(pc), print(pd), print(pe), print(pf),
                  ncol = 3,
                  nrow = 2,
                  labels = c("a)", "b)", "c)", "d)", "e)", "f)"),
                  font.label = list(size= 10),
                  common.legend = F)


                  # common.legend = T,
                  # legend.grob = l,
                  # legend= "right")
figure2

```


## Figure 4. Structural attributes and tree size contribution to AGB and stem density at plot level

```{r}

# Tree size contribution to stem density

## Data frame for plot 1

p1_df <- plots_cf_4p %>%
  select(plot_id, tree_no, forest_classification) %>%
  right_join(stems_class_plot, by = "plot_id") %>%
  mutate(stem_prop = stems/tree_no)  %>%
  # mutate(succ = ifelse(str_class_plots == "1", "F",
  #                      ifelse(str_class_plots == "2", "Y", "M"))) %>%
  mutate(size_class = ifelse(diam_class == "1", "<10",
                             ifelse(diam_class == "2", "10-20",
                                    ifelse(diam_class == "3", "20-30",
                                           ifelse(diam_class == "4", "30-40",
                                                  ifelse(diam_class == "5", "40-50", ">50"))))))

## Plot

p1 <- p1_df %>%
  group_by(forest_classification, diam_class) %>%
  # plot
  ggplot(aes(x = factor(size_class, 
                        levels =c("<10", "10-20", "20-30", "30-40", "40-50", ">50"),
                        ordered = TRUE),
             y = stem_prop)) +
  geom_boxplot(aes(fill = factor(forest_classification, levels = c('undisturbed', 'regrowth', 'degraded', 'deforested'),ordered = TRUE)),
             color = "gray25",
               outlier.alpha = 0.5,
               outlier.color = "gray40",
               outlier.size = 1.5) +
  scale_color_manual(values = c("#1b9e77", "#7570b3", "#e7298a", "#d95f02")) +
  scale_fill_manual(values = c("#1b9e77", "#7570b3", "#e7298a", "#d95f02")) +
  scale_y_continuous(breaks = c(0.0, 0.25, 0.50, 0.75, 1.0)) +
  theme_classic() +
  labs(x = "Size class",
       y = "Proportion of stems") +
  theme(axis.title.x = element_text(size = 11)) +
  theme(axis.title.y = element_text(size = 10)) +
  theme(axis.text = element_text(size = 9)) +
  theme(legend.position = c(0.9,0.7)) +
  theme(legend.title = element_blank()) +
  theme(plot.margin = margin(b= 10, t= 9)) +
  annotate("text", x = c(1,2,3,5,6),
           y = 1,
           label = c("NS","NS", "NS", "NS", "NS"),
           size = 4,
           color = "gray10") +
  annotate("text", x = 4,
           y = 1,
           label = "*",
           size = 6,
           color = "gray10")


# Stats:
# Diam 4. undisturbed - deforested == 0  -0.8813     0.2980  -2.958   0.0152 *


# Tree size contribution to AGB

## Data frame for plot 2

p2_df <- plots_cf_4p %>%
  select(plot_id, agb_plot, forest_classification) %>%
  right_join(agb_class_plots, by = "plot_id") %>%
  mutate(agb_prop = agb_class_plot/agb_plot)  %>%
  mutate(size_class = ifelse(diam_class == "1", "<10",
                             ifelse(diam_class == "2", "10-20",
                                    ifelse(diam_class == "3", "20-30",
                                           ifelse(diam_class == "4", "30-40",
                                                  ifelse(diam_class == "5", "40-50", ">50"))))))

## Stats


## Plot
p2 <-  p2_df %>%
  group_by(forest_classification, diam_class) %>%
  ggplot(aes(x = factor(size_class, 
                        levels =c("<10", "10-20", "20-30", "30-40", "40-50", ">50"),
                        ordered = TRUE),
             y = agb_prop)) +
  geom_boxplot(aes(fill = factor(forest_classification, levels = c('undisturbed', 'regrowth', 'degraded', 'deforested'),ordered = TRUE)),
             color = "gray25",
               outlier.alpha = 0.5,
               outlier.color = "gray40",
               outlier.size = 1.5) +
  scale_color_manual(values = c("#1b9e77", "#7570b3", "#e7298a", "#d95f02")) +
  scale_fill_manual(values = c("#1b9e77", "#7570b3", "#e7298a", "#d95f02")) +
  theme_classic() +
  scale_y_continuous(breaks = c(0.0, 0.25, 0.50, 0.75, 1.0)) +
  labs(x = "Size class",
       y = "Proportion of AGB") +
  theme(axis.title.x = element_text(size = 11)) +
  theme(axis.title.y = element_text(size = 10)) +
  theme(axis.text = element_text(size = 9)) +
  theme(legend.position = "none") +
  theme(plot.margin = margin(b= 10, t= 9)) +
  annotate("text", x = c(1,2,3,5,6),
           y = 1,
           label = c("NS","NS", "NS", "NS", "NS"),
           size = 4,
           color = "gray10") +
  annotate("text", x = 4,
           y = 1,
           label = "*",
           size = 6,
           color = "gray10")

# Stats:
# Diam4. undisturbed - deforested == 0  -1.1575     0.3308  -3.499  0.00244 **

p1
p2  
   
figure3 <- plot_grid(p1, p2,
                     nrow = 2,
                     ncol = 1,
                     heights = c(1, 1),
                     label_size = 10, labels = c("a)", "b)"))
figure3

####################################################################################################

## Make figure with tukey test groups

# AGB

anova3 <- aov(agb_prop ~ diam_class*succ, data = agb_prop_str)
summary(anova3)
t <- TukeyHSD(anova3)
multcompLetters4(anova3, TukeyHSD(anova3), reversed = T)

#### Groupings resulted from tukey test

c1 <- rep(c(1, 2, 3), each = 6)
c2 <- rep(seq(1, 6), times = 3)
c3 <- list(c("ab", "ef", "def", "efg", "g", "g",
       "a", "cde", "cde", "bcd", "bcde", "defg",
       "a", "ab", "ab", "abc", "bcde", "fg"))

groupings_aov <- data.frame(c1, c2, c3) 
colnames(groupings_aov) <- c("str_class_plots", "diam_class", "groupings")

groupings_aov <- groupings_aov %>%
  as_tibble() %>%
  arrange(diam_class) %>%
  mutate(succ = ifelse(str_class_plots == "1", "F",
                       ifelse(str_class_plots == "2", "Y", "M"))) %>%
  mutate(size_class = ifelse(diam_class == "1", "<10",
                             ifelse(diam_class == "2", "10-20",
                                    ifelse(diam_class == "3", "20-30",
                                           ifelse(diam_class == "4", "30-40",
                                                  ifelse(diam_class == "5", "40-50", ">50"))))))

groupings_aov <- groupings_aov %>%
  mutate(x_groupings = c(0.7333333, 1.000000, 1.266667, 1.7333333, 2.000000, 2.266667,
               2.7333333, 3.000000, 3.266667, 3.7333333, 4.000000, 4.266667,
               4.7333333, 5.000000, 5.266667, 5.7333333, 6.000000, 6.266667))

groupings_aov <- groupings_aov %>%
  as_tibble() %>%
  mutate(str_class_plots = as.character(str_class_plots),
         diam_class = as.character(diam_class))

p2_df <- p2_df %>%
  select(-groupings, -x_groupings)

p2b_df <- p2_df %>%
  left_join(groupings_aov, by = c("str_class_plots", "diam_class", "succ", "size_class"))

p2b <-  p2b_df %>%
  group_by(str_class_plots, diam_class) %>%
  ggplot(aes(x = factor(size_class, 
                        levels =c("<10", "10-20", "20-30", "30-40", "40-50", ">50"),
                        ordered = TRUE),
             y = agb_prop)) +
  geom_boxplot(aes(fill = factor(succ,
                           levels = c("F","Y", "M"),
                           ordered = TRUE)),
             color = "gray25",
               outlier.alpha = 0.5,
               outlier.color = "gray40",
               outlier.size = 1.5) +
  # stat_pvalue_manual(stat_test_p2,
  #                    tip.length = 0.01,
  #                    hide.ns = T,
  #                    bracket.size = 0.5,
  #                    step.increase = 0.001,
  #                    bracket.nudge.y = -0.015,
  #                    bracket.shorten = -0.05) +
  geom_text(aes(x= x_groupings,
                y= -0.05,
                label = groupings),
                # color = factor(succ,
                #            levels = c("F","Y", "M"),
                #            ordered = TRUE)),
                color = "gray40",
            size = 3.5) +
  scale_color_manual(values = c("#d95f02", "#7570b3", "#1b9e77")) +
  scale_fill_manual(values = c("#d95f02", "#7570b3", "#1b9e77")) +
  theme_classic() +
  scale_y_continuous(breaks = c(0.0, 0.25, 0.50, 0.75, 1.0)) +
  labs(x = "Size class",
       y = "Proportion of AGB") +
  theme(axis.title.x = element_text(size = 10)) +
  theme(axis.title.y = element_text(size = 9)) +
  theme(axis.text = element_text(size = 8)) +
  theme(legend.position = "none") +
  theme(plot.margin = margin(b= 10, t= 9))
p2b

# Stem density

anova4 <- aov(stems_prop ~ diam_class*succ, data = stem_prop_str)
summary(anova4)
TukeyHSD(anova4)
multcompLetters4(anova4, TukeyHSD(anova4))

t_stems <- tukey_hsd(anova4)

#### Groupings resulted from tukey test

c1 <- rep(c(1, 2, 3), each = 6)
c2 <- rep(seq(1, 6), times = 3)
c3 <- list(c("a", "a", "bc", "bcd", "bcdef", "bcdef",
       "b", "a", "cde", "f", "f", "ef",
       "bcd", "a", "cdef", "cdef", "def", "def"))

groupings_aov <- data.frame(c1, c2, c3) 
colnames(groupings_aov) <- c("str_class_plots", "diam_class", "groupings")

groupings_aov <- groupings_aov %>%
  as_tibble() %>%
  arrange(diam_class) %>%
  mutate(succ = ifelse(str_class_plots == "1", "F",
                       ifelse(str_class_plots == "2", "Y", "M"))) %>%
  mutate(size_class = ifelse(diam_class == "1", "<10",
                             ifelse(diam_class == "2", "10-20",
                                    ifelse(diam_class == "3", "20-30",
                                           ifelse(diam_class == "4", "30-40",
                                                  ifelse(diam_class == "5", "40-50", ">50"))))))

groupings_aov <- groupings_aov %>%
  mutate(x_groupings = c(0.7333333, 1.000000, 1.266667, 1.7333333, 2.000000, 2.266667,
               2.7333333, 3.000000, 3.266667, 3.7333333, 4.000000, 4.266667,
               4.7333333, 5.000000, 5.266667, 5.7333333, 6.000000, 6.266667))

groupings_aov <- groupings_aov %>%
  as_tibble() %>%
  mutate(str_class_plots = as.character(str_class_plots),
         diam_class = as.character(diam_class))

p1_df <- p1_df %>%
  select(-groupings, -x_groupings)

p1b_df <- p1_df %>%
  left_join(groupings_aov, by = c("str_class_plots", "diam_class", "succ", "size_class"))

p1b <-  p1b_df %>%
  group_by(str_class_plots, diam_class) %>%
  ggplot(aes(x = factor(size_class, 
                        levels =c("<10", "10-20", "20-30", "30-40", "40-50", ">50"),
                        ordered = TRUE),
             y = stem_prop)) +
  geom_boxplot(aes(fill = factor(succ,
                           levels = c("F","Y", "M"),
                           ordered = TRUE)),
             color = "gray25",
               outlier.alpha = 0.5,
               outlier.color = "gray40",
               outlier.size = 1.5) +
  # stat_pvalue_manual(stat_test_p2,
  #                    tip.length = 0.01,
  #                    hide.ns = T,
  #                    bracket.size = 0.5,
  #                    step.increase = 0.001,
  #                    bracket.nudge.y = -0.015,
  #                    bracket.shorten = -0.05) +
  geom_text(aes(x= x_groupings,
                y= -0.05,
                label = groupings),
                # color = factor(succ,
                #            levels = c("F","Y", "M"),
                #            ordered = TRUE)),
                color = "gray40",
            size = 3.5) +
  scale_color_manual(values = c("#d95f02", "#7570b3", "#1b9e77")) +
  scale_fill_manual(values = c("#d95f02", "#7570b3", "#1b9e77")) +
  theme_classic() +
  scale_y_continuous(breaks = c(0.0, 0.25, 0.50, 0.75, 1.0)) +
  labs(x = "Size class",
       y = "Proportion of stems") +
  theme(axis.title.x = element_text(size = 10)) +
  theme(axis.title.y = element_text(size = 9)) +
  theme(axis.text = element_text(size = 8)) +
  theme(legend.position = c(0.9,0.8)) +
  theme(legend.title = element_blank()) +
  theme(plot.margin = margin(b= 10, t= 9))

p1b

figure3b <- plot_grid(p1b, p2b,
                     nrow = 2,
                     ncol = 1,
                     heights = c(1, 1),
                     label_size = 8, labels = c("a)", "b)"))
figure3b

anova3
summary(anova3)
t <- tukey_hsd(anova3)

```

## Figure 3. 

```{r}

sp_plot <- trees_cf_4p %>% # make a table of community data for estimating diversity
  filter(life_form == "Arbol" |
           is.na(life_form)) %>%
  #separate(plot_id, c("site", "plot"), sep = "_", remove = F) %>%
  dplyr::group_by(plot_id, species) %>%
  dplyr::summarize(abundance = sum(n())) %>%
  filter(!is.na(species)) %>%
  spread(species, abundance, fill = 0) %>%
  column_to_rownames(var = "plot_id")

sp_plot2 <- trees_cf_4p %>% # make a table of community data for estimating diversity
  filter(life_form == "Arbol" |
           is.na(life_form)) %>%
  #separate(plot_id, c("site", "plot"), sep = "_", remove = F) %>%
  dplyr::group_by(plot_id, species) %>%
  dplyr::summarize(abundance = sum(n())) %>%
  filter(!is.na(species))

fig4_df <- trees_cf_4p %>%
  filter(!is.na(mean_wd)) %>%
  dplyr::group_by(plot_id, species) %>%
  dplyr::summarize(abundance = sum(n()),
                   mean_wd = as.numeric(mean_wd),
                   height = as.numeric(height)) %>%
  right_join(plots_cf_4p) %>%
  dplyr::group_by(plot_id) %>%
  dplyr::summarize(ave_wd = weighted.mean(mean_wd, abundance),
                   sd_wd = sd(mean_wd, na.rm = T),
                   se_wd = calcSE(mean_wd),
                   max_height = max(height),
                   mean_height = mean(height),
                   sd_height = sd(height)) %>%
  right_join(plots_cf_4p) %>%
  dplyr::select(altitude, tree_density, basal_area_ha, loreys_height, mean_height, max_height, ave_wd, se_wd, agb_plot_ha, agb_se_ha, forest_classification) 

f4a <- fig4_df %>%
  ggplot(aes(x = tree_density, y = agb_plot_ha))  +
  geom_point(aes(color = factor(forest_classification, levels = c('undisturbed', 'regrowth', 'degraded', 'deforested'),ordered = TRUE)),
             size = 2, alpha = 0.75) +
  scale_color_manual(values = c("#1b9e77", "#7570b3", "#e7298a", "#d95f02")) +
  geom_smooth(method = "lm", span= 0.9, color = "gray20", se = F, linetype = "solid", size = 0.75, alpha= 0.3) +
  scale_x_continuous(limits = c(0, 2000)) +
  theme_classic() +
  theme(legend.position = "none") +
  labs(x = (bquote('Tree density ('*"tree"~ha^-1*')')),
       y = (bquote('log AGB ('*"Mg "~ha^-1*')'))) +
  theme(legend.position = "none") +
  scale_y_continuous(trans = "log1p",
                     limits = c(0,1000),
                     breaks = c(0,100,500,1000)) +
  annotate("text", x = 400, y = 1000, label = "Adjusted R-squared:  0.332", size = 3.3)
  
summary(lm(log1p(fig4_df$agb_plot_ha) ~ fig4_df$tree_density))

f4b <- fig4_df %>%
  ggplot(aes(x = basal_area_ha, y = agb_plot_ha))  +
  geom_point(aes(color = factor(forest_classification, levels = c('undisturbed', 'regrowth', 'degraded', 'deforested'),ordered = TRUE)),
             size = 2, alpha = 0.75) +
  scale_color_manual(values = c("#1b9e77", "#7570b3", "#e7298a", "#d95f02")) +
  geom_smooth(method = "lm", span=1, color = "gray20", se = F, linetype = "solid", size = 0.75, alpha= 0.3) +
  theme_classic() +
  labs(x = (bquote('Basal area ('*"m"^2 ~ha^-1*')')),
       y = (bquote('log AGB ('*"Mg "~ha^-1*')'))) +
  theme(legend.title = element_blank(),
      legend.position = c(0.9, 0.2),
      legend.background = element_rect(linetype = "solid", size = 0.01, color = "gray50"),
      legend.margin = margin(0.1,5,0.1,0.1)) +
  scale_y_continuous(trans = "log1p",
                     limits = c(0,1000),
                     breaks = c(0,100,500,1000)) +
  annotate("text", x = 20, y = 1000, label = "Adjusted R-squared:  0.721", size = 3.3)

summary(lm(log1p(fig4_df$agb_plot_ha) ~ fig4_df$basal_area_ha))

f4c <- fig4_df %>%
  ggplot(aes(x = loreys_height, y = agb_plot_ha))  +
  geom_point(aes(color = factor(forest_classification, levels = c('undisturbed', 'regrowth', 'degraded', 'deforested'),ordered = TRUE)),
             size = 2, alpha = 0.75) +
  scale_color_manual(values = c("#1b9e77", "#7570b3", "#e7298a", "#d95f02")) +
  geom_smooth(method = "lm", span= 0.9, color = "black", se = F, linetype = "solid", size = 0.75, alpha= 0.3) +
  theme_classic() +
  theme(legend.position = "none") +
  labs(x = "Lorey's height (m)",
       y = (bquote('log AGB ('*"Mg "~ha^-1*')'))) +
  scale_y_continuous(trans = "log1p",
                     limits = c(0,1000),
                     breaks = c(0,100,500,1000)) +
  annotate("text", x = 5, y = 1000, label = "Adjusted R-squared:  0.525", size = 3.3)

summary(lm(log1p(fig4_df$agb_plot_ha) ~ fig4_df$loreys_height))

f4f <- fig4_df %>%
  ggplot(aes(x = ave_wd, y = agb_plot_ha))  +
  geom_point(aes(color = factor(forest_classification, levels = c('undisturbed', 'regrowth', 'degraded', 'deforested'),ordered = TRUE)),
             size = 2, alpha = 0.75) +
  scale_color_manual(values = c("#1b9e77", "#7570b3", "#e7298a", "#d95f02")) +
  geom_smooth(method = "lm", span=1, color = "gray20", se = F, linetype = "solid", size = 0.75, alpha= 0.3) +
  theme_classic() +
  theme(legend.position = "none") +
  labs(x = (bquote('Wood density ('*"g"~cm^-3*')')),
       y = (bquote('log AGB ('*"Mg "~ha^-1*')'))) +
  scale_y_continuous(trans = "log1p",
                     limits = c(0,1000),
                     breaks = c(0,100,500,1000)) +
  annotate("text", x = 0.35, y = 1000, label = "Adjusted R-squared:  0.17", size = 3.3)

summary(lm(log1p(fig4_df$agb_plot_ha) ~ fig4_df$ave_wd))

# figure 4

figure4 <- ggarrange(f4a, f4b, f4c, f4f,
                  ncol = 2,
                  nrow = 2,
                  labels = c("a)", "b)", "c)", "d)"),
                  font.label = list(size= 10),
                  common.legend = F)

figure4
```



## Figure . Relationship between AGB and landscape composition, and diversity and landscape composition

```{r}

# AGB ~ landscape composition

p1 <- sites_cf_4p %>%
  ggplot(aes(x = landscape, y = av_agb_site)) +
  geom_errorbar(aes(ymax = av_agb_site + se_agb_site,
                    ymin = av_agb_site - se_agb_site),
                width = 0.02,
                color = "gray30",
                alpha = 0.8) +
  geom_point(color = "gray30", size = 3, alpha = 0.8) +
  geom_smooth(method = "lm", color = "black", se = F, linetype = "solid", size = 0.75) +
  # geom_smooth(method = "loess", span=1, color = "black", se = F, linetype = "dashed", size = 0.75) +
  theme_classic() +
  labs(x = "Land-use intensity",
       y = (bquote('AGB ('*"Mg"~ha^-1*')')))
p1

# Tree diversity ~ landscape composition

p2 <- sites_cf_4p %>%
  ggplot(aes(x = landscape, y = shannon)) +
  # geom_errorbar(aes(ymax = shannon_av_site + shannon_se_site,
  #                   ymin = shannon_av_site - shannon_se_site),
  #               width = 0.02,
  #               color = "gray60",
  #               alpha = 0.8) +
  geom_smooth(method = "lm", color = "black", se = F, linetype = "dashed", size = 0.75) +
  # geom_smooth(method = "loess", span=1, color = "black", se = F, linetype = "dashed", size = 0.75) +
  geom_point(color = "gray30", size = 3, alpha = 0.8) +
  theme_classic() +
  labs(x = "Land-use intensity",
       y = "H")

# p3 <- sites_cf_4p %>%
#   ggplot(aes(x = landscape, y = sp_richness_site_av)) +
#   geom_errorbar(aes(ymax = sp_richness_site_av + sp_richness_site_se,
#                     ymin = sp_richness_site_av - sp_richness_site_se),
#                 width = 0.02,
#                 color = "gray60",
#                 alpha = 0.8) +
#   geom_smooth(method = "lm", color = "black", se = F, linetype = "solid", size = 0.75) +
#   geom_smooth(method = "loess", span=1, color = "gray40", se = F, linetype = "dashed") +
#   geom_point(color = "gray60", size = 3, alpha = 0.8) +
#   theme_classic() +
#   labs(x = "Landscape composition",
#        y = "S")
# 
# p2
# p3

figure4 <- plot_grid(p1, p2,
          labels = c("a)", "b)"),
          label_size = 10)

figure4

summary(lm(av_agb_site ~ landscape, data = sites_cf_4p))
```

## Figure .  Regressions between AGB and env_pc1 and land-use intensity and env_pc1

```{r}

# AGB ~ env_pc1

p1 <- ggplot(sites_cf_4p, aes(x = env_pc1, y = av_agb_site)) +
  geom_smooth(alpha= 0.1, method = "lm", color = "darkgray", linetype = "solid") +
  geom_point(aes(color = landscape), size = 4) +
  scale_color_gradient(low = "forestgreen", high =  "lightgoldenrod2") +
  theme_classic() +
  labs(x = "Environmental gradient",
       y = (bquote('AGB ('*"Mg"~ha^-1*')')),
       color = "Land-use \nintensity   ") +
  theme(legend.position='top', 
        legend.justification='right',
        legend.direction='horizontal',
        legend.key.height = unit(0.25, "cm"),
        legend.key.width = unit(0.75, "cm"),
        legend.text = element_text(size= 8),
        legend.title = element_text(size = 9),
        legend.title.align = 0)


# Landscape composition ~ env_pc1

p2 <- ggplot(sites_cf_4p, aes(x = env_pc1, y = landscape)) +
  geom_smooth(alpha= 0.1, method = "lm", color = "darkgray", linetype = "solid") +
  geom_point(aes(color = av_agb_site), size = 4) +
  scale_color_gradient(high = "darkgreen", low = "darkseagreen1") +
  theme_classic() +
  labs(x = "Environmental gradient",
       y = "Land-use intensity",
       color = (bquote('AGB ('*"Mg"~ha^-1*')'))) +
  scale_y_continuous(limits = c(0.0, 1.0), breaks = c(0.0, 0.25, 0.50, 0.75, 1.00)) +
  theme(legend.position='top', 
        legend.justification='right',
        legend.direction='horizontal',
        legend.key.height = unit(0.25, "cm"),
        legend.key.width = unit(0.75, "cm"),
        legend.text = element_text(size= 8),
        legend.title = element_text(size = 9),
        legend.title.align = 0)

figure5 <- plot_grid(p1, p2,
          labels = c("a)", "b)"),
          label_size = 10,
          hjust = -1)

figure5
```

# Figure . Loess curves of AGB and tree diversity at site and plot levels

```{r}

# Sites

## Shannon

p1 <- sites_cf_4p %>%
  ggplot(aes(y = av_agb_site, x = hill_shannon)) +
  geom_smooth(alpha = 0.1, method = "gam", formula = y ~ t2(x, bs = "tp"), span = 1, color= "black", linetype = "dashed", se = T) +
  geom_point(color = "gray30", size=2) +
  theme_classic() +
  labs(y = (bquote('log AGB ('*"Mg"~ha^-1*')')),
       x = "Hill-Shannon") +
  scale_y_continuous(trans = "log1p") +
  scale_x_continuous(trans = "sqrt") +
  geom_point(aes(x = 7, y = 1.5), color = "gray30", size = 2) +
  annotate("text", x = 7.5, y = 1.5, label = "sites", size = 3.3)

#method = "gam", formula = y ~ t2(x, bs = "tp")
p1_line <- gam(log1p(av_agb_site) ~ t2(hill_shannon, bs= "tp", k= 4), data = sites_cf_4p)
summary(p1_line)

# Plots

## Shannon 

p3 <- plots_cf_4p %>%
  #filter(agb_plot_ha < 600) %>%
  #filter(hill_shannon <7) %>%
  filter(!is.na(hill_shannon)) %>%
  filter(!is.na(agb_plot_ha)) %>%
  # group_by(str_class_plots) %>%
  # summarise(mean_shannon = mean(shannon),
  #           mean_agb = mean(agb_plot_ha),
  #           agb_plot_ha = as.numeric(agb_plot_ha),
  #           shannon = as.numeric(shannon)) %>%
  ggplot(aes(y = agb_plot_ha, x = hill_shannon)) +
  #geom_point(aes(x= mean_agb, y = mean_shannon, color = factor(succ, levels = c("F","Y", "M"),ordered = TRUE), size = 3)) +
  geom_point(aes(color = factor(forest_classification, levels = c('undisturbed', 'regrowth', 'degraded', 'deforested'),ordered = TRUE)), size= 2, alpha = 0.75) +
  geom_smooth(alpha = 0.15, method = "gam", formula = y ~ s(x, bs = "tp"), span = 1, color= "black", linetype = "dashed", se= T) +
  theme_classic() +
  scale_color_manual(values = c("#1b9e77", "#7570b3", "#e7298a", "#d95f02")) +
  labs(y = (bquote('log AGB ('*"Mg"~ha^-1*')')),
       x = "Hill-Shannon") +
  theme(legend.position = c(0.8,0.13)) +
  theme(legend.title = element_blank()) +
  scale_y_continuous(trans = "log1p") +
  scale_x_continuous(trans = "sqrt")

# p3_line <- gam(log1p(agb_plot_ha) ~ s(hill_shannon, bs= "tp", k= 4), data = plots_cf_4p)
# summary(p3_line)
# p3_line_re <- gamm4(log1p(agb_plot_ha) ~ s(hill_shannon, bs= "tp", k= 4), random = ~(1|site), data = plots_cf_4p_wo)
# summary(p3_line_re$gam)
# summary(p3_line_re$mer)
#None have adj-R2 > 0
#The only significant relationship is AGB~H without the outlier (it's linear)

figure6 <- ggarrange(p3,p1,
                  ncol = 2,
                  nrow = 1,
                  labels = c("a)", "b)"),
                  font.label = list(size= 10))
figure6
```



```{r}

# Sites

## Shannon

p1 <- sites_cf_4p %>%
  ggplot(aes(y = log1p(av_agb_site), x = shannon)) +
  geom_smooth(method = "lm", span = 1, color= "black", linetype = "dashed", se= F) +
  geom_point(color = "gray30", size=2) +
  theme_classic() +
  labs(y = (bquote('log AGB ('*"Mg"~ha^-1*')')),
       x = "H")

## Sp richness

p2 <- sites_cf_4p %>%
  ggplot(aes(y = log1p(av_agb_site), x = sp_richness_site)) +
  geom_smooth(method = "lm", span = 1, color= "black", linetype = "dashed", se = F) +
  geom_point(color = "gray30", size=2) +
  scale_x_continuous(breaks = c(0, 2, 4, 6, 8, 10, 12, 14, 16, 18)) +
  theme_classic() +
  labs(y = (bquote('log AGB ('*"Mg"~ha^-1*')')),
       x = "S") +
  geom_point(aes(x = 16.3, y = 1.5), color = "gray30", size = 2) +
  geom_text(x = 17.5, y = 1.5, label = "Sites", size = 3.3)

# Plots

## Shannon 

p3 <- plots_cf_4p %>%
  filter(agb_plot_ha < 600) %>%
  filter(!is.na(shannon)) %>%
  filter(!is.na(agb_plot_ha)) %>%
  group_by(str_class_plots) %>%
  summarise(mean_shannon = mean(shannon),
            mean_agb = mean(agb_plot_ha),
            agb_plot_ha = as.numeric(agb_plot_ha),
            shannon = as.numeric(shannon)) %>%
  mutate(succ = ifelse(str_class_plots == "1", "F",
                       ifelse(str_class_plots == "2", "Y", "M"))) %>%
  ggplot(aes(y = log1p(agb_plot_ha), x = shannon)) +
  #geom_point(aes(x= mean_agb, y = mean_shannon, color = factor(succ, levels = c("F","Y", "M"),ordered = TRUE), size = 3)) +
  geom_point(aes(color = factor(succ, levels = c("F","Y", "M"),ordered = TRUE)), size= 2, alpha = 0.75) +
  geom_smooth(method = "lm", span = 1, color= "black", linetype = "solid", se = F) +
  theme_classic() +
  scale_color_manual(values = c("#d95f02", "#7570b3", "#1b9e77")) +
  labs(y = (bquote('log AGB ('*"Mg"~ha^-1*')')),
       x = "H") +
  theme(legend.position = "none")

## Species richness

p4 <- plots_cf_4p %>%
  filter(agb_plot_ha < 600) %>%
  filter(!is.na(simpson)) %>%
  filter(!is.na(agb_plot_ha)) %>%
  group_by(str_class_plots) %>%
  summarise(mean_sp = mean(sp_richness_plot),
            mean_agb = mean(agb_plot_ha),
            agb_plot_ha = as.numeric(agb_plot_ha),
            sp_richness_plot = as.numeric(sp_richness_plot)) %>%
  mutate(succ = ifelse(str_class_plots == "1", "F",
                       ifelse(str_class_plots == "2", "Y", "M"))) %>%
  ggplot(aes(y = log1p(agb_plot_ha),
             x = sp_richness_plot,
             color = factor(succ, levels = c("F","Y", "M"), ordered = TRUE))) +
  geom_point(aes(color = factor(succ, levels = c("F","Y", "M"), ordered = TRUE)), size= 2, alpha = 0.75) +
  geom_smooth(method = "lm", span = 1, color= "black", linetype = "solid", se = F) +
  theme_classic() +
  scale_color_manual(values = c("#d95f02", "#7570b3", "#1b9e77")) +
  scale_x_continuous(breaks = c(0, 2, 4, 6, 8, 10, 12, 14)) +
  labs(size = "Average H in successional stage",
       y = (bquote('log AGB ('*"Mg"~ha^-1*')')),
       x = "S") +
  theme(legend.title = element_blank(),
        #legend.direction = "horizontal",
        legend.position = c(0.9,0.2),
        legend.background = element_rect(linetype = "solid", size = 0.01, color = "gray50"),
        legend.key.size = unit(0.25, "cm"))

figure6_2 <- ggarrange(p3,p4,p1,p2,
                  ncol = 2,
                  nrow = 2,
                  labels = c("a)", "b)", "c)", "d)"),
                  font.label = list(size= 10))
figure6_2

summary(lm(sites_cf_4p$av_agb_site ~ sites_cf_4p$shannon))
summary(lm(log1p(sites_cf_4p$av_agb_site) ~ sites_cf_4p$shannon))

summary(lm(sites_cf_4p$av_agb_site ~ sites_cf_4p$sp_richness_site))
summary(lm(log1p(sites_cf_4p$av_agb_site) ~ sites_cf_4p$sp_richness_site))

summary(lm(plots_cf_4p$agb_plot_ha ~ plots_cf_4p$shannon))
summary(lm(log1p(plots_cf_4p$agb_plot_ha) ~ plots_cf_4p$shannon))

summary(lm(plots_cf_4p$agb_plot_ha ~ plots_cf_4p$sp_richness_plot))
summary(lm(log1p(plots_cf_4p$agb_plot_ha) ~ plots_cf_4p$sp_richness_plot))
```


# Extra figure: is the relationship between sp diversity and AGB scale dependent?

```{r}

plots_cf_4p %>%
  #filter(agb_plot_ha < 600) %>%
  ggplot(aes(x = hill_shannon, y = log1p(agb_plot_ha))) +
  geom_point(aes(color = forest_classification)) +
  geom_smooth(method = "lm")

plots_cf_4p %>%
  filter(!str_class_plots == "1") %>%
  ggplot(aes(x = shannon, y = log1p(agb_plot_ha))) +
  geom_point(aes(color = str_class_plots)) +
  geom_smooth(method = "lm")


sites_cf_4p %>%
  ggplot(aes(x = shannon, y = log1p(av_agb_site))) +
  geom_point() +
  geom_smooth(method = "lm")
  
```

# Another extra figure: relationship between all structural attributes and env gradient

```{r}

ggplot(plots_cf_4p, aes(x = env_pc1, y = tree_density)) +
  geom_smooth(alpha= 0.1, method = "loess", color = "darkgray", linetype = "dashed") +
  geom_point(aes(color = factor(forest_classification, levels = c('undisturbed', 'regrowth', 'degraded', 'deforested'),ordered = TRUE)), size = 2) +
  scale_color_manual(values = c("#1b9e77", "#7570b3", "#e7298a", "#d95f02")) +
  labs(y = "Stem density") +
  theme_classic() +
  theme(legend.title = element_blank(),
      legend.position = c(0.9, 0.9),
      legend.background = element_rect(linetype = "solid", size = 0.01, color = "gray50"),
      legend.margin = margin(0.1,5,0.1,0.1))

ggplot(plots_cf_4p, aes(x = env_pc1, y = loreys_height)) +
  geom_smooth(alpha= 0.1, method = "loess", color = "darkgray", linetype = "dashed") +
  geom_point(aes(color = factor(forest_classification, levels = c('undisturbed', 'regrowth', 'degraded', 'deforested'),ordered = TRUE)), size = 2) +
  scale_color_manual(values = c("#1b9e77", "#7570b3", "#e7298a", "#d95f02")) +
  labs(y = "Lorey's height") +
  theme_classic() +
  theme(legend.title = element_blank(),
      legend.position = c(0.9, 0.9),
      legend.background = element_rect(linetype = "solid", size = 0.01, color = "gray50"),
      legend.margin = margin(0.1,5,0.1,0.1))

ggplot(plots_cf_4p, aes(x = env_pc1, y = tree_height)) +
  geom_smooth(alpha= 0.1, method = "loess", color = "darkgray", linetype = "dashed") +
  geom_point(aes(color = factor(forest_classification, levels = c('undisturbed', 'regrowth', 'degraded', 'deforested'),ordered = TRUE)), size = 2) +
  scale_color_manual(values = c("#1b9e77", "#7570b3", "#e7298a", "#d95f02")) +
  labs(y = "Average tree height") +
  theme_classic() +
  theme(legend.title = element_blank(),
      legend.position = c(0.9, 0.9),
      legend.background = element_rect(linetype = "solid", size = 0.01, color = "gray50"),
      legend.margin = margin(0.1,5,0.1,0.1))

ggplot(plots_cf_4p, aes(x = env_pc1, y = basal_area_ha)) +
  geom_smooth(alpha= 0.1, method = "loess", color = "darkgray", linetype = "dashed") +
  geom_point(aes(color = factor(forest_classification, levels = c('undisturbed', 'regrowth', 'degraded', 'deforested'),ordered = TRUE)), size = 2) +
  scale_color_manual(values = c("#1b9e77", "#7570b3", "#e7298a", "#d95f02")) +
  labs(y = "Basal area") +
  theme_classic() +
  theme(legend.title = element_blank(),
      legend.position = c(0.9, 0.9),
      legend.background = element_rect(linetype = "solid", size = 0.01, color = "gray50"),
      legend.margin = margin(0.1,5,0.1,0.1))


```

# Histograms for AGB and log-transformed AGB by plots and sites

```{r}

plots_cf_4p %>%
  ggplot(aes(x= agb_plot_ha)) +
  geom_histogram(aes(y=..density..), color = "gray70", fill = "gray90") +
  geom_line(stat = "density", size = 0.8) +
  geom_rug() +
  theme_classic()

plots_cf_4p %>%
  mutate(log_agb = log1p(agb_plot_ha)) %>%
  ggplot(aes(x= log_agb)) +
  geom_histogram(aes(y=..density..), color = "gray70", fill = "gray90") +
  geom_line(stat = "density", size = 0.8) +
  geom_rug() +
  theme_classic()

sites_cf_4p %>%
  ggplot(aes(x = av_agb_site)) +
  geom_histogram(aes(y=..density..), binwidth = 20, color = "gray70", fill = "gray90") +
  geom_line(stat = "density", size = 0.8) +
  geom_rug() +
  theme_classic()

sites_cf_4p %>%
  mutate(log_agb_s = log1p(av_agb_site)) %>%
  ggplot(aes(x = log_agb_s)) +
  geom_histogram(aes(y=..density..), binwidth = 0.15, color = "gray70", fill = "gray90") +
  geom_line(stat = "density", size = 0.8) +
  geom_rug() +
  theme_classic()

```

# Scatterplots of log-AGB with table 3 predictors

```{r}
sites_cf_4p %>%
  mutate(log_agb = log1p(av_agb_site),
         land_use = as.numeric(landscape)) %>%
  select(log_agb, land_use, slope_gee, agriculture, grazing, env_pc1, shannon, sp_richness_site) %>%
  na.omit() %>%
  pivot_longer(cols = c("land_use", "slope_gee", "agriculture", "grazing", "env_pc1", "shannon", "sp_richness_site"),
               names_to = "variable",
               values_to = "values") %>%
  mutate(variable = factor(variable, levels = c("land_use", "agriculture", "slope_gee", "env_pc1", "grazing", "shannon", "sp_richness_site"))) %>%
  ggplot(aes(y = log_agb, x = values)) +
  geom_point() +
  geom_smooth(method = "lm", color = "gray60", alpha= 0.3) +
  labs(x = NULL, y = (bquote('log AGB ('*"Mg"~ha^-1*')'))) +
  facet_wrap(~variable, scales = "free", nrow = 2)
  
```

# GAMMs plots

```{r}

plots_cf_4p %>%
  filter(!is.na(hill_shannon)) %>%
  gather(variable, value, agriculture, slope, env_pc1, hill_shannon) %>%
  ggplot(aes(x = value, y = agb_plot_ha)) +
  geom_point(aes(color = factor(forest_classification, levels = c('undisturbed', 'regrowth', 'degraded', 'deforested'),ordered = TRUE)),
             size = 1.5,
             alpha = 0.7) +
  scale_color_manual(values = c("#1b9e77", "#7570b3", "#e7298a", "#d95f02")) +
  geom_smooth(method = "lm", se = F, color = "gray20") +
  facet_wrap(~variable, scales = "free") +
  scale_y_continuous(trans = "log1p") +
  theme_bw() +
  theme(legend.title = element_blank(),
      legend.position = c(1.1, 1.1),
      legend.background = element_rect(linetype = "solid", size = 0.01, color = "gray50"),
      legend.margin = margin(0.1,5,0.1,0.1)) +
  labs(x = NULL, y = (bquote('log AGB ('*"Mg"~ha^-1*')'))) 
  
```
```{r}
tibble(
  agriculture = seq(from=min(plots_cf_4p$agriculture), 
              to=max(plots_cf_4p$agriculture), 
              length.out = 1000),
  slope = seq(from=min(plots_cf_4p$slope), 
              to=max(plots_cf_4p$slope), 
              length.out = 1000),
  env_pc1 = seq(from=min(plots_cf_4p$env_pc1), 
              to=max(plots_cf_4p$env_pc1), 
              length.out = 1000),
  hill_shannon = seq(from= 1, 
              to= 9, 
              length.out = 1000)
) %>% 
  bind_cols(as_tibble(
    predict(gamm4_4_int$gam, newdata = ., se.fit=TRUE)
  )) %>% 
  gather(variable, value, agriculture, slope, env_pc1, hill_shannon) %>%
  ggplot(aes(x = value, y = fit)) +
  facet_wrap(~variable, scales = "free") +
  geom_ribbon(aes(ymin=fit-se.fit, ymax=fit+se.fit),
              alpha=.2) +
  geom_line(aes(colour = "gamm"))

gam.check(gamm4_4_int$gam)


tibble(
  agriculture = seq(from=min(plots_cf_4p$agriculture), 
              to=max(plots_cf_4p$agriculture), 
              length.out = 1000),
  slope = seq(from=min(plots_cf_4p$slope), 
              to=max(plots_cf_4p$slope), 
              length.out = 1000),
  env_pc1 = seq(from=min(plots_cf_4p$env_pc1), 
              to=max(plots_cf_4p$env_pc1), 
              length.out = 1000),
  hill_shannon = seq(from= 1, 
              to= 9, 
              length.out = 1000)
) %>% 
  bind_cols(as_tibble(
    predict(gamm4_4_int$gam, newdata = ., se.fit=TRUE)
  )) %>% 
  gather(variable, value, agriculture, slope, env_pc1, hill_shannon) %>%
  ggplot(aes(x = value)) +
  geom_point(aes(x = plots_cf_4p$agriculture, y = log1p(plots_cf_4p$agb_plot_ha))) +
  geom_ribbon(aes(ymin=fit-se.fit, ymax=fit+se.fit),
              alpha=.2) +
  geom_line(aes(y = fit)) +
  facet_wrap(~variable, scales = "free")
  



pred <- data.frame(
  agb = log1p(plots_cf_4p$agb_plot_ha),
  agriculture = plots_cf_4p$agriculture,
  slope = plots_cf_4p$slope,
  env_pc1 = plots_cf_4p$env_pc1,
  hill_shannon = plots_cf_4p$hill_shannon,
  pred_values = predict(gamm4_4_int$gam, newdata = plots_cf_4p, se.fit=TRUE)
)

pred %>%
  filter(!is.na(pred_values.fit)) %>%
  gather(variable, value, agriculture, slope, env_pc1, hill_shannon) %>%
  ggplot(aes(x = value)) +
  geom_point(aes(y = agb)) +
  geom_line(aes(y= pred_values.fit), color = "gray30") +
  geom_ribbon(aes(ymin=pred_values.fit-pred_values.se.fit, ymax=pred_values.fit+pred_values.se.fit),
              alpha=.2) +
  facet_wrap(~variable, scales = "free")

plot.gam(gamm4_4_int$gam, residuals = T)

predict_values <- ggeffects::ggpredict(gamm4_4_int, optional = T)
predict_values$env_pc1$group <- c("Environmental gradient", "Environmental gradient", "Environmental gradient", "Environmental gradient", "Environmental gradient")
predict_values$agriculture$group <- c("Agriculture", "Agriculture", "Agriculture", "Agriculture")
predict_values$slope$group <- c("Slope" ,"Slope" ,"Slope" ,"Slope" ,"Slope" ,"Slope" ,"Slope")
predict_values$hill_shannon$group <- c("Hill-Shannon", "Hill-Shannon", "Hill-Shannon", "Hill-Shannon", "Hill-Shannon", "Hill-Shannon")

gamm_plot <- plot(predict_values, 
     add.data = T,
     facets = TRUE,
     colors = c("black", "black", "black", "black"),
     show.x.title = F,
     show.y.title = TRUE,
     jitter = 0.05
     ) +
  scale_y_continuous(trans = "log1p", breaks = c(0, 50, 100, 500, 1000)) +
  labs(y = (bquote('log AGB ('*"Mg"~ha^-1*')')))

# New facet label names for supp variable
labs <- c("Environmental gradient", "Agriculture", "Slope", "Hill-Shannon")
names(labs) <- c("env_pc1", "agriculture", "slope", "hill_shannon")

#gratia::draw(gamm4_4_int)


pred_values_ag <- data.frame(predict_values$agriculture$x,
                             predict_values$agriculture$predicted,
                             predict_values$agriculture$conf.low,
                             predict_values$agriculture$conf.high)

plots_cf_4p %>%
  #filter(!is.na(hill_shannon)) %>%
  #gather(variable, value, agriculture, slope, env_pc1, hill_shannon) %>%
  ggplot() +
  geom_point(aes(x = agriculture, y = agb_plot_ha)) +
  geom_line(aes(x= predict_values.agriculture.x, y = predict_values.agriculture.predicted), data = pred_values_ag) +
  #facet_wrap(~variable, scales = "free") +
  #scale_y_continuous(trans = "log1p") +
  theme_bw()

```
